% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frxst_gage_points_add.R
\name{AddRouteLinkGage}
\alias{AddRouteLinkGage}
\title{Add a 'gages' column to a RouteLink.nc file}
\usage{
AddRouteLinkGage(rlFile, gageIds, comIds, newCopyId, gageMiss = "",
  overwrite = FALSE)
}
\arguments{
\item{rlFile}{Character path+file for the desired Route_Link.nc file}

\item{gageIds}{Vector of character identifiers for the gages, max length = 15. Must collocate with comIds.}

\item{comIds}{Vector of integer comIds for identifying the links (existing in rlFile). Must collocate with comIds.}

\item{newCopyId}{Character identifier for the new copy of the fullDomFile with specified frxst_pts.}

\item{gageMiss}{Value which indicates no gage at this link.}

\item{overwrite}{Logical If the output path/file created from fullDomFile and newCopyId already exists, then overwrite it.}
}
\description{
Add a 'gages' column to a RouteLink.nc file
}
\examples{
\dontrun{
rlFile <- '~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.nc'
gageIds <- c('06730200', #BOULDER_CREEK_AT_NORTH_75TH_ST._NEAR_BOULDER_CO
             '06730160', #FOURMILE_CANYON_CREEK_NEAR_SUNSHINE_CO
             # '06727410', #FOURMILE_CREEK_AT_LOGAN_MILL_ROAD_NEAR_CRISMAN_CO
             '06727500'  #FOURMILE_CREEK_AT_ORODELL_CO
             )
comIds <- c( 42,  #BOULDER_CREEK_AT_NORTH_75TH_ST._NEAR_BOULDER_CO
             38,   #FOURMILE_CANYON_CREEK_NEAR_SUNSHINE_CO
             # 200,  #FOURMILE_CREEK_AT_LOGAN_MILL_ROAD_NEAR_CRISMAN_CO
             82   #FOURMILE_CREEK_AT_ORODELL_CO
              )
newCopyId <- 'threeRealGagesTEST'
AddRouteLinkGage(rlFile, gageIds, comIds, new=identifier)
}
}

