% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AllSame}
\alias{AllSame}
\title{Are all vector entries the same/identical.}
\usage{
AllSame(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical Remove NAs from output?}
}
\value{
Logical
}
\description{
\code{AllSame} check if all vector entries are same/identical.
}
\examples{
AllSame( 1:5 )
AllSame( 0*(1:5) )
}
\keyword{internal}

