% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameriflux_get.R
\docType{data}
\name{amfMeta}
\alias{amfMeta}
\title{Ameriflux site metadata}
\format{data.frame:  166 obs. of  11 variables:
\describe{
 \item{name}{Site name}
 \item{id_txt}{Site ID (used to access datasets)}
 \item{state}{State where site is located}
 \item{cntry}{Country where site is located}
 \item{lon}{Site longitude}
 \item{lat}{Site latitude}
 \item{tz_us}{Time zone for US sites only}
 \item{tz_glob}{General time zones for all sites (no daylight savings)}
 \item{UTC_offset}{UTC offset w/o daylight savings}
 \item{UTC_DST_offset}{UTC offset with daylight savings}
 \item{UTC_offset_int}{UTC offset w/o daylight savings in integer format (for manual time adjustments to UTC)}
}}
\usage{
amfMeta
}
\description{
Metadata for the Ameriflux sites (as of June 2015) including lat/lon and timezone.
}
\seealso{
Other Ameriflux: \code{\link{GetAmeriflux}},
  \code{\link{tzLookup}}
}
\concept{
data
}
\keyword{data}

