% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{AsCharLongInt}
\alias{AsCharLongInt}
\title{Convert long-ish integers to characters without mutilation}
\usage{
AsCharLongInt(vec)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Vector of characters.
}
\description{
\code{AsCharLongInt} is simply a usage of format.
}
\keyword{internal}
\keyword{utilities}

