% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Bias}
\alias{Bias}
\title{Calculate Bias}
\usage{
Bias(m, o, na.rm = TRUE)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
bias
}
\description{
\code{Bias} calculates bias (m-o).
Calculate the bias for vectors
of modelled and observed values.
}
\keyword{internal}

