% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_chan_connectivity.R
\name{CalcChanConnect}
\alias{CalcChanConnect}
\title{Calculate the channel connectivity for gridded routing.}
\usage{
CalcChanConnect(hydroGridFile)
}
\description{
This is simply an off-line implementation of how the code solves the connectivity from the
hydrogrid file with some reformatting for less repetition.
trunk/NDHMS/Routing/module_HYDRO_io.F : SUBROUTINE READ_ROUTEDIM, v3.0 prerelease line ~574, ~5225
}
\examples{
hydroFile4Mile <- '~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/Fulldom_hydro_OrodellBasin_100m.nc'
conn4Mile <- CalcChanConnect(hydroFile4Mile)
fromTo <- conn4Mile$toNode; names(fromTo) <- conn4Mile$fromNode
chrtFile <- paste0('~/wrfHydroTestCases/Fourmile_Creek/RUN.RTTESTS/OUTPUT_CHRT_DAILY/201308010000.CHRTOUT_DOMAIN1')
Viz4Mile <- VisualizeChanNtwk(chrtFile, plot=FALSE)
chrtData <- Viz4Mile()
}

