% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_chan_connectivity2.R
\name{CalcChanConnect2}
\alias{CalcChanConnect2}
\title{Calculate the channel connectivity for gridded routing.}
\usage{
CalcChanConnect2(hydroGridFile, quiet = FALSE)
}
\arguments{
\item{hydroGridFile}{character, the path/fileName to the "Fulldom" file, aka "the hydro grid file"}

\item{quiet}{Logical Print information about the channel connectivity?}
}
\description{
This is simply an off-line implementation of how the code solves the connectivity from the
hydrogrid file with some reformatting for less repetition.
trunk/NDHMS/Routing/module_HYDRO_io.F : SUBROUTINE READ_ROUTEDIM, v3.0 prerelease line ~574, ~5225
}
\examples{
conusFullDom <- '~/chimayoSpace/WRF_Hydro/CONUS/DOMAIN/Fulldom_hires_netcdf_file.nc'
connConus <- CalcChanConnect2(conusFullDom)

hydroFile4Mile <- '~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/Fulldom_hydro_OrodellBasin_100m.nc'
conn4Mile  <- CalcChanConnect2(hydroFile4Mile)
conn4Mile0 <-  CalcChanConnect(hydroFile4Mile)
identical(conn4Mile0, conn4Mile)

fromTo <- conn4Mile$toNode; names(fromTo) <- conn4Mile$fromNode
toFrom <- conn4Mile$fromNode; names(fromTo) <- conn4Mile$toNode

#Watch the "flows" go up stream from outlet!
chrtFile <- paste0('~/wrfHydroTestCases/Fourmile_Creek/RUN.RTTESTS/OUTPUT_CHRT_DAILY/201308010000.CHRTOUT_DOMAIN1')
Viz4Mile <- VisualizeChanNtwk(chrtFile, plot=FALSE)
chrtGg <- Viz4Mile()

toTo=350
cc=1
while(length(toTo)) {
whTf <- c('chLat','chLon','fromNode')
tfSub <- subset(conn4Mile, toNode \%in\% toTo | fromNode \%in\% toTo )[whTf]
tfSub$var <- ifelse(tfSub$fromNode \%in\% toTo, 'to', 'from')
print("---------")
print(cc)
print(tfSub)
png(file=paste0('~/tmpPngs/fourmile.',formatC(cc,,dig=3,flag='0'),'.png'), width=210*5*2,height=70*5*2, pointsize=3)
print(chrtGg$plot + ggplot2::geom_point(data=tfSub,
                        ggplot2::aes(x=chLon, y=chLat, color=var) ))
dev.off()
toTo <- subset(tfSub, var=='from')$fromNode
cc=cc+1
#if(cc==30) break
print(toTo)
#Sys.sleep(2)
}
}

