% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcCOM}
\alias{CalcCOM}
\title{Calculate center-of-mass.}
\usage{
CalcCOM(x, na.rm = TRUE)
}
\arguments{
\item{x}{The time series vector.}
}
\value{
The center-of-mass time step.
}
\description{
\code{CalcCOM} calculates the time step of center of mass.
Calculate the time step when the center-of-mass of
a time series of values occurs.
}
\keyword{internal}

