% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_daily_ghcn.R
\name{CalcDailyGhcn}
\alias{CalcDailyGhcn}
\title{To aggregate the hourly data into daily data comparable to GHCN daily data.}
\usage{
CalcDailyGhcn(sg, prcp, reportTime = 700, parallel = FALSE)
}
\arguments{
\item{sg}{A dataframe of the selectedGauges, it should have at least one column
with name of siteIds.
siteIds should match the the standardized GHCN IDs (for example : ACW00011604).
See \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}
for a list of siteIds.}

\item{prcp}{A dataframe of hourly data. It should have at least two columns of POSIXct,
statArg, and DEL_ACCPRCP.POSIXct has the timing at UTC. statArg has the siteIds, and
DEL_ACCPRCP is the depth of rainfall since wrf_hydro report the accumulated rainfall,
one extra step is required to calculate the depth of rainfall for an interval.
The column will be with the same name if one use GetMultiNcdf, ReshapeMultiNcdf and
CalcNoahmpFluxes to get the precipitation from WRF_hydro files.}

\item{reportTime}{A vector of reportTime matching with sg$siteIds.
reportTime is a unique reporting time for each elelment of each gauge.
It may vary for different elemenets for a single gauge.
The report time values are retrived when reading the daily GHCN data
using  \code{GetGhcn2}. Note, use 700 wherever the report time is missing.
It will be overwritten if sg$reportTime exits.}
}
\value{
Daily precipitation comparable with daily GHCN data.
}
\description{
This function inputs the hourly data and aggregate the hourly data into daily info.
GHCN gauges report data usually at 7:00 AM, however, taht is not the case for all
gauges. The reporting time may even differ for the same gauge but different elements.
One can obtain the report time using \code{GetGhcn2}.
}

