% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcDates}
\alias{CalcDates}
\title{Calculate standard date breaks.}
\usage{
CalcDates(x)
}
\arguments{
\item{x}{The input dataframe.}
}
\value{
The input dataframe with date columns added.
}
\description{
\code{CalcDates} calculates standard date breaks.
Calculate standard date breaks.
}

