% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcDateTrunc}
\alias{CalcDateTrunc}
\title{Calculate date object from POSIXct time}
\usage{
CalcDateTrunc(timePOSIXct, timeZone = "UTC")
}
\arguments{
\item{timePOSIXct}{Time in POSIXct format}

\item{timeZone}{Time zone (DEFAULT="UTC")}
}
\value{
Date object
}
\description{
\code{CalcDateTrunc} takes a POSIXct object and outputs
a corresponding date object. POSIXct times are truncated
to a date, not rounded (e.g., 03/15/2014 23:00 will
become 03/15/2014).
}
\keyword{internal}
\keyword{utilities}

