% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_fdc.R
\name{CalcFdc}
\alias{CalcFdc}
\title{Calculate flow duration curve statistics}
\usage{
CalcFdc(strflowDf, strflowCol = "q_cms")
}
\arguments{
\item{strflowDf}{The streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}} or \code{\link{ReadUsgsGage}}).
The data frame must contain a column of streamflow values.}

\item{strflowCol}{The name of the column containing the streamflow values (DEFAULT="q_cms").}
}
\value{
A dataframe containing the original input data with a new column added called "<strflowCol>.fdc"
containing the flow exceedance probabilities.
}
\description{
\code{CalcFdc} calculates flow exceedance probabilities and adds them as a new column to the dataframe.
}
\details{
\code{CalcFdc} reads a streamflow time series dataframe (modeled or observed) and outputs a modified dataframe with
calculated percent exceedances.
}
\examples{
## Take the time series of observed 5-minute streamflow values for Fourmile Creek (stored in the dataframe
## obsStr5min.fc in the column "q_cms") and return the same dataframe with a new column called "q_cms.fdc"
## containing the flow exceedance probabilities.

obsStr5min.fc <- CalcFdc(obsStr5min.fc, "q_cms")
}

