% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_fdc.R
\name{CalcFdcSpline}
\alias{CalcFdcSpline}
\title{Generate a spline-fit funtion for a flow duration curve}
\usage{
CalcFdcSpline(strDf, strCol = "q_cms")
}
\arguments{
\item{strDf}{The streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}} or \code{\link{ReadUsgsGage}})
already processed through \code{\link{CalcFdc}}. The data frame must contain a column of streamflow values and a
corresponding column (named "<strCol>.fdc") of flow exceedance probabilities.}

\item{strCol}{The name of the column containing the streamflow values (DEFAULT="q_cms").}
}
\value{
A spline function fitted to flow (y) vs. flow exceedance probabilities (x).
}
\description{
\code{CalcFdcSpline} generates a spline fit function for a flow duration curve.
}
\details{
\code{CalcFdcSpline} reads the percent exceedances generated by \code{\link{CalcFdc}} and outputs a fitted spline function
useful for plotting or estimating flow at specified exceedance thresholds (e.g., 20\% exceedance probability).
}
\examples{
## Take a time series of observed 5-minute streamflow values for Fourmile Creek (stored in the dataframe
## obsStr5min.fc in the column "q_cms") that has already been processed through "CalcFdc" (so also contains
## a column named "q_cms.fdc") and return a spline fit function.

fdc.obsStr5min.fc <- CalcFdcSpline(obsStr5min.fc, "q_cms")

## Use the spline function to estimate the flow value that is exceeded 20\% of the time.

fdc.obsStr5min.fc(0.2)
> 0.72
}

