% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcMeanMinrm}
\alias{CalcMeanMinrm}
\title{Calculate mean with enforced minimum valid value.
\code{CalcMeanMinrm} calculates a mean with an enforced minimum valid value.
Read a vector and calculate the mean with all values below
a specified minimum value set to NA (and therefore ignored).}
\usage{
CalcMeanMinrm(x, minValid = -1e+30)
}
\arguments{
\item{x}{The vector of values.}

\item{minValid}{The minimum valid value.}
}
\value{
The mean.
}
\description{
Calculate mean with enforced minimum valid value.
\code{CalcMeanMinrm} calculates a mean with an enforced minimum valid value.
Read a vector and calculate the mean with all values below
a specified minimum value set to NA (and therefore ignored).
}

