% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcMeanNarm}
\alias{CalcMeanNarm}
\title{Calculate mean with forced NA removal.
\code{CalcMeanNarm} calculates a mean with forced NA removal.
Read a vector and calculate the mean with forced NA removal.}
\usage{
CalcMeanNarm(x)
}
\arguments{
\item{x}{The vector of values.}
}
\value{
The mean.
}
\description{
Calculate mean with forced NA removal.
\code{CalcMeanNarm} calculates a mean with forced NA removal.
Read a vector and calculate the mean with forced NA removal.
}

