% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_MET_measures.R
\name{CalcMetContGrid}
\alias{CalcMetContGrid}
\title{Calculate MET verification measures for grid continuous variables.}
\usage{
CalcMetContGrid(obs, mod, .funObs = NULL, .funMod = NULL,
  statList = list("numPaired", "meanObs", "meanMod", "multiBias", "RMSE",
  "pearsonCor"), conRange = Inf, probs = NULL, parallel = FALSE,
  ncors = 100)
}
\arguments{
\item{obs}{Array of observation time series (first dimension is for time), or a list of observation files.}

\item{mod}{Array of model time series (first dimension is for time), or a list of observation files.}

\item{.funObs}{Function for reading observation file and extracting the observation data from it.}

\item{.funMod}{Function for reading model file and extracting the model data from it.}

\item{statList}{List of statistics to be calculated.}

\item{conRange}{Numeric vector containing two elements.
Values are used as the lower and upper boundary for conditional statistics.
If conditioning only at one tail, leave the second value as Inf.}

\item{probs}{Numeric vector of probabilities with values in [0,1], only required if quantile is in the statList}

\item{parallel}{Logical}

\item{ncors}{Numeric to define how many file to be read in a chunk}
}
\value{
A list of all statistics requested.
}
\description{
\code{CalcMetContGrid}  computes MET verification measures for continous
variables conditionally or unconditionally.
\code{CalcMetContGrid} can be called in two ways. First, user reads all the observation
and model values into memory as 3D arrays which the first array is the time and the
the other two are the spatial dimensions.
Second, providing the address to locations observation and mmodel files reside
as well as functions for reading the files. This option is for cases which the data is
too big to be stored in memory.
Pair matching is not done in this function. It assumes the two time series are
comparable one by one.
}
\details{
The calculated statistics are the following: \itemize{
\item numPaired: number of paierd observed and model data
\item minObs: min of observation time series
\item minMod: min of model simulation/forecast time series
\item maxObs: max of observation time series
\item maxMod: max of model simulation/forecast time series
\item meanObs: mean of observation time series
\item meanMod: mean of model simulation/forecast time series
\item stdObs: standard deviation of observation time series
\item stdMod: standard deviation of model simulation/forecast time series
\item pearsonCor: Pearson correlation coefficient
\item spearmanCor: spearman correlation coefficient
\item kendallCor: kendall tau correlation coefficient (tau-b)
\item ME:mean error
\item multiBias: mutliplicative bias
\item MSE: mean square error
\item RMSE: root mean square error
\item MAE: mean absolute error
\item MAD: median absolute deviation
\item IQR: inter quantile range of the errors
\item quantiles: quantiles of error for given probabilities
} For more information refer to Model Evaluation Tool (MET) documentation
}
\examples{
obs<-array(rnorm(300*4*5),c(300,4,5))
mod<-array(rnorm(300*4*5),c(300,4,5))
stat<-CalcMetContGrid(obs,mod, statList = list('multiBias','RMSE','pearsonCor'))
}
\seealso{
Other modelEvaluation: \code{\link{CalcMetCont}};
  \code{\link{CalcModPerfMulti}};
  \code{\link{CalcModPerf}};
  \code{\link{CalcNoahmpFluxes}};
  \code{\link{CalcNoahmpWatBudg}}
}
\concept{
modelEval
}

