% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{CalcModPerfMulti}
\alias{CalcModPerfMulti}
\title{Computes model performance statistics for WRF-Hydro flux output}
\usage{
CalcModPerfMulti(flxDf.mod, flxDf.obs, flxCol.mod = "q_cms",
  flxCol.obs = "q_cms", stdate = NULL, enddate = NULL)
}
\arguments{
\item{flxDf.mod}{The flux output dataframe (required). Assumes only one forecast
point per file, so if you have multiple forecast points in your output dataframe, use
subset to isolate a single forecast point's data. Also assumes model output and observation
both contain POSIXct fields (called "POSIXct").}

\item{flxDf.obs}{The observed flux dataframe. Assumes only one observation point per file, so if
you have multiple observation points in your dataframe, use subset to isolate a single point's data.
Also assumes model output and observation both contain POSIXct fields (called "POSIXct").}

\item{flxCol.mod}{The column name for the flux time series for the MODEL data (default="q_cms")}

\item{flxCol.obs}{The column name for the flux time series for the OBSERVED data (default="q_cms")}

\item{stdate}{Start date for plot/statistics (DEFAULT=NULL, all records will be used).
Date MUST be specified in POSIXct format with appropriate timezone
(e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC"))}

\item{enddate}{End date for plot/statistics (DEFAULT=NULL, all records will be used).
Date MUST be specified in POSIXct format with appropriate timezone
(e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC"))}
}
\value{
A new dataframe containing the model performance statistics.
}
\description{
\code{CalcModPerfMulti} calculates model performance statistics for flux output.
}
\details{
\code{CalcModPerfMulti} reads a model flux time series (i.e., created using \code{\link{ReadFrxstPts}}) and
an observation time series (i.e., created using \code{\link{ReadUsgsGage}}) and calculates model performance
statistics (Nash-Sutcliffe Efficiency, Rmse, etc.) at various time scales and for low
and high fluxes. The tool will subset data to matching time periods (e.g., if the
observed data is at 5-min increments and modelled data is at 1-hr increments, the tool
will subset the observed data to select only observations on the matching hour break).

\code{CalcModPerfMulti} calculates the same statistics as \code{\link{CalcModPerf}}) but returns
results as a single-row vs. multi-row dataframe. This is intended to streamline compiling statistics
for multiple data runs or multiple sites.

Performance Statistics:
\cr (mod = model output, obs = observations, n = sample size)
\itemize{
\item n: sample size
\item nse: Nash-Sutcliffe Efficiency
\deqn{nse = 1 - ( sum((obs - mod)^2) / sum((obs - mean(obs))^2) ) }
\item nselog: log-transformed Nash-Sutcliffe Efficiency
\deqn{nselog = 1 - ( sum((log(obs) - log(mod))^2) / sum((log(obs) - mean(log(obs)))^2) ) }
\item cor: correlation coefficient
\deqn{cor = cor(mod, obs) }
\item rmse: root mean squared error
\deqn{rmse = sqrt( sum((mod - obs)^2) / n ) }
\item rmsenorm: normalized root mean squared error
\deqn{rmsenorm = rmse / (max(obs) - min(obs)) }
\item bias: percent bias
\deqn{bias = sum(mod - obs) / sum(obs) * 100 }
\item mae: mean absolute error
\deqn{mae = mean(abs(mod - obs)) }
\item errcom: error in the center-of-mass of the flux, where center-of-mass is the
hour/day when 50\% of daily/monthly/water-year flux has occurred. Reported as number of hours for
daily time scale and number of days for monthly and yearly time scales.
\item errmaxt: Error in the time of maximum flux. Reported as number of hours for daily time scale
and number of days for monthly and yearly time scales).
\item errfdc: Error in the integrated flow duration curve between 0.05 and 0.95 exceedance thresholds
(in native flow units).
}

Time scales/Flux types:
\itemize{
\item t = native model/observation time step (e.g., hourly)
\item dy = daily time step
\item mo = monthly time step
\item yr = water-year time step
\item max10 = high flows; restricted to the portion of the time series where the observed flux
is in the highest 10\% (native time step)
\item min10 = low flows; restricted to the portion of the time series where the observed flux
is in the lowest 10\% (native time step)
}
}
\examples{
## Take forecast point model output for Fourmile Creek (modStrh.mod1.fc) and a corresponding
## USGS gage observation file (obsStrh.fc), both at an hourly time step, and calculate
## model performance statistics. The model forecast point data was imported using ReadFrxstPts
## and the gage observation data was imported using ReadUsgsGage.

CalcModPerfMulti(modStr1h.allrt.fc, obsStr5min.fc)

> Output:
t_n t_nse t_nselog t_cor t_rmse t_rmsenorm t_bias t_mae t_errfdc dy_n dy_nse dy_nselog
744  0.63     0.66   0.8   0.29      13.29    0.7  0.19     0.04   32   0.69      0.74
dy_cor dy_rmse dy_rmsenorm dy_bias dy_mae dy_errcom dy_errmaxt dy_errfdc mo_n
  0.83    0.26       13.23     0.6   0.18      0.03      -0.56      0.03    2
mo_nse mo_nselog mo_cor mo_rmse mo_rmsenorm mo_bias mo_mae mo_errcom mo_errmaxt
     1      0.99      1    0.01        2.46    -0.8   0.01       0.5        1.5
yr_n yr_bias yr_mae yr_errcom yr_errmaxt max10_n max10_nse max10_nselog max10_cor
   1     0.7   0.01         0          3      75     -5.05         -6.9     -0.84
max10_rmse max10_rmsenorm max10_bias max10_mae min10_n min10_nse min10_nselog
      0.72          80.01      -27.9       0.6     138     -2.64        -2.43
min10_cor min10_rmse min10_rmsenorm min10_bias min10_mae
     0.59       0.11          53.89       -7.3      0.09
}
\concept{
aconcept
}
\keyword{ts}
\keyword{univar}

