% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcMonthDays}
\alias{CalcMonthDays}
\title{Calculate number of days in a month.}
\usage{
CalcMonthDays(mo, yr)
}
\arguments{
\item{mo}{The month.}

\item{yr}{The year.}
}
\value{
The day count.
}
\description{
\code{CalcMonthDays} calculates the number of days in a month.
Calculate the number of days in the month specified
by the given month and year.
}
\keyword{internal}
\keyword{utilities}

