% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{CalcNoahmpWatBudg}
\alias{CalcNoahmpWatBudg}
\title{Calculate water balance from WRF-Hydro (w/NoahMP) output}
\usage{
CalcNoahmpWatBudg(ldasoutDf, rtoutDf = NULL, gwoutDf = NULL,
  sfcrt = FALSE, soildeps = c(100, 300, 600, 1000), basarea = NULL)
}
\arguments{
\item{ldasoutDf}{The LDASOUT dataframe (required)}

\item{rtoutDf}{The RTOUT dataframe, if overland or subsurface routing was
turned on (default=NULL)}

\item{gwoutDf}{The GW_OUT dataframe, if groundwater model was turned on
(default=NULL)}

\item{sfcrt}{A flag whether surface overland flow routing was active. All
other routing options are determined based on the input dataframes, as
needed (e.g., if gwoutDf is provided, it is assumed that the groundwater
model was active). (default=FALSE)}

\item{soildeps}{A list of soil layer depths in mm (top to bottom,
default=c(100, 300, 600, 1000))}

\item{basarea}{The basin area in square km (necessary only if gwoutDf is
provided)}
}
\value{
A new dataframe containing the water budget components in mm.
}
\description{
\code{CalcNoahmpWatBudg} calculates water budget components from WRF-Hydro
(w/NoahMP) model output.
}
\details{
\code{CalcNoahmpWatBudg} reads dataframes derived from WRF-Hydro output
(i.e., using \code{\link{GetMultiNcdf}}) and calculates water budget
partitioning (e.g., surface runoff, evaporation, groundwater). Assumes
WRF-Hydro output dataframes have already been masked to the desired basin.
See \code{\link{GetMultiNcdf}} documentation for examples of how to do this.
NOTE: Currently only works for model runs using NoahMP as the LSM.

REQUIRED variables (these terms must be in your input dataframes): \itemize{
\item LDASOUT: ACCPRCP, ACCECAN, ACCETRAN, ACCEDIR, SFCRNOFF, UGDRNOFF,
SOIL_M (all layers), SNEQV, CANICE, CANLIQ \item RTOUT (optional, use if
overland or subsurface routing were activated): QSTRMVOLRT, SFCHEADSUBRT,
QBDRY \item GWOUT (optional, use if groundwater bucket model was activated):
q_cms, POSIXct }

OUTPUT water budget terms (may vary depending on model configuration):
\itemize{ \item LSM_PRCP: Total precipitation (mm) \item LSM_ECAN: Total
canopy evaporation (mm) \item LSM_ETRAN: Total transpiration (mm) \item
LSM_EDIR: Total surface evaporation (mm) \item LSM_DELSWE: Change in snowpack
snow water equivalent (mm) \item LSM_DELCANWAT: Change in canopy water
storage (liquid + ice) (mm) \item LSM_SFCRNOFF: Surface runoff from LSM
\emph{(for an LSM-only run)} (mm) \item LSM_UGDRNOFF: Subsurface runoff from
LSM \emph{(for an LSM-only run)} (mm) \item LSM_DELSOILM: Change in total
soil moisture storage (mm) \item HYD_QSTRMVOL: Total runoff into channel from
land \emph{(routing model only)}  (mm) \item HYD_DELSFCHEAD: Change in
surface storage \emph{(routing model only)} (mm) \item HYD_QBDRY: Total flow
outside of domain \emph{(routing model only)} (mm) \item HYD_GWOUT: Total
groundwater outflow \emph{(routing model only)} (mm) \item
WB_SFCRNOFF: Total surface runoff used in the water budget calculation
(\emph{either} LSM_SFCRNOFF or HYD_QSTRMVOL) (mm) \item WB_GWOUT: Total
groundwater outflow used in the water budget calculation (\emph{either}
LSM_UGDRNOFF or HYD_GWOUT) (mm) \item WB_DELGWSTOR: Change in groundwater
storage (adjusted by surface runoff when no surface routing active) (mm)
\item ERROR: Remainder in water budget (mm)
\item RUN_FRAC: Runoff fraction, runoff/precipitation \item EVAP_FRAC:
Evaporative fraction, evapotranspiration/precipitation \item STOR_FRAC:
Change in storage fraction, storagechange/precipitation }
}
\examples{
## Take a NoahMP LDASOUT dataframe for a model run of Fourmile Creek with no routing
## options turned on and return a water budget summary.

\dontrun{
wb.nort.fc <- CalcNoahmpWatBudg(modLDASOUT1d.nort.fc)
wb.nort.fc
##
## Take NoahMP LDASOUT, HYDRO model RTOUT, and GW_outflow dataframes for a model
## run of Fourmile Creek with subsurface, overland, and groundwater routing options
## turned on and return a water budget summary. The default soil depths were used
## and the basin is 63.1 km2. NOTE: We MUST specify with the sfcrt flag that overland
## flow routing was turned on. Otherwise the LSM surface runoff term is used.

wb.allrt.fc <- CalcNoahmpWatBudg(modLDASOUT1d.allrt.fc,
                                 modRTOUT1h.allrt.fc,
                                 modGWOUT1h.allrt.fc,
                                 sfcrt=TRUE, basarea=63.1)
wb.allrt.fc
}
}
\seealso{
Other modelEvaluation: \code{\link{CalcMetContGrid}};
  \code{\link{CalcMetCont}};
  \code{\link{CalcModPerfMulti}};
  \code{\link{CalcModPerf}};
  \code{\link{CalcNoahmpFluxes}};
  \code{\link{CalcStatCategorical}};
  \code{\link{CalcStatCont}}
}
\concept{
modelEval
}
\keyword{manip}

