% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcRunningMean}
\alias{CalcRunningMean}
\title{Calculate a running mean}
\usage{
CalcRunningMean(x, n, sides = 2, ts = FALSE)
}
\arguments{
\item{x}{Vector of values}

\item{n}{Number of records to use}

\item{sides}{Whether to use both sides (2=past and 
future) or just one side (1=past only).}

\item{ts}{Flag whether or not to return a time
series object or just a vector (DEFAULT=FALSE,
which returns a vector)}
}
\value{
Vector of moving averages
}
\description{
\code{CalcRunningMean} takes series and calculates
a running mean over specified number of records
}
\keyword{internal}
\keyword{utilities}

