% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{CalcSnodasCoords}
\alias{CalcSnodasCoords}
\title{CalcSnodasCoords}
\usage{
CalcSnodasCoords()
}
\value{
A list of lon and lat
}
\description{
\code{CalcSnodasCoords} Get and unpack the SNODAS snow depth and SWE tarball for a given date.
}
\examples{
snodasCoords <- CalcSnodasCoords()
}

