% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{CalcStrPerf}
\alias{CalcStrPerf}
\title{Computes model performance statistics for WRF-Hydro streamflow output}
\usage{
CalcStrPerf(stroutDf, obsDf, strCol = "q_cms", obsCol = "q_cms")
}
\arguments{
\item{stroutDf}{The forecast point output dataframe (required). Assumes only one forecast
point per file, so if you have multiple forecast points in your output dataframe, use
subset to isolate a single forecast point's data. Also assumes model output time step and
observation time step match, and both contain POSIXct fields (called "POSIXct").}

\item{obsDf}{The observed streamflow dataframe. Assumes only one gage per file, so if
you have multiple forecast points in your output dataframe, use subset to isolate
a single gage's data. Also assumes model output time step and observation time step match,
and both contain POSIXct fields (called "POSIXct").}

\item{strCol}{The column name for the streamflow time series for the MODEL data (default="q_cms")}

\item{obsCol}{The column name for the streamflow time series for the OBSERVED data (default="q_cms")}
}
\value{
A new dataframe containing the model performance statistics.
}
\description{
\code{CalcStrPerf} calculates model performance statistics for streamflow output.
}
\details{
\code{CalcStrPerf} reads a model forecast point streamflow timeseries (i.e., created using \code{\link{ReadFrxstPts}}) and
a streamflow observation timeseries (i.e., created using \code{\link{ReadUsgsFlow}}) and calculates model performance
statistics (Nash-Sutcliffe Efficiency, Rmse, etc.) at various time scales and for low
and high flows. Assumes model output and observation datasets have the same time step
(e.g., both hourly).

Performance Statistics:
\cr (mod = model output, obs = observations, n = sample size)
\itemize{
\item nse: Nash-Sutcliffe Efficiency
\deqn{nse = 1 - ( sum((obs - mod)^2) / sum((obs - mean(obs))^2) ) }
\item nselog: log-transformed Nash-Sutcliffe Efficiency
\deqn{nselog = 1 - ( sum((log(obs) - log(mod))^2) / sum((log(obs) - mean(log(obs)))^2) ) }
\item cor: correlation coefficient
\deqn{cor = cor(mod, obs) }
\item rmse: root mean squared error
\deqn{rmse = sqrt( sum((mod - obs)^2) / n ) }
\item rmsenorm: normalized root mean squared error
\deqn{rmsenorm = rmse / (max(obs) - min(obs)) }
\item bias: percent bias
\deqn{bias = sum(mod - obs) / sum(obs) * 100 }
\item mae: mean absolute error
\deqn{mae = mean(abs(mod - obs)) }
\item errcom: error in the center-of-mass of streamflow, where center-of-mass is the
hour/day when 50\% of daily/monthly/water-year streamflow has occurred. Reported as number of hours for
daily time scale and number of days for monthly and yearly time scales.
\item errmaxt: Error in the time of maximum streamflow. Reported as number of hours for daily time scale
and number of days for monthly and yearly time scales).
\item errfdc: Error in the integrated flow duration curve between 0.05 and 0.95 exceedance thresholds
(in native flow units).
}

time scales/Flow Types:
\itemize{
\item ts = native model/observation time step (e.g., hourly)
\item daily = daily time step
\item monthly = monthly time step
\item yearly = water-year time step
\item max10 = high flows; restricted to the portion of the time series where the observed streamflow is in the highest 10%
\item min10 = low flows; restricted to the portion of the time series where the observed streamflow is in the lowest 10%
}
}
\examples{
## Take forecast point model output for Fourmile Creek (modStrh.mod1.fc) and a corresponding
## USGS gage observation file (obsStrh.fc), both at an hourly time step, and calculate
## model performance statistics. The model forecast point data was imported using ReadFrxstPts
## and the gage observation data was imported using ReadUsgsFlow.

CalcStrPerf(modStrh.mod1.fc, obsStrh.fc)

> Output:
          nse nselog  cor rmse rmsenorm  bias  mae errcom errmaxt errfdc
ts       0.57   0.61 0.79 1.43     9.48 -28.0 0.70     NA      NA  -0.42
daily    0.71   0.64 0.87 1.17     9.86 -28.1 0.61   0.19   -2.25  -0.37
monthly  0.80   0.70 0.93 0.89    12.73 -26.6 0.53  -0.18   -0.96     NA
yearly   0.05   0.37 0.36 0.55    41.50  -6.5 0.45  -1.50   -3.38     NA
max10   -7.50 -15.94 0.19 3.82    38.89 -24.5 0.04     NA      NA     NA
min10   -2.84  -1.83 0.10 0.05    33.36 -23.7   NA     NA      NA     NA
}

