% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcWaterYear}
\alias{CalcWaterYear}
\title{Returns the water year or the day of water year for a given POSIXct.}
\usage{
CalcWaterYear(POSIXct, dayOf = FALSE)
}
\arguments{
\item{POSIXct}{is a POSIXct variable.}

\item{dayOf}{signals if you want to get back the day of the water year instead of the water year.}
}
\description{
\code{CalcWaterYear} Returns the water year or the day of water year for a given POSIXct.
}
\examples{
CalcWaterYear(as.POSIXct(c("2011-09-30", "2011-10-01"), tz='US/Pacific'))
CalcWaterYear(as.POSIXct(c("2011-09-30", "2011-10-01"), tz='US/Pacific'), dayOf=TRUE)
}

