% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_MET_measures.R
\name{CalMetCont}
\alias{CalMetCont}
\title{Calculate the MET verification measures for continuous variables.}
\usage{
CalMetCont(obs, mod, conRange = NULL)
}
\arguments{
\item{obs}{A numeric vector of Observation time series.}

\item{mod}{A numeric vector of Model simulation or forecast time series.}

\item{conRange}{A numeric vector containing two elements.
Values are used as the lower and upper boundary for conditional statistics.
If conditioning only at one tail, leave the second value as NA.}
}
\value{
A matrix containing most of the MET verification measures
for continuous variable.
}
\description{
It inputs two times series of observation and model/simlulated and
computes the MET verification measures for continous variables.
Pair matching is not done in this subroutine and should be performed
before introducing to this function. It assumes the two time series are
comparable one by obe.
}
\examples{
DF <- data.frame(obs=seq(1:5), mod=seq(6:11), stringsAsFactors=FALSE)
stat<-CalMetVeriCont(DF$obs,DF$sim)
}

