% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{ChanConnToRouteLink}
\alias{ChanConnToRouteLink}
\title{Reformat a CHANNEL_CONNECTIVITY.nc file as a Route_Link.nc file
The idea here is to simply reexpress the CHANNEL_CONNECTIVITY.nc file as
a pseudo - Route_Link.nc file with only the basic information, so that it can
be passed to all the existing network processing routines as any Route_Link.nc}
\usage{
ChanConnToRouteLink(chanConnFile)
}
\description{
Reformat a CHANNEL_CONNECTIVITY.nc file as a Route_Link.nc file
The idea here is to simply reexpress the CHANNEL_CONNECTIVITY.nc file as
a pseudo - Route_Link.nc file with only the basic information, so that it can
be passed to all the existing network processing routines as any Route_Link.nc
}

