% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_to_timeslice.R
\name{ChanObsToTimeSlice}
\alias{ChanObsToTimeSlice}
\title{Process model CHANOBS to observation timeSlice format for nudging DA.}
\usage{
ChanObsToTimeSlice(files, sliceResolutionMin, outputDir, na.rm = FALSE,
  parallel = FALSE)
}
\arguments{
\item{files}{Character vector path/file to CHANOBS files.}

\item{sliceResolutionMin}{Integer must be <= 60 and evenly divide 60.}

\item{outputDir}{Character the directory where the ouput files are to be written}

\item{na.rm}{Remove NA's from the timeseries?}

\item{parallel}{Create the output files in parallel?}
}
\description{
Process model CHANOBS to observation timeSlice format for nudging DA.
}
\examples{
\dontrun{
  files <- list.files('~/WRF_Hydro/testChanobs/', pattern = 'CHANOBS_DOMAIN', full.names=TRUE)
  sliceResolutionMin <- 12
  outputDir <- '~/WRF_Hydro/testChanobs/timeSliceTest/'
  ChanObsToTimeSlice(files, sliceResolutionMin, outputDir)
}
}
\author{
James McCreight
}
\seealso{
Other nudging: \code{\link{AddRouteLinkGage}};
  \code{\link{EditFrxstPts}}; \code{\link{MkNudgingParams}}
}
\concept{
nudging
}
\keyword{IO}
\keyword{manip}

