% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{CheckConn}
\alias{CheckConn}
\title{CheckConn checks a re-expressed network.}
\usage{
CheckConn(ind, upstream = TRUE, printInds = FALSE)
}
\arguments{
\item{ind}{The indices to check.}

\item{upstream}{Logical, check connectivity upstream (TRUE) or downstream (FALSE).}

\item{printInds}{Logical, print the indices checked.}
}
\value{
Logical, code halts at first FALSE.
}
\description{
\code{CheckConn} checks that a re-expressed network matches it's original expression.
}
\examples{
\dontrun{
  for (ii in seq(1,2720000,1000)) { print(ii); print(CheckConn(ii)) }
  for (ii in seq(1,2720000,1000)) { print(ii); print(CheckConn(ii), up=FALSE) }
  for (ii in seq(1,2000)) { print(ii); print(CheckConn(ii)) }
  for (ii in seq(1,2000)) { print(ii); print(CheckConn(ii),up=FALSE) }

load("/home/jamesmcc/WRF_Hydro/DOMAIN_library/BoCr_100m_1km_NHDPlus_2015_08_11/Route_Link.reInd.Rdb")

load("/home/jamesmcc/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink_2015_07_31.reExpFrom.Rdb")
load("/home/jamesmcc/WRF_Hydro/DOMAIN_library/BoCr_100m_1km_NHDPlus_2015_08_11/Route_Link.reExpFrom.Rdb")
## number of contributing/upstream links.
nContrib<-from$end-from$start
nContrib[which(from$start>0)] <-nContrib[which(from$start>0)] +1
table(nContrib)
for (ii in which(nContrib >3)) { print(ii); print(CheckConn(ii),up=FALSE) }
comIdWhContribGt3 <-
  data.frame(nContrib = nContrib[which(nContrib > 16)],
             comId = reInd$comId[which(nContrib > 16)] )
comIdWhContribGt3 <- comIdWhContribGt3[order(comIdWhContribGt3$nContrib),]
write.table(comIdWhContribGt3, row.names=FALSE,
            file='~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink4.comIdWhContribGt3.txt')

load("/home/jamesmcc/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink_2015_07_31.reExpTo.Rdb")
load("/home/jamesmcc/WRF_Hydro/DOMAIN_library/BoCr_100m_1km_NHDPlus_2015_08_11/Route_Link.reExpFrom.Rdb")
## number of downstream/outflow links.
nOut<-to$end-to$start
nOut[which(to$start>0)] <-nOut[which(to$start>0)] +1
table(nOut)
for (ii in which(nOut >1)) { print(ii); print(CheckConn(ii),up=FALSE) }
comIdWhOutGt1 <-
  data.frame(nOut = nOut[which(nOut > 1)],
             comId = reInd$comId[which(nOut > 1)] )
comIdWhOutGt1 <- comIdWhOutGt1[order(comIdWhOutGt1$nOut),]
write.table(comIdWhOutGt1, row.names=FALSE,
            file='~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink4.comIdWhOutGt1.txt')

load("/home/jamesmcc/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink_2015_07_31.reInd.Rdb")
}  #dontrun
}
\concept{
dataMgmt nudging
}
\keyword{manip}

