% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{CheckReGridConnUp}
\alias{CheckReGridConnUp}
\title{Check for re-rxpressed grid connectivity upstream.}
\usage{
CheckReGridConnUp(ind, printInds = FALSE)
}
\description{
Check for re-rxpressed grid connectivity upstream.
}

