% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_RS.R
\name{ConvertBrick2NC}
\alias{ConvertBrick2NC}
\title{Convert a raster brick to a NetCDF file.}
\usage{
ConvertBrick2NC(inBrick, outFile = NULL, varName = NULL, varUnit = NULL,
  varLong = NULL, varNA = -1e+36)
}
\arguments{
\item{inBrick}{The name of the raster brick to export.}

\item{outFile}{Name for an output NetCDF file.}

\item{varName}{Name for the NetCDF export variable.}

\item{varUnit}{Units for the NetCDF export variable.}

\item{varLong}{Long name for the NetCDF export variable.}

\item{varNA}{Value to set for "NA" or "no data". Default is -1.e+36.}
}
\description{
\code{ConvertBrick2NC} takes a raster brick of RS images and outputs a NetCDF file
with a time dimension and specified variable name & metadata.
}
\details{
\code{ConvertBrick2NC} converts a raster brick to an output NetCDF file. The raster brick
should be already processed through the \code{\link{ConvertRS2Brick}} tool or follow the
same layer (date) naming convention.
}
\examples{
## Export the raster brick of LAI images created through ConvertRS2Brick to a NetCDF file. Use the full time series of images.

ConvertBrick2NC(lai.b, outFile="BCNED_LAI.nc", varName="LAI", varUnit="(m^2)/(m^2)", varLong="Leaf area index")
}

