% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RS_get.R
\name{ConvertRS2Stack}
\alias{ConvertRS2Stack}
\title{Convert a set of MODIS images to a raster stack and, optionally, a NetCDF file.}
\usage{
ConvertRS2Stack(inPath, matchStr, begin = NULL, end = NULL, noData = NULL,
  noDataQual = "exact", valScale = 1, valAdd = 0, outFile = NULL,
  varName = NULL, varUnit = NULL, varLong = NULL, varNA = -1e+36)
}
\arguments{
\item{inPath}{The path to the directory that holds the already processed MODIS TIF files.}

\item{matchStr}{The regular expression for filename matching (e.g., "*Lai_1km.tif").}

\item{begin}{Date string for the start date to include. The
date string should follow \href{http://r-forge.r-project.org/projects/modis/}{MODIS} package convention (e.g., "2011.06.01").
(DEFAULT=NULL, all files are processed).}

\item{end}{Date string for the end date to include. The
date string should follow \href{http://r-forge.r-project.org/projects/modis/}{MODIS} package convention (e.g., "2011.06.01").
(DEFAULT=NULL, all files are processed).}

\item{noData}{Value for specifying "no data" per the MODIS product. Should be combined
with a qualifier. For example, for the MOD15A2 (LAI) product, valid value range is 0-100.
So the noData value should be set to 100 and the noDataQual to "max". (DEFAULT=NULL, no
additional "no data" screening is applied)}

\item{noDataQual}{Qualifier for specifying "no data" per the MODIS product. Should be combined
with a nodata value. For example, for the MOD15A2 LAI product, valid value range is 0-100.
So the noData value should be set to 100 and the noDataQual to "max".
Options are: "exact" (value == noData are converted to NA), "min" (value < noData are converted to NA),
"max" (value > noData are converted to NA). (DEFAULT="exact", only exact matches are converted to NA)}

\item{valScale}{The scale factor to apply to the image values per the MODIS product. The
adjustment is applied as VALUE * valScale + valAdd. For example, for the MOD15A2 LAI product,
the scale factor is 0.1. (DEFAULT = 1)}

\item{valAdd}{The addition value to apply to the image values per the MODIS product. The
adjustment is applied as VALUE * valScale + valAdd. (DEFAULT = 0)}

\item{outFile}{OPTIONAL name for an output NetCDF file. A NetCDF file will only be created if
this file name is provided. Images will be exported after the
"no data" and scale adjustments are made. If you want to do smoothing or other time series
processing, do not export a NetCDF file here but do processing on the raster stack and then
use ConvertStack2NC to export the processed brick to a NetCDF file.}

\item{varName}{Name for the NetCDF export variable. Only required if outFile is provided.}

\item{varUnit}{Units for the NetCDF export variable. Only required if outFile is provided.}

\item{varLong}{Long name for the NetCDF export variable. Only required if outFile is provided.}

\item{varNA}{Value to set for "NA" or "no data". Default is -1.e+36. Only required if outFile is provided.}
}
\value{
A raster stack.
}
\description{
\code{ConvertRS2Stack} takes a set of pre-processed MODIS TIF files and creates
a raster stack and, optionally, a NetCDF file.
}
\details{
\code{ConvertRS2Stack} scans the specified directory and imports pre-processed
MODIS TIF files matching the specified expression and combines them into an R raster
stack object and, optionally, an output NetCDF file. Files in the input directory
should be already processed through the \code{\link{GetMODIS}} tool or follow the
same file naming convention used by the MODIS \code{\link[MODIS]{runGdal}} tool.
See the R \href{http://r-forge.r-project.org/projects/modis/}{MODIS} package for
more on those specifications.
}
\examples{
## Import the already processed LAI TIF images into a raster stack. Use the full time series of images.

lai.b <- ConvertRS2Stack("/d6/adugger/WRF_Hydro/RS/MODIS_ARC/PROCESSED/BCNED_LAI", "*Lai_1km.tif", noData=100, noDataQual="max", valScale=0.1, valAdd=0)

## Export a subset of the already processed LAI TIF images into an output netcdf file

lai.b <- ConvertRS2Stack("/d6/adugger/WRF_Hydro/RS/MODIS_ARC/PROCESSED/BCNED_LAI", "*Lai_1km.tif", begin=c("2011.06.01", end="2011.06.30", noData=100, noDataQual="max", valScale=0.1, valAdd=0, outFile="BCNED_LAI.nc", varName="LAI", varUnit="(m^2)/(m^2)", varLong="Leaf area index")
}
\concept{
aconcept
}
\keyword{IO}

