% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_get.R
\name{ConvertStack2NC}
\alias{ConvertStack2NC}
\title{Convert a raster stack to a NetCDF file.}
\usage{
ConvertStack2NC(inStack, outFile = NULL, varName = NULL, varUnit = NULL,
  varLong = NULL, varNA = -1e+36, flipY = TRUE)
}
\arguments{
\item{inStack}{The name of the raster stack to export.}

\item{outFile}{Name for an output NetCDF file.}

\item{varName}{Name for the NetCDF export variable.}

\item{varUnit}{Units for the NetCDF export variable.}

\item{varLong}{Long name for the NetCDF export variable.}

\item{varNA}{Value to set for "NA" or "no data". Default is -1.e+36.}

\item{flipY}{Reverse y indices, e.g., to match geogrid S->N 
orientation (DEFAULT=TRUE).}
}
\description{
\code{ConvertStack2NC} takes a raster stack of RS images and outputs a NetCDF
file with a time dimension and specified variable name & metadata.
}
\details{
\code{ConvertStack2NC} converts a raster stack to an output NetCDF file. The
raster stack should be already processed through the
\code{\link{ConvertRS2Stack}} tool or follow the same layer (date) naming
convention.
}
\examples{
## Export the raster stack of LAI images created through ConvertRS2Stack
## to a NetCDF file. Use the full time series of images.

\dontrun{
ConvertStack2NC(lai.b, outFile="BCNED_LAI.nc", varName="LAI", 
                varUnit="(m^2)/(m^2)", varLong="Leaf area index")
}
}
\seealso{
Other MODIS: \code{\link{CalcStatsRS}},
  \code{\link{ConvertRS2Stack}}, \code{\link{GapFillRS}},
  \code{\link{GetMODIS}}, \code{\link{InsertRS}},
  \code{\link{SmoothStack}}
}
\concept{
MODIS dataMgmt
}
\keyword{IO}

