% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{CreateBasinMask}
\alias{CreateBasinMask}
\title{Create a coarse-resolution basin mask grid.}
\usage{
CreateBasinMask(ncfile, mskvar = "basn_msk", basid = 1, aggfact = 1)
}
\arguments{
\item{ncfile}{The full pathname to the WRF-Hydro high-res routing domain file.}

\item{mskvar}{The variable name for the high-res basin mask (DEFAULT="basn_msk")}

\item{basid}{The basin ID to generate a mask file for (DEFAULT=1)}

\item{aggfact}{The aggregation factor for downsampling the high-res grid (e.g.,
aggfact=1 for going from a 100-m routing grid to a 1km geogrid) (DEFAULT=1)}
}
\value{
A matrix containing the basin mask weights on the resampled grid.
}
\description{
\code{CreateBasinMask} reads in a high-res domain file and outputs a resampled
weighted basin mask grid for generating LSM-grid statistics.
}
\details{
\code{CreateBasinMask} reads in a high-res domain file and outputs a resampled
weighted basin mask grid for generating LSM-grid statistics. The output grid will
contain 1 for cells that are completely within the basin, 0 for cells that are
completely outside of the basin, and fractions (based on area) for cells that are
partially within and partially outside of the basin.
}
\examples{
## Take the high-res 100-m routing domain for Fourmile and generate a matrix of
## area weights on the 1km geogrid domain.
\dontrun{
geoMsk <- CreateBasinMask("~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/Fulldom_hydro_OrodellBasin_100m.nc", aggFact=10)
}
}
\seealso{
Other modelDataReads: \code{\link{ReadFrxstPts}};
  \code{\link{ReadGwOut}}; \code{\link{ReadLdasoutAll}};
  \code{\link{ReadLdasoutWb}}; \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}

