% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{CumsumNa}
\alias{CumsumNa}
\title{Calculate cumulative sum with forced NA=0.}
\usage{
CumsumNa(x)
}
\arguments{
\item{x}{The vector of values.}
}
\value{
The cumulative sum vector.
}
\description{
\code{CumsumNa} calculates a cumulative sum with NAs converted to 0s.
Read a vector and calculate the cumulative sum with NAs converted to 0s.
}

