% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{ExtractGRIBGrid}
\alias{ExtractGRIBGrid}
\title{Extract grid of GRIB data given a variable name and level.}
\usage{
ExtractGRIBGrid(fileIn, var, levType, level, nx, ny, numFTimes = 1)
}
\arguments{
\item{fileIn}{GRIB file to be opened for reading.}

\item{var}{GRIB variable to be extracted.}

\item{levType}{Level type to extract GRIB data. Examples of
levType are "surface", "heightAboveGround","cloudTop"}

\item{level}{GRIB level to be extracted.}

\item{nx}{Integer number of GRIB data columns.}

\item{ny}{Integer number of GRIB data rows.}

\item{numFTimes}{Optional number of forecast times in
GRIB file. If not specified, defaults to 1.}
}
\value{
dataOut List of metadata and data to user.
}
\description{
\code{ExtractGRIBGrid} Extract grid of GRIB data.
}
\examples{
datOut <- ExtractGRIBGrid('rap.t01z.awp130bgrbf16.grib2','acpcp',
           'surface',0,451,337)
}
\concept{
GRIB
}
\keyword{GRIB}
\keyword{IO}

