% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{ExtractHucMeta}
\alias{ExtractHucMeta}
\title{For an indivudal product, grab the HUC metadata to be stored in the database.}
\usage{
ExtractHucMeta(hucProdDf)
}
\arguments{
\item{hucProdDf}{dataframe from dataRetrieval::readNWISuv}
}
\value{
list of metadata for a
}
\description{
\code{ExtractHucMeta} extracts the metadata to stash in a usgsDataRetrieval database for an individual product.
It gathers the attributes returned by dataRetrieval::readNWISuv , supplements siteInfo with
startTime and endTime for each station in UTC to assist appending the records.
}

