% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FillOutliers}
\alias{FillOutliers}
\title{Fill outliers based on change between steps}
\usage{
FillOutliers(x, thresh)
}
\arguments{
\item{x}{Vector of values}

\item{thresh}{Threshold for change between timesteps
(in absolute value of x units)}
}
\value{
Vector of values with outliers replaced by
previous value
}
\description{
\code{FillOutliers} is a simple utility
to fill outliers in a time series. Values that are 
more than some threshold change from the previous
time step are replaced by the value from that
previous time step.
}
\keyword{internal}
\keyword{utilities}

