% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{FindUsgsStns}
\alias{FindUsgsStns}
\title{Discover USGS stations using huc8 code or lat/lon/radius.}
\usage{
FindUsgsStns(stnLon = NULL, stnLat = NULL, within = NULL, huc8 = NULL,
  siteType = "ST", hasDataTypeCd = "iv")
}
\arguments{
\item{stnLon}{optional}

\item{stnLat}{optional}

\item{within}{optional, goes with stnLon and stnLat and specifies a search radius in decimal degrees.}

\item{huc8}{optional EIGHT digit HUC code.}

\item{siteType}{the type of USGS site to look for.}

\item{hasDataTypeCd}{the kind of data of interest (iv=instantaneous, dv=daily, etc.)}
}
\description{
\code{FindUsgsStns} wraps \code{dataRetrieval::whatNWISsites} with common options for finding USGS gauges.
See the dataRetrieval package for more deails on what else could be passed or how this could be modified.
}
\examples{
stnDf <- FindUsgsStns(huc='10190005')
stnDf <- FindUsgsStns(stnLon=254.67374999999998408,stnLat=40.018666670000001773,within=.001)
}

