% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{FlipLR}
\alias{FlipLR}
\title{Flip a matrix from left to right.}
\usage{
FlipLR(matrix)

FlipLR(matrix)
}
\arguments{
\item{matrix}{A matrix.}

\item{matrix}{A matrix.}
}
\description{
\code{flipLR} Flips a matrix from left to to right.

\code{FlipLR} Flips a matrix from left to to right.
}
\examples{
x <- matrix[1:9,3]
x
flipLR(x)
x <- matrix(1:9,3)
x
FlipLR(x)
}
\keyword{internal}

