% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{FlipUD}
\alias{FlipUD}
\title{Flip a matrix vertically, that is over it's horizontal axis (y1 <-> yn).}
\usage{
FlipUD(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
This is the transformation which 2D WRF Hydro fields need upon import.
\code{FlipVert} Flips a matrix vertically (over it's horizontal axis).
}
\examples{
x <- matrix(1:9, 3)
x
FlipUD(x)
}
\keyword{internal}

