% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{FormalsToDf}
\alias{FormalsToDf}
\title{General purpose helper for handling vector arguments.}
\usage{
FormalsToDf(theFunc, envir = parent.frame())
}
\description{
General purpose helper for handling vector arguments.
}
\keyword{internal}
\keyword{util}

