% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gages_2.R
\docType{data}
\name{gages2Attr}
\alias{gages2Attr}
\title{Gages-II attributes}
\format{data.frame:  9322 obs. of  16 variables:
\describe{
 \item{X}{}
 \item{Y}{}
 \item{STAID}{}
 \item{STANAME}{}
 \item{CLASS}{}
 \item{AGGECOREGI}{}
 \item{DRAIN_SQKM}{}
 \item{HUC02}{}
 \item{LAT_GAGE}{}
 \item{LNG_GAGE}{}
 \item{STATE}{}
 \item{HCDN_2009}{}
 \item{ACTIVE09}{}
 \item{FLYRS1900}{}
 \item{FLYRS1950}{}
 \item{FLYRS1990}{}
}}
\usage{
gages2Attr
}
\description{
These are the attributes taken from the USGS Gages-II data set. \cr
\link[http://water.usgs.gov/GIS/metadata/usgswrd/XML/gagesII_Sept2011.xml]{http://water.usgs.gov/GIS/metadata/usgswrd/XML/gagesII_Sept2011.xml}
}
\section{Usage}{

 \code{gages2Attr} \cr
 \code{data(gages2Attr)}
}

\section{Citation}{

Originator: James Falcone \cr
Publication_Date: 20111012 \cr
Title: GAGES-II: Geospatial Attributes of Gages for Evaluating Streamflow \cr
Geospatial_Data_Presentation_Form: vector digital data
Publication_Information: \cr
Publication_Place: Reston, Virginia \cr
Publisher: U.S. Geological Survey \cr
Online_Linkage: \link[http://water.usgs.gov/lookup/getspatial?gagesII_Sept2011]{http://water.usgs.gov/lookup/getspatial?gagesII_Sept2011} \cr
}
\concept{
data
}
\keyword{data}

