% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_get.R
\name{GapFillRS}
\alias{GapFillRS}
\title{Gap fill a series of MODIS images using interpolation.}
\usage{
GapFillRS(inPath, matchStr, outDir, begin = NULL, end = NULL,
  maskFile = NULL, maskVals = NULL, nodataVal = -9999,
  scriptPath = NULL, pos1 = 10, pos2 = 16, format = "\%Y\%j")
}
\arguments{
\item{inPath}{The path to the directory that holds the already processed
MODIS TIF files.}

\item{matchStr}{The regular expression for filename matching (e.g.,
"*Lai_1km.tif").}

\item{outDir}{Pathname to the directory to store the output files.}

\item{begin}{Date string for the start date to include. The date string
should follow \href{http://r-forge.r-project.org/projects/modis/}{MODIS}
package convention (e.g., "2011.06.01"). (DEFAULT=NULL, all files are
processed).}

\item{end}{Date string for the end date to include. The date string should
follow \href{http://r-forge.r-project.org/projects/modis/}{MODIS} package
convention (e.g., "2011.06.01"). (DEFAULT=NULL, all files are processed).}

\item{maskFile}{Mask grid file. File must be in a GDAL-friendly format and 
should match the dimensions and resolution of the inputs.}

\item{maskVals}{List of value sets to pull from mask file and convert. 
The list should be in the form: list("maskValue"=outputValue), 
e.g., list("1"=0, "16"=0) to convert all cells in the mask grid with mask 
values of 1 and 16 to 0 in the output files.}

\item{nodataVal}{Value to use to represent no data. Make sure this value 
is NOT in your mask raster. (DEFAULT=-9999)}

\item{scriptPath}{OPTIONAL path to where gdal_calc.py script lives in case
it is not in the same default path as gdal executables}

\item{pos1}{From MODIS orgTime: Start position of date in the filename (DEFAULT=10).}

\item{pos2}{From MODIS orgTime: End position of date in the filename (DEFAULT=16).}

\item{format}{From MODIS orgTime: How is the date formatted in the file, default 
expects: YYYYDDD (\%Y\%j). Read ?as.Date for for more information.}
}
\value{
null
}
\description{
\code{GapFillRS} takes a set of pre-processed MODIS TIF files and
fills no-data gaps using interpolation.
}
\details{
\code{GapFillRS} scans the specified directory and runs the GDAL python
script gdal_fillnodata.py. If a mask file is provided, GapFillRS will
convert specified mask cells to specified values AFTER the gap-filling.
}
\examples{
## Gap fill a set of MODIS LAI images. We will convert all non-vegetation 
## land use types into 0.

\dontrun{ 
maskVals <- list("1"=0, "16"=0, "19"=0, "23"=0, 
                 "24"=0, "25"=0, "26"=0, "27"=0)
GapFillRS("~/wrfHydroTestCases/MODIS_ARC/PROCESSED/FOURMILE_LAI",
                         outDir="~/wrfHydroTestCases/MODIS_ARC/PROCESSED/FOURMILE_LAI_GAPFILL",
                         matchStr="*Lai_1km.tif", 
                         begin="2013.05.01", end="2013.07.31", 
                         maskFile="~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/geo_LUINDEX.tif",
                         maskVals=maskVals)
}
}
\seealso{
Other MODIS: \code{\link{CalcStatsRS}},
  \code{\link{ConvertRS2Stack}},
  \code{\link{ConvertStack2NC}}, \code{\link{GetMODIS}},
  \code{\link{InsertRS}}, \code{\link{SmoothStack}}
}
\concept{
MODIS dataMgmt
}
\keyword{IO}
