% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gage_gage_dist_gridded.R
\name{GatherStream}
\alias{GatherStream}
\title{Gather upstream or downstream distance from a given starting location}
\usage{
GatherStream(stream, length, start, indDist = list(ind = c(), dist = c()))
}
\arguments{
\item{stream}{list of stream information containing either from/to and start and end positions}

\item{length}{vector of lengths for each re-indexed reach}

\item{start}{Indexed location (NOT comID) of where stream starts}
}
\value{
List containing indices and accumulated distance from start
}
\description{
Gather upstream or downstream distance from a given starting location
}

