% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_mod.R
\name{genWghtFile}
\alias{genWghtFile}
\title{High level call to generate regridding weight NetCDF file.}
\usage{
genWghtFile(geoFile, nxIn, nyIn, latSrc, lonSrc, method, srcDummy, ndv,
  wghtFile)
}
\arguments{
\item{geoFile}{geogrid file of WRF-Hydro domain.}

\item{nxIn}{Integer number of columns of source data.}

\item{nyIn}{Integer number of rows of source data.}

\item{latSrc}{Grid of center-stagger latitude values for source data.}

\item{lonSrc}{Grid of center-stagger longitude values for soure data.}

\item{method}{Integer of regridding method. Options are 2 - "conserve",
1 - "bilinear", 3 - "nneighbor".}

\item{srcDummy}{Array of sample source data to be used in generation of
weights. This will be used to generate the mask.}

\item{ndv}{Source NDV value used in masking.}

\item{wghtFile}{Output weight file to be generated.}
}
\description{
\code{genWghtFile} High level call to generate regridding
weight file.
}
\examples{
\dontrun{
genWghtFile('./geo_em.d02.nc',nxIn,nyIn,latGRIB,lonGRIB,1,dataIn[,,],
9999,'./wght_hrrr_rio_grande.nc')
}
}
\seealso{
Other regrid: \code{\link{regrid}}
}
\concept{
regrid geospatial
}
\keyword{ESMF}
\keyword{geospatial}
\keyword{io}
\keyword{regrid}

