% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geospatial.R
\name{geogridSpatialInfo}
\alias{geogridSpatialInfo}
\title{Pull necessary geospatial information from geogrid file used for
regridding and deprojection.}
\usage{
geogridSpatialInfo(geoFile)
}
\arguments{
\item{geoFile}{The geogrid file information is being pulled from}
}
\value{
spatialDF data frame containing geospatial information
}
\description{
\code{geogridSpatialInfo} Pull geospatial information about WRF-Hydro
 modeling domain from geogrid file.
}
\examples{
\dontrun{
lccGeoDF <- geogridSpatialInfo('./geo_em.d02.nc')
}
}
\concept{
GEOSPATIAL
}
\keyword{GEOSPATIAL}
\keyword{METADATA}

