% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwmChannelOut.R
\name{get_nwmChannelOut}
\alias{get_nwmChannelOut}
\title{Stash NWM channel data in a SQLlite database}
\usage{
get_nwmChannelOut(files, features = NULL, variables = c("streamflow",
  "velocity", "nudge", "q_lateral"))
}
\arguments{
\item{files}{A character vector of file paths to CHRTOUT files.}

\item{features}{Optional, A numeric vector of features for subsetting channel data prior to storing in database.}

\item{variables}{Optional, character vector of variabel names to extract}
}
\value{
data.frame
}
\description{
Stash NWM channel data in a SQLlite database
}
\details{
This function can only be used for NWM model output from versions 1.1-1.2. 
Given a character vector of paths to channel route netcdf files, \code{get_nwmChannel} will extract selected data and return a dataframe.
}
\examples{
\dontrun{
TODO - 1333
}
}
