% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameriflux_get.R
\name{GetAmeriflux}
\alias{GetAmeriflux}
\title{Get Ameriflux data for specified sites.}
\usage{
GetAmeriflux(siteIDs, gaps = TRUE, timeZones = NULL, utcOffsets = NULL,
  startYr = NULL, endYr = NULL)
}
\arguments{
\item{siteIDs}{A single site ID or vector of site IDs to download and process. Site 
IDs should match the standardized AmerifluxIDs (e.g., "US-NR1" for Niwot Ridge). 
See amfMeta$id_txt for a list of site IDs (as of June 2015).}

\item{gaps}{(DEFAULT=TRUE) Boolean flag whether to download data with gaps 
(TRUE) or with gaps filled (FALSE). Gap-filled data is not available for all 
sites.}

\item{timeZones}{(OPTIONAL) A single time zone or vector of time zones for each
site provided in the siteIDs. Note that Ameriflux timestamps are (allegedly!)
in local time WITHOUT daylight savings time adjustments. If neither timeZones nor
utcOffsets are provided, the tool will attempt to lookup the site in the master
database (amfMeta) and get a UTC adjustment from there. These offsets were
estimated based on site lat/lon coordinates and are provided with a 
use-at-your-own-risk(!) caveat.}

\item{utcOffsets}{(OPTIONAL) A single UTC offset or vector of UTC offsets for each 
site provided in the siteIDs. These should be in positive or negative hours 
(e.g., -7 for "America/Denver" with no DST shift). These values will supercede 
any provided time zones. See tzLookup for a list of time zone UTC offsets. If 
neither timeZones nor utcOffsets are provided, the tool will attempt to lookup the 
site in the master database (amfMeta) and get a UTC adjustment from there. These 
offsets were estimated based on site lat/lon coordinates and are provided with a 
use-at-your-own-risk(!) caveat.}

\item{startYr}{(OPTIONAL) The first year to download files for.}

\item{endYr}{(OPTIONAL) The last year to download files for.}
}
\value{
dataframe
}
\description{
\code{GetAmeriflux} downloads Ameriflux data tables (Level 2 standardized CSV
files) and creates a dataframe. Based on Koen Hufkens's Ameriflux download tool.
}
\details{
\code{GetAmeriflux} downloads Ameriflux Level 2 standardized CSV files for
specified sites and outputs a dataframe with consistent date and data columns 
for use with other rwrfhydro tools. Borrowed generously from Koen Hufkens's
Ameriflux download tool: 
\link[https://bitbucket.org/khufkens/ameriflux-download-tool]{https://bitbucket.org/
khufkens/ameriflux-download-tool}
}
\examples{
\dontrun{
amf.An <- GetAmeriflux(c("US-An1","US-An2","US-An3"), startYr=2009, endYr=2010)
}
}
\seealso{
Other Ameriflux: \code{\link{amfMeta}},
  \code{\link{tzLookup}}
}
\concept{
Ameriflux
}
\keyword{IO}
