% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_strflow_get.R
\name{GetCoDwrData}
\alias{GetCoDwrData}
\title{Get CO DWR data for specified sites.}
\usage{
GetCoDwrData(siteIDs, paramCodes, timeInt = 1, startDate = NULL,
  endDate = NULL)
}
\arguments{
\item{siteIDs}{A single site ID or vector of site IDs to download and process.
Site IDs should match the standardized CO DWR IDs (e.g., "RIOALACO" for RIO  
GRANDE RIVER AT ALAMOSA).}

\item{paramCodes}{A single or list of parameter codes to download. CO DWR
parameters include (list not inclusive) GAGE_HT, DISCHRG, AIRTEMP, COND, DO,  
DO_SAT, PH, PRECIP, TDS, TURBID, WATTEMP, WATTEMP2.}

\item{timeInt}{Time interval to download. Options are "raw", "hourly", or "daily".
(DEFAULT="raw"). Note that the hourly and daily options are means.}

\item{startDate}{(OPTIONAL) The start date to download for (format "mm/dd/yy").}

\item{endDate}{(OPTIONAL) The end date to download for (format "mm/dd/yy").}
}
\value{
dataframe
}
\description{
\code{GetCoDwrData} downloads and processes CO DWR data from web feed.
}
\details{
\code{GetCoDwrData} downloads Colorado Department of Water Resources data for
specified sites and outputs a dataframe with consistent date and data columns 
for use with other rwrfhydro tools.
}
\examples{
\dontrun{}
str.RIO <- GetCoDwrData(c("RIOSFKCO", "RIOWAGCO","RIODELCO"), 
           paramCodes=c("GAGE_HT", "DISCHRG"), timeInt="hourly", 
           startDate="06/01/14", endDate="06/30/14")
}
\concept{Streamflow}
\keyword{IO}
