% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_USCRN.R
\name{GetCRN}
\alias{GetCRN}
\title{Get the US Climate Reference Network data from the NCEP ftp server.}
\usage{
GetCRN(gageNames = NULL, siteIds = NULL, elements = NULL,
  timestep = "hourly", domain = NULL, startDate = Sys.time() - 10 *
  24 * 3600, endDate = Sys.time(), realTime = FALSE,
  parallel = FALSE)
}
\arguments{
\item{gageNames}{Character vector of the gage names which is used in the file names (or any part of the file name)}

\item{siteIds}{Characcter vector of the stations IDs so called WBANNO.}

\item{elements}{Character vector containing all elements required from the data.}

\item{timestep}{Character (DEFAULT = "hourly"). It can be subhourly, hourly, daily and monthly}

\item{domain}{data.frame, containing four columns named lonmax, latmin, latmax, latmin. 
It will be used to subset the data to only those gauges falling into the rectangle domain}

\item{startDate}{POSIXct, ending date}

\item{endDate}{POSIXct, strating date}

\item{realTime}{Logical (DEFAULT = FALSE)}

\item{parallel}{Logical (DEFAULT = FALSE)}
}
\value{
A data.table containing the US CRN data.
}
\description{
\code{GetCRN} Given the information provided to the function will download the data from NCEP ftp serverr
and then subset the data to meet the criteria defined such as being in a domian or being in specific time period. 
Return a dataframe containing.
}
\details{
The contiguous U.S. network of 114 stations was completed in 2008. 
There are two USCRN stations in Hawaii and deployment of a network of 29 stations in Alaska continues. 
The vision of the USCRN program is to maintain a sustainable high-quality climate observation network.

These data are provided in 4 different temporal resolution (subhourly, hourly, daily and monthly),
and depending on the temporal resolution, the variables provided changes. Below is a summary of all the 
variables in each temporal resolution and their units.

subhourly data contains: \itemize{
\item chech this README.txt \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/subhourly01/README.txt}
\item AIR_TEMPERATURE                Celsius
\item PRECIPITATION                  (mm)
\item SOLAR_RADIATION                (W/m^2)
\item SR_FLAG                        (X)
\item SURFACE_TEMPERATURE            (Celsius)
\item ST_TYPE                        (X)
\item ST_FLAG                        (X)
\item RELATIVE_HUMIDITY              (\%)
\item RH_FLAG                        (X)
\item SOIL_MOISTURE_5                (m^3/m^3)
\item SOIL_TEMPERATURE_5             (Celsius)
\item WETNESS                        (Ohms)
\item WET_FLAG                       (X)
\item WIND_1_5                       (m/s)
\item WIND_FLAG                      (X)
}

hourly data contains: \itemize{
\item chech this README.txt \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/hourly02/README.txt}
\item T_CALC                         (Celsius)
\item T_HR_AVG                       (Celsius)
\item T_MAX                          (Celsius)
\item T_MIN                          (Celsius)
\item P_CALC                         (mm)
\item SOLARAD                        (W/m^2)
\item SOLARAD_FLAG                   (X)
\item SOLARAD_MAX                    (W/m^2)
\item SOLARAD_MAX_FLAG               (X)
\item SOLARAD_MIN                    (W/m^2)
\item SOLARAD_MIN_FLAG               (X)
\item SUR_TEMP_TYPE                  (X)
\item SUR_TEMP                       (Celsius)
\item SUR_TEMP_FLAG                  (X)
\item SUR_TEMP_MAX                   (Celsius)
\item SUR_TEMP_MAX_FLAG              (X)
\item SUR_TEMP_MIN                   (Celsius)
\item SUR_TEMP_MIN_FLAG              (X)
\item RH_HR_AVG                      (\%)
\item RH_HR_AVG_FLAG                 (X)
\item SOIL_MOISTURE_5                (m^3/m^3)
\item SOIL_MOISTURE_10               (m^3/m^3)
\item SOIL_MOISTURE_20               (m^3/m^3)
\item SOIL_MOISTURE_50               (m^3/m^3)
\item SOIL_MOISTURE_100              (m^3/m^3)
\item SOIL_TEMP_5                    (Celsius)
\item SOIL_TEMP_10                   (Celsius)
\item SOIL_TEMP_20                   (Celsius)
\item SOIL_TEMP_50                   (Celsius)
\item SOIL_TEMP_100                  (Celsius)
}

daily data contains: \itemize{
\item chech this README.txt \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/daily01/README.txt}
\item T_DAILY_MAX                    (Celsius)
\item T_DAILY_MIN                    (Celsius)
\item T_DAILY_MEAN                   (Celsius)
\item T_DAILY_AVG                    (Celsius)
\item P_DAILY_CALC                   (mm)
\item SOLARAD_DAILY                  (MJ/m^2)
\item SUR_TEMP_DAILY_TYPE            (X)
\item SUR_TEMP_DAILY_MAX             (Celsius)
\item SUR_TEMP_DAILY_MIN             (Celsius)
\item SUR_TEMP_DAILY_AVG             (Celsius)
\item RH_DAILY_MAX                   (\%)
\item RH_DAILY_MIN                   (\%)
\item RH_DAILY_AVG                   (\%)
\item SOIL_MOISTURE_5_DAILY          (m^3/m^3)
\item SOIL_MOISTURE_10_DAILY         (m^3/m^3)
\item SOIL_MOISTURE_20_DAILY         (m^3/m^3)
\item SOIL_MOISTURE_50_DAILY         (m^3/m^3)
\item SOIL_MOISTURE_100_DAILY        (m^3/m^3)
\item SOIL_TEMP_5_DAILY              (Celsius)
\item SOIL_TEMP_10_DAILY             (Celsius)
\item SOIL_TEMP_20_DAILY             (Celsius)
\item SOIL_TEMP_50_DAILY             (Celsius)
\item SOIL_TEMP_100_DAILY            (Celsius)
}

monthly data contains: \itemize{
\item chech this README.txt \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/monthly01/README.txt}
\item T_MONTHLY_MAX                  (Celsius)
\item T_MONTHLY_MIN                  (Celsius)
\item T_MONTHLY_MEAN                 (Celsius)
\item T_MONTHLY_AVG                  (Celsius)
\item P_MONTHLY_CALC                 (mm)
\item SOLRAD_MONTHLY_AVG             (MJ/m^2)
\item SUR_TEMP_MONTHLY_TYPE          (X)
\item SUR_TEMP_MONTHLY_MAX           (Celsius)
\item SUR_TEMP_MONTHLY_MIN           (Celsius)
\item SUR_TEMP_MONTHLY_AVG           (Celsius)
}
}
\examples{
\dontrun{
# example 1: getting the precipition data at the subhourly time steps, 
# for the first three months of 2016
startDate = as.POSIXct("2016-01-01 00:00:00", format= "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
endDate   = as.POSIXct("2016-04-01 00:00:00", format= "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
output <- GetCRN(elements = "P_CALC", timestep = "hourly", 
startDate = startDate, endDate= endDate)

# choose only gauge in the following domain
domain <- data.frame(minlon = -80, maxlon = -60, minlat = 30, maxlat = 50)
output <- GetCRN(elements = "P_CALC", timestep = "hourly", 
startDate = startDate, endDate= endDate, domain = domain)

# choose only gauges in the Colorado 
outputHourly <- GetCRN(gageNames = "CO", elements = "P_CALC", timestep = "hourly", 
startDate = startDate, endDate= endDate)

# get daily data only for basins in colorado
outputDaily <- GetCRN(gageNames = "CO", elements = "P_DAILY_CALC", timestep = "daily", 
startDate = startDate, endDate= endDate)

# get monthly data only for basins in colorado
outputMonthly <- GetCRN(gageNames = "CO", elements = "P_MONTHLY_CALC", timestep = "monthly", 
startDate = startDate, endDate= endDate)

# get subhourly data only for basins in colorado
outputsubhourly <- GetCRN(gageNames = "CO", elements = "PRECIPITATION", 
timestep = "subhourly", startDate = startDate, endDate= endDate)

# get hourly soik moisure data
outputHourly <- GetCRN(gageNames = "CO", elements = c("SOIL_MOISTURE_5", 
"SOIL_MOISTURE_10", "SOIL_MOISTURE_20", "SOIL_MOISTURE_50", "SOIL_MOISTURE_100"), 
timestep = "hourly")
}
}
\concept{USCRN}
\keyword{IO}
