% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_domain.R
\name{GetDomainCoordsProj}
\alias{GetDomainCoordsProj}
\title{\code{GetDomainCoordsProj} collects and reprojects spatial data from geo and hydro files.}
\usage{
GetDomainCoordsProj(file, proj4Str = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{file}{A path/name to a geo or hydro domain file.}

\item{proj4Str}{the desired output projection.}
}
\value{
An sp::SpatialPoints object.
}
\description{
Handle the projection and datum iformation associated with WRF-Hydro geo and hydro grid files.
}
\seealso{
Other domain: \code{\link{VisualizeChanNtwk}},
  \code{\link{VisualizeDomain}}
}
\concept{
plot
}
\keyword{hplot}

