% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ghcn.R
\name{GetGhcn}
\alias{GetGhcn}
\title{Get GHCN data for specified sitesId.}
\usage{
GetGhcn(siteIds, elements, startDate = NULL, endDate = NULL,
  parallel = FALSE, fileAdd = NULL)
}
\arguments{
\item{siteIds}{A single siteId or vector of siteIds to download and process.
SiteIds should match the standardized GHCN IDs (for example : ACW00011604).
See \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}
for a list of siteIds.}

\item{elements}{A Character vector defining what type of observation you
are interested in. There are five core elements as well as a number of
addition elements. The five core elements are:
\describe{
\item{PRCP}{ = Precipitation (tenths of mm)}
\item{SNOW}{ = Snowfall (mm)}
\item{SNWD}{ = Snow depth (mm)}
\item{MAX}{ = Maximum temperature (tenths of degrees C)}
\item{TMIN}{ = Minimum temperature (tenths of degrees C)}
}
For the full list of elemenst refer to
\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}}

\item{startDate,endDate}{Date.}

\item{parallel}{Logical}

\item{fileAdd}{Address to the url containg all the daily GHCN data,
default is \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/all/}}
}
\value{
A dataframes containing the date and the correponding daily GHCN value and the qFlag value.
}
\description{
\code{\link{GetGhcn}} downloads the daily GHCN (Global Historic Climatology Network)
 data for each siteIds, and creates a dataframe containing four fields of siteIds, date,
 daily GHCN value and the qFlag. If there are so many gauges, then
 \code{\link{GetGhcn2}} would be much faster.
}
\examples{
siteIds=c("ACW00011604","AJ000037579","AJ000037883","ASN00005095")
startDate="1949/02/01"
endDate="1949/10/01"
element="PRCP"
obsPrcp<-GetGhcn(siteIds,element,startDate,endDate,parallel=FALSE)

Or you could use the results of the SelectGhcnGauges

countryCodeList=c("US")
networkCodeList=c("1")
statesList=c("WY")
selectedGauges<-SelectGhcnGauges(countryCode=countryCodeList,
                                 networkCode=networkCodeList,
                                 states=statesList)
obsPrcp<-GetGhcn(selectedGauges$siteIds,element,startDate,endDate,parallel=FALSE)
}

