% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/latest_version.R
\name{GetGithubPat}
\alias{GetGithubPat}
\title{Retrieve Github personal access token.
Looks in env var \code{GetGithubPat}.}
\usage{
GetGithubPat()
}
\description{
Retrieve Github personal access token.
Looks in env var \code{GetGithubPat}.
}
\keyword{internal}

