% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_multi_ncdf.R
\name{GetMultiNcdf}
\alias{GetMultiNcdf}
\title{Get WRF Hydro output/restart timeseries spread over multiple files.}
\usage{
GetMultiNcdf(filesList, variableList, indexList, parallel = FALSE)
}
\arguments{
\item{filesList}{The list of file groups. Names must match those in the other lists.}

\item{variableList}{The list of variables for each file group. Names must match filesList.}

\item{indexList}{The list of indices or statistics to be applied to each variable.}

\item{parallel}{Logical, this is the .parallel argument of plyr functions.
Parallelization is at the file level (not file group).Typcially we achieve
parallelization using the DoMC package. See examples.}
}
\value{
A dataframe in an awesome format (, Aubrey!).
}
\description{
\code{GetMultiNcdf} is designed to get *all* your output/restart data which
are spread over multiple files. Three collated lists specify
1) file groups, 2) variables for each file group, and 3) indices or statistics
for each variable in each file group. The names of the lists must match. See
examples for details.
}
\examples{
This is to do.
}

