% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_multi_ncdf.R
\name{GetMultiNcdfFile}
\alias{GetMultiNcdfFile}
\title{The GetMultiNcdf for individual file groups.}
\usage{
GetMultiNcdfFile(filesInd, filesList, variableList, indexList,
  env = parent.frame(), parallel = FALSE)
}
\arguments{
\item{filesInd}{The index of the file groups.}

\item{filesList}{The list of the file groups.}

\item{variableList}{The variable list.}

\item{indexList}{The index list.}

\item{env}{The environment where the stat function lives}

\item{parallel}{Logical, this is the .parallel argument of plyr functions.}
}
\value{
A dataframe
}
\description{
\code{GetMultiNcdfFile}: For this file group, get each files variable.
Typically calls GetMultiNcdfVariable to then get the individual files.
Typically called by GetMultiNcdf.
}
\seealso{
Other getMultiNcdf: \code{\link{GetFileStat}};
  \code{\link{GetMultiNcdfVariable}}
}
\concept{
dataGet
}
\keyword{internal}

