% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_multi_ncdf.R
\name{GetMultiNcdfVariable}
\alias{GetMultiNcdfVariable}
\title{The GetMultiNcdf for getting variables out of individual files.}
\usage{
GetMultiNcdfVariable(varInd, indexList, variableList, files, parallel = FALSE)
}
\arguments{
\item{varInd}{The variable index.}

\item{indexList}{The variable list.}

\item{variableList}{The variable list.}

\item{files}{The files vector.}

\item{parallel}{Logical, this is the .parallel argument of plyr functions.}
}
\value{
A dataframe
}
\description{
\code{GetMultiNcdfVariable}: For a file, get all variables.
Typically calls GetFileStat to get the individual variables.
Typically called by GetMultiNcdfFile.
}

