% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldas2_get.R
\name{GetNldas2}
\alias{GetNldas2}
\title{Get NLDAS2 forcings or model outputs.}
\usage{
GetNldas2(prodName, timeInt, outDir, startDate, endDate, quiet = TRUE)
}
\arguments{
\item{prodName}{Product to download. Options are: \itemize{
\item "FORA" (primary forcings)
\item "FORB" (secondary forcings)
\item "MOS" (Mosaic model output)
\item "NOAH" (Noah model output)
\item "VIC" (VIC model output)}}

\item{timeInt}{Time interval. Options are: \itemize{
\item "H" (hourly)
\item "M" (monthly)
\item "MC" (monthly climatology)}}

\item{outDir}{Output directory. This directory will be checked to see if the file
already exists.}

\item{startDate}{Date to start downloading data}

\item{endDate}{Date to end downloading data}

\item{quiet}{Flag whether or not to output messages (DEFAULT=TRUE)}
}
\value{
raster stack object
}
\description{
\code{GetNldas2} downloads NLDAS-2 grids and creates a raster stack.
}
\details{
\code{GetNldas2} downloads NLDAS-2 for specified time period
and outputs a raster stack.
}
\examples{
sno.santafe <- GetSnotel(c("921","922"), series="Daily", startYr=2009, endYr=2010)
sno.niwot <- GetSnotel(c("663"), series="Hourly", startYr=2009, month=4)
sno.santafe <- GetSnotel(c("921","922"), intervaltype="Current", current="WEEK")
}
\concept{
NLDAS
}
\keyword{IO}

