% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nohrsc_get.R
\name{GetNWCSnowPoints}
\alias{GetNWCSnowPoints}
\title{Get and unpack National Water Center SWE and snow depth 
point observations for a given hour.}
\usage{
GetNWCSnowPoints(datePOSIXct, outDir = ".", overwrite = FALSE,
  quiet = TRUE)
}
\arguments{
\item{datePOSIXct}{the date in POSIXct format for which the
data is desired. The format needs to be YYYY-MM-DD HH:MM:SS.}

\item{outDir}{The directory where the data is to be archived.
This directory is checked to see if the data exist when
overwrite=FALSE}

\item{overwrite}{when false: If depth and SWE files exist on
disk, don't grab the zip file. \cr When false: If the zip
file exists on disk but depth and SWE files don't, just unzip
the zip files. \cr When true: Pull new zip and overwrite any
existing files with the same date.}

\item{quiet}{logical passed to curl , to show it's progress.}
}
\value{
Logical was the file "got".
}
\description{
\code{GetNWCSnowPoints} Get and unpack National Water Center
snow observations for a given hour and unpack the data. Data
is stored in text file format.
}
\examples{
\dontrun{
NWCSnowGot <- GetNWCSnowPoints(as.POSIXct('2015-02-01 12:00:00'),
 outDir = '/home/karsten/snow_obs')
 }
}
\concept{
dataGet
}
\keyword{IO}

