% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetPkgDataPath}
\alias{GetPkgDataPath}
\title{GetPkgDataPath}
\usage{
GetPkgDataPath(theFile)
}
\arguments{
\item{file}{The external data file to load (this is in dirrefent places before, rwrfhydro/inst/extdata, and after build, rwrfhydro/).}
}
\value{
The full path to the file.
}
\description{
\code{GetPkgData} is a simplified wrapper (for system.file) for loading external rwrfhydro data included with the package.
}
\examples{
GetPkgDataPath('Fourmile_test_case_AD.hydro_OrodellBasin_100m_8.nc')
#'
}

