% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetPkgMeta}
\alias{GetPkgMeta}
\title{Get a package's metadata fields and associated entries.}
\usage{
GetPkgMeta(meta = c("concept", "keyword"), package = "rwrfhydro",
  keyword = NULL, concept = NULL, quiet = FALSE, listMetaOnly = FALSE,
  byFunction = "")
}
\arguments{
\item{meta}{Character the metadata field.}

\item{package}{Character The package to query for metadata.}

\item{keyword}{Character A specific keyword to look for.}

\item{concept}{Character A specific concept to look for.}

\item{quiet}{Logical Do not print summary to screen.}

\item{listMetaOnly}{Logical Just return the meta categories (without functons)?}

\item{byFunction}{Character Vector of functions for which concepts and keywords are desired.}
}
\value{
List of metadata fields in alphabetical order with corresponding entries.
}
\description{
\code{GetPkgMeta} Get metadata fields and associated entries from a package's documentation (e.g. "keyword" or "concepts".)
}
\examples{
GetPkgMeta()
GetPkgMeta('keyword', package='ggplot2')
GetPkgMeta(concept = 'foo', key='hplot' )
print( GetPkgMeta(concept = c('dataMgmt','foo'), key=c('foo','hplot') , quiet=TRUE))
str( GetPkgMeta(concept = c('dataMgmt','foo'), key=c('foo','hplot') , quiet=TRUE))
str( GetPkgMeta(concept = c('dataMgmt','DART'), key=c('internal','hplot') , quiet=TRUE))
GetPkgMeta(byFunction=c('MkDischargeVariance','SaveHucData'))
}
\keyword{utilities}

