% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetPkgRawDataPath}
\alias{GetPkgRawDataPath}
\title{Simplifed loading of rwrfhydro data included with the package.}
\usage{
GetPkgRawDataPath(theFile = "")
}
\arguments{
\item{theFile}{The external data file to load (this is in dirrefent places
before, rwrfhydro/inst/extdata, and after build, rwrfhydro/).}
}
\value{
The full path to the file.
}
\description{
\code{GetPkgRawDataPath} is a simplified wrapper (for system.file) for
loading external rwrfhydro data included with the package.
}
\examples{
GetPkgRawDataPath('gagesII_all.csv')
}
\keyword{internal}

