% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poly.R
\name{GetPoly}
\alias{GetPoly}
\title{Return the polygon attribute for any point.}
\usage{
GetPoly(points, polygon = NULL,
  proj4 = "+init=epsg:4269 +proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs +towgs84=0,0,0",
  polygonAddress = NULL, polygonShapeFile = NULL, join = NULL,
  parallel = FALSE)
}
\arguments{
\item{points}{A dataframe of the points. The dataframe should contain at least two
fields: "latitude" and "longitude".}

\item{polygon}{SpatialPolygonDataFrame}

\item{proj4}{Character: Projection to be used for points's dataframe.}

\item{polygonAddress}{Characetr: Address to where polygonShapeFile is located.}

\item{polygonShapeFile}{Character: Name of a polygon shapefile.}

\item{join}{Character: Attribute from the polygon shapefile which will be added as a column to points dataframe}

\item{parallel}{Logical (DEFAULT= FALSE)}
}
\value{
Dataframe with an extra column containing the attribute for each point.
}
\description{
\code{GetPoly} takes a dataframe of points containing at least
two fields of latitude and longitude, overlays the points with a 
shapefile, and adds the attribute as a field to the input dataframe.
}
\examples{
\dontrun{
sg<-GetPoly(sg,  polygonAddress= "/glade/scratch/arezoo/QPF_verification_rwrfhydro/gis/",
polygonShapeFile= "huc6", join="HUC6")
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}},
  \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}},
  \code{\link{GetProj}}, \code{\link{GetRfc}},
  \code{\link{GetTimeZone}}, \code{\link{PolyToRaster}}
}
\concept{dataMgmt}
\concept{geospatial}
\keyword{IO}
