% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geospatial.R
\name{GetProj}
\alias{GetProj}
\title{Pull projection information from geogrid file}
\usage{
GetProj(geoFile)
}
\arguments{
\item{geoFile}{The geogrid file information is being pulled from}
}
\value{
Character Projection
}
\description{
\code{GetProj} Pull projection information of WRF-Hydro
 modeling domain from geogrid file.
}
\examples{
\dontrun{
proj4 <- GetProj('./geo_em.d02.nc')
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetRfc}};
  \code{\link{GetTimeZone}}; \code{\link{PolyToRaster}}
}
\concept{
geospatial getData
}
\keyword{IO}

