% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poly.R
\name{GetRfc}
\alias{GetRfc}
\title{Return the RFC name for any point.}
\usage{
GetRfc(points,
  proj4 = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  parallel = FALSE)
}
\arguments{
\item{points}{A dataframe of the points. The dataframe should contain at least two
fields: "latitude" and "longitude".}

\item{proj4}{Character: Projection to be used for points's dataframe.
(DEFAULT = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")}

\item{parallel}{Logical (DEFAULT= FALSE)}
}
\value{
Dataframe with an extra column containing the RFC name.
}
\description{
\code{GetRfc} takes a dataframe of points containing at least
two fields of latitude and longitude, overlays the points with RFC
shapefile, and adds the RFC name as a field to the input dataframe.
}
\examples{
\dontrun{
sg <- SelectGhcnGauges(countryCode="US",networkCode="C")[1:100,]
sg <- GetRfc(sg)
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}},
  \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}},
  \code{\link{GetPoly}}, \code{\link{GetProj}},
  \code{\link{GetTimeZone}}, \code{\link{PolyToRaster}}
}
\concept{
dataMgmt
}
\keyword{IO}
