% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{getSnodasDepthSweDate}
\alias{getSnodasDepthSweDate}
\title{GetSnodasDepthSweDate}
\usage{
getSnodasDepthSweDate(datePOSIXct, outputDir = ".", overwrite = FALSE,
  quiet = TRUE)
}
\arguments{
\item{datePOSIXct}{The date in POSIXct format for which data is desired. Only whole days matter.}

\item{outputDir}{The directory where the data are to be archived. This directory is checked to see if the
data exist when overwrite=FALSE}

\item{overwrite}{When false: If the depth and SWE files exist on disk, dont grab the tarball. \cr
When false: If the tarball exists on disk but depth and SWE files dont, just unpack the tarball. \cr
When true: Pull new tarball and overwrite any existing files with the same date.}

\item{quiet}{Passed to curl, to show it's progress (typcially too fast to matter).}
}
\value{
Success or failure.
}
\description{
\code{GetSnodasDepthSweDate} Get and unpack the SNODAS snow depth and SWE tarball for a given date.
}
\examples{
snodasGot <- getSnodasDepthSweDate(as.POSIXct('2015-02-28'))
}

