% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snodas.R
\name{GetSnodasPointTs}
\alias{GetSnodasPointTs}
\title{Pull SNODAS snowdepth and SWE time series for a given lat/lon pair.}
\usage{
GetSnodasPointTs(bDatePOSIXct, eDatePOSIXct, snodasDir, lat, lon,
  quiet = TRUE)
}
\arguments{
\item{bDatePOSIXct}{The date in POSIXct format for the beginning of
the time series.}

\item{eDatePOSIXct}{The date in POSIXct format for the ending of
the time series.}

\item{snodasDir}{The directory containing SNODAS NetCDF files.}

\item{lat}{The latitude of the point of interest to pull data for.}

\item{lon}{The longitude of the point of interest to pull data for.}

\item{quiet}{logical to print updates to screen. Default is TRUE}
}
\value{
Data frame with time series of SWE and snow depth data.
}
\description{
\code{GetSnodasPointTs} Pull SWE and snow depth values for a specific
point for a given time period.
}
\examples{
snodasExtracted <- GetSnodasPointTs(as.POSIXct('2014-12-01'),as.POSIXct('2015-04-01'),
                                    snodasDir='/d4/karsten/snodas',lat=39.9,
                                    lon=-105.1)
}
\seealso{
Other SNODAS: \code{\link{CalcSnodasCoords}},
  \code{\link{GetSnodasDepthSweDate}},
  \code{\link{PutSnodasCoordsNcdf}},
  \code{\link{PutSnodasNcdf}},
  \code{\link{ReadSnodasDepthSweDate}}
}
\concept{SNODAS}
\keyword{IO}
