% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snotel_get.R
\name{GetSnotel}
\alias{GetSnotel}
\title{Get SNOTEL data for specified sites.}
\usage{
GetSnotel(siteIDs, report = "STAND", series = "Daily",
  intervaltype = "Historic", current = "DAY", duration = "WY",
  startYr = NULL, endYr = NULL, month = NULL, day = NULL,
  quiet = FALSE)
}
\arguments{
\item{siteIDs}{A single site ID or vector of site IDs to download and process. Site IDs
should match the standardized SNOTEL IDs (integers). See snotelMeta for a list of
site IDs (as of June 2015).}

\item{report}{(DEFAULT="STAND") Type of report to generate. Options (from NRCS)
are: \itemize{ \item "STAND" (standard snotel) \item "SOIL" (soil temp and moisture)
\item "SCAN" (standard scan) \item "ALL" (all) \item "WEATHER" (atmospheric) }}

\item{series}{(DEFAULT="Daily") Time series to generate. Options (from NRCS)
are: \itemize{ \item "Daily" \item "Hourly" \item "Hour:HH" (a single hour specified
by HH) }}

\item{intervaltype}{(DEFAULT="Historic") Options (from NRCS)
are: \itemize{ \item "Historic" \item "Current" }}

\item{current}{(DEFAULT="DAY") Only used if intervaltype="Current". Options (from NRCS)
are: \itemize{ \item "DAY" \item "WEEK" \item "YEAR" \item "WATERYEAR" }}

\item{duration}{(DEFAULT="WY") Only used if intervaltype="Historic". Options
are: \itemize{ \item "CY" (calendar year) \item "WY" (water year) }}

\item{startYr}{(OPTIONAL) The first year to download files for.}

\item{endYr}{(OPTIONAL) The last year to download files for.}

\item{month}{(OPTIONAL) The month to download files for.}

\item{day}{(OPTIONAL) The day of month to download files for.}

\item{quiet}{(OPTIONAL) Whether to suppress output (TRUE=suppress output;
DEFAULT=FALSE).}
}
\value{
dataframe
}
\description{
\code{GetSnotel} downloads SNOTEL data tables and creates a dataframe.
}
\details{
\code{GetSnotel} downloads SNOTEL for
specified sites and outputs a dataframe with consistent date and data columns
for use with other rwrfhydro tools.
}
\examples{
sno.santafe <- GetSnotel(c("921","922"), series="Daily", startYr=2009, endYr=2010)
sno.niwot <- GetSnotel(c("663"), series="Hourly", startYr=2009, month=4)
sno.santafe <- GetSnotel(c("921","922"), intervaltype="Current", current="WEEK")
}
\seealso{
Other SNOTEL: \code{\link{snotelMeta}}
}
\concept{
SNOTEL
}
\keyword{IO}

