% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_observations.R
\name{GetStage4Files}
\alias{GetStage4Files}
\title{Download Stage IV and II data files}
\usage{
GetStage4Files(startDate, endDate, dataTyp, dataDir, keepDownloads = FALSE,
  runParallel = FALSE)
}
\arguments{
\item{startDate}{start date + hour in the form of YYYYMMDDHH}

\item{endDate}{end date + hour in the form of YYYYMMDDHH}

\item{dataTyp}{name of the dataset: stage4 or stage2}

\item{dataDir}{directory to store the data, subdirectories YYYY/MM will be created}

\item{keepDownloads}{logical for whether to keep original download files;
if FALSE, dummy empty files will be used for tracking instead
to conserve space (DEFAULT=FALSE)}

\item{runParallel}{logical for running in parallel mode (must have a parallel
backend installed and registered (e.g., doMC or doParallel) (DEFAULT=FALSE)}
}
\description{
\code{GetStage4Files} downloads Stage IV and Stage II data files
}
\details{
\code{GetStage4Files} downloads Stage IV and Stage II data files from NCEP 
ftp site based on user configurations. The function downloads Stage IV or 
Stage II data from NCEP ftp site, unpacks them, and removes uncessary 
files (e.g., *.gif). Note for Stage II, only the multi-sensor ("ml") data 
files are kept while others are removed. Both Stage IV & Stage II datasets 
contain hourly, 6-hourly, and daily precipitation data.
}
\examples{
## Pull data for May 1-5, 2015.
\dontrun{
datadir <- "~/wrfHydroTestCases/OBS/Stage_II/"
GetStage4Files(startDate="2015050100", endDate="2015050500", 
                 dataTyp="stage2", dataDir=datadir, 
                 runParallel=FALSE)
}
}
\seealso{
Other obsDataReads: \code{\link{ReadAmerifluxCSV}},
  \code{\link{ReadAmerifluxNC}},
  \code{\link{ReadCoDwrGage}}, \code{\link{ReadUsgsGage}}
}
\author{
Yuqiong.Liu@noaa.gov (updated by aubreyd)
}
\concept{
dataGet
}
\keyword{IO}
