% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_time_zone.R
\name{GetTimeZone}
\alias{GetTimeZone}
\title{Return the time zone for gauge locations.}
\usage{
GetTimeZone(points, address = NULL, polygonShapeFile = NULL)
}
\arguments{
\item{points}{A dataframe of the gauges, there should be at least two
fields of latitude ad longitude in the dataframe.}

\item{address}{Address to where polygonShapeFile is located.}

\item{polygonShapeFile}{Polygon shapefile containing the boundaries of
timezones (can be downloaded from \url{http://efele.net/maps/tz/world/}.}
}
\value{
Point datafram with an extra column having the time zone for each gauge.
}
\description{
\code{GetTimeZone} gets a dataframe of the gauge locations having at least
two fields of latitude and longitude. it overlays the points with timezone
shapefile (can be downloded from \url{http://efele.net/maps/tz/world/})
and add the timezone as a field to the input dataframe.
Original R-script is from
\url{https://www.nceas.ucsb.edu/scicomp/usecases/point-in-polygon}.
}

