% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{GetUsgsHucData}
\alias{GetUsgsHucData}
\title{Get all the USGS streamgage information within a HUC8.}
\usage{
GetUsgsHucData(huc8, outPath = NULL,
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{huc8}{Character The eight-digit HUC code.}

\item{outPath}{Character The path to the database directory}

\item{metaDBFileName}{Character The name of the database metadata file.}
}
\description{
\code{GetUsgsHucData} gets all the USGS streamgage information within a HUC8. If an
output path is supplied, an existing database is examined for existing records and
the data retrieved from the USGS only extends the available and is saved in the data base.
Could eventually use startDate and endDate to tweak this for calls which dont care about archival.
}
\examples{
bldrHucData    <- GetUsgsHucData(huc='10190005')
satilpaHucData <- GetUsgsHucData(huc='03160203')
vansHucData    <- GetUsgsHucData(huc='03020104')
}

