% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{GetUsgsIvProduct}
\alias{GetUsgsIvProduct}
\title{For a single product, get instantaneous USGS data and separate data and metadata/attributes into a list.}
\usage{
GetUsgsIvProduct(prodDf)
}
\arguments{
\item{prodDf}{is a dataframe returned by \code{dataRetrieval::whatNWISdata(stns$site_no, service = "uv")}
subet to an individual product code.}
}
\value{
list(data=,meta=)
}
\description{
\code{GetUsgsIvProduct} gets instantaneous USGS data for a single product code
and separates the returned dataframe with attributes from \code{dataRetrieval::readNWISuv} into a list with
separate data and metadata/attributes. The meta data are updated with startTime and endTime information.
}

