% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgeospatial}
\alias{GRIBgeospatial}
\title{Use projection information to extract geospatial information from a GRIB file.}
\usage{
GRIBgeospatial(proj, fileIn)
}
\arguments{
\item{proj}{character string specifying projection. Can be obtained from
GRIBgridType.}

\item{fileIn}{character string specifying GRIB file to pull geospatial info
from.}
}
\value{
GRIBgeo data frame with relevant geospatial information
}
\description{
\code{GRIBgeospatial} Extract relevant geospatial information associated with a
GRIB file based on it's projection
}
\examples{
GRIBgeo <- GRIBgeospatial('lambert','rap.t01z.awp130bgrbf16.grib2')
}
\concept{
GRIB geospatial
}
\keyword{geospatial}
\keyword{io}

