% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgeospatialGauss}
\alias{GRIBgeospatialGauss}
\title{Extract regular Gaussian geospatial parameters from GRIB file.}
\usage{
GRIBgeospatialGauss(fileIn)
}
\arguments{
\item{fileIn}{Character string specifying GRIB file to extract
information from.}
}
\value{
GRIBgeo data frame containing relevant geospatial info.
}
\description{
\code{GRIBgeospatialGaus} Pull regular gaussian geospatial parameters
from GRIB file.
}
\examples{
GRIBgeo <- GRIBgeospatialGaus('gfs.t00z.f01.grib2')
}
\seealso{
Other geospatial GRIB: \code{\link{GRIBgeospatialLL}};
  \code{\link{GRIBgeospatialLambert}};
  \code{\link{GRIBgeospatialPolar}}
}
\concept{
geospatial GRIB
}
\keyword{metadata}

