% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgeospatialLambert}
\alias{GRIBgeospatialLambert}
\title{Extract Lambert Conformal geospatial parameters from GRIB file.}
\usage{
GRIBgeospatialLambert(fileIn)
}
\arguments{
\item{fileIn}{Character string specifying GRIB file to extract
information from.}
}
\value{
GRIBgeo data frame containing relevant geospatial info.
}
\description{
\code{GRIBgeospatialLambert} Pull Lambert Conformal geospatial
parameters from GRIB file.
}
\examples{
GRIBgeo <- GRIBgeospatialLambert('rap.t01z.awp130bgrbf16.grib2')
}
\seealso{
Other geospatial GRIB: \code{\link{GRIBgeospatialGauss}};
  \code{\link{GRIBgeospatialLL}};
  \code{\link{GRIBgeospatialPolar}}
}
\concept{
geospatial GRIB
}
\keyword{metadata}

