% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgeospatialLL}
\alias{GRIBgeospatialLL}
\title{Extract lat/lon projection parameters from GRIB file.}
\usage{
GRIBgeospatialLL(fileIn)
}
\arguments{
\item{fileIn}{Character string specifying GRIB file to extract
information from.}
}
\value{
GRIBgeo data frame containing relevant geospatial info.
}
\description{
\code{GRIBgeospatialLL} Pull lat/lon projection geospatial
parameters from GRIB file.
}
\examples{
GRIBgeo <- GRIBgeospatialLL('nldas.force0125A.t00z.f01.grib')
}
\seealso{
Other geospatial GRIB: \code{\link{GRIBgeospatialGauss}};
  \code{\link{GRIBgeospatialLambert}};
  \code{\link{GRIBgeospatialPolar}}
}
\concept{
geospatial GRIB
}
\keyword{metadata}

