% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgeospatialPolar}
\alias{GRIBgeospatialPolar}
\title{Extract Polar Stereographic projection parameters from GRIB file.}
\usage{
GRIBgeospatialPolar(fileIn)
}
\arguments{
\item{fileIn}{Character string specifying GRIB file to extract 
information from.}
}
\value{
GRIBgeo data frame containing relevant geospatial info.
}
\description{
\code{GRIBgeospatialPolar} Pull Polar Stereographic geospatial
parameters from GRIB file.
}
\examples{
GRIBgeo <- GRIBgeospatialPolar('stage4.t00z.f01.grib')

}
\seealso{
Other geospatial GRIB: \code{\link{GRIBgeospatialGauss}},
  \code{\link{GRIBgeospatialLL}},
  \code{\link{GRIBgeospatialLambert}}
}
\concept{
geospatial GRIB
}
\keyword{metadata}

