% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBgridType}
\alias{GRIBgridType}
\title{Extract GRIB grid type using ECMWF GRIB_API.}
\usage{
GRIBgridType(fileIn)
}
\arguments{
\item{fileIn}{The GRIB file of interest.}
}
\value{
gridType The GRIB projection
}
\description{
\code{GRIBgridType} Obtain grid projection from GRIB file for a given file.
ECMWF GRIB_API INSTALLATION ON YOUR SYSTEM IS REQUIRED.
}
\examples{
gridType <- GRIBgridType('rap.t01z.awp130bgrbf16.grib2')
}
\concept{
geospatial GRIB
}
\keyword{GRIB}
\keyword{geospatial}
\keyword{io}

