% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBLatLon}
\alias{GRIBLatLon}
\title{Extract lat/lon values as a grid from a GRIB file.}
\usage{
GRIBLatLon(fileIn, nx, ny)
}
\arguments{
\item{fileIn}{GRIB file to extract values from.}

\item{nx}{number of columns of GRIB data.}

\item{ny}{number of rows of GRIB data.}
}
\value{
dataOut 3D grid (nx,ny,2) of lat/lon values.
 (,,1) is latitude, (,,2) is longitude.
}
\description{
\code{GRIBLatLon} Extract lat/lon grids from GRIB file.
}
\examples{
latLon <- GRIBLatLon('hrrr.grib2',1200,800)
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}}

Other metadata: \code{\link{GRIBNumForecastTimes}}
}
\keyword{GRIB}
\keyword{geospatial}

