% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{GRIBNumForecastTimes}
\alias{GRIBNumForecastTimes}
\title{Extract number of forecast times for a particular variable from
a GRIB file.}
\usage{
GRIBNumForecastTimes(fileIn, var, levType, lev)
}
\arguments{
\item{fileIn}{GRIB file to extract times from.}

\item{var}{Variable of interest.}

\item{levType}{Level type variable resides on.}

\item{lev}{Level variable resides on.}
}
\value{
numTimes Number of variable steps in file.
}
\description{
\code{GRIBNumForecastTimes} Extract number of forecast times
from a GRIB file.
}
\seealso{
Other metadata: \code{\link{GRIBLatLon}}
}
\keyword{GRIB}
\keyword{IO}

