% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_mod.R
\name{HRAPLatLon}
\alias{HRAPLatLon}
\title{Calculate a grid of center-stagger latitude/longitude values for the HRAP grid.}
\usage{
HRAPLatLon(nx, ny, dx, lat1, lon1, lonV)
}
\arguments{
\item{nx}{Integer number of HRAP columns.}

\item{ny}{Integer number of HRAP rows.}

\item{dx}{Float of HRAP grid cell spacing (in meters).}

\item{lat1}{Float of lower left latitude (degrees).}

\item{lon1}{Float of lower left longitude (degrees).}

\item{lonV}{Float of standard parallel HRAP longitude (degrees).}
}
\value{
A grid of HRAP lat/lon center-stagger latitude/longitude values.
}
\description{
\code{HRAPLatLon} Calculate HRAP lat/lon center pixel values.
}
\examples{
\dontrun{
stageIVLatLon <- HRAPLatLon(1121,881,4762.5,23.117,-119.023,-105.0)
}
}
\concept{
regrid, HRAP
}
\keyword{HRAP}
\keyword{geospatial,}
\keyword{regrid,}

