% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{ImproveHucMeta}
\alias{ImproveHucMeta}
\title{For an indivudal product, Improve site metadata to be stored in the database and file.}
\usage{
ImproveHucMeta(hucProdDf)
}
\arguments{
\item{hucProdDf}{dataframe from dataRetrieval::readNWISuv}
}
\value{
list of metadata for a
}
\description{
\code{ImproveHucMeta} extracts the metadata to stash in a usgsDataRetrieval database for an individual product.
It gathers the attributes returned by \code{dataRetrieval::readNWISuv} , supplements siteInfo with
startTime and endTime for each station in UTC to assist appending the records.
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataGet usgsStreamObs
}
\keyword{IO}
\keyword{internal}

