% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RS_get.R
\name{InsertRS}
\alias{InsertRS}
\title{Inserts pre-processed images into appropriate forcing NetCDF files by date.}
\usage{
InsertRS(inFile, forcPath, forcName = "LDASIN_DOMAIN1", varName = NULL,
  varUnit = NULL, varLong = NULL, varNA = -1e+36, overwrite = TRUE)
}
\arguments{
\item{inFile}{The name of the raster stack/brick or NetCDF file (full pathname) to export.}

\item{forcPath}{Path to the forcing data you want to modify. Forcing data files MUST
match the size/resolution of the images in the inFile.}

\item{forcName}{The suffix for the forcing data files to modify (DEFAULT="LDASIN_DOMAIN1")}

\item{varName}{Name for the NetCDF variable to export. The varibale will be copied as-is,
so make sure it matches the variable name needed in the forcing data.}

\item{varUnit}{Units for the NetCDF export variable. Only required if the inFile
is a raster stack/brick. If the inFile is a NetCDF file, the units will carry over.}

\item{varLong}{Long name for the NetCDF export variable. Only required if the inFile
is a raster stack/brick. If the inFile is a NetCDF file, the longname will carry over.}

\item{varNA}{Value to set for "NA" or "no data". Default is -1.e+36.}

\item{overwrite}{Boolean to allow the tool to overwrite existing variables if found in
the forcing data. (DEFAULT=TRUE)}
}
\description{
\code{InsertRS} takes a raster stack or brick of RS images and exports individual
images to matching (by date) forcing NetCDF files.
}
\details{
\code{InsertRS} takes a raster stack or brick  (as created by \code{\link{ConvertRS2Stack}} or \code{\link{SmoothStack}})
or a NetCDF file (as created by \code{\link{ConvertStack2NC}}) of RS images and
exports each layer (time step) to the appropriate time step forcing file.
Only looks for the date (not time) and inserts at the 00:00 hour on that date.
The input stack, brick, or file should be already processed through the \code{\link{ConvertRS2Stack}}, \code{\link{SmoothStack}},
or \code{\link{ConvertStack2NC}} tools or follow the same layer (date) naming convention.
}
\examples{
## Export the raster stack of LAI images created through ConvertRS2Stack to the forcing data.

InsertRS(lai.b, forcPath="FORCING", forcName="LDASIN_DOMAIN3", varName="LAI", varUnit="(m^2)/(m^2)", varLong="Leaf area index")

## Export the NetCDF of LAI images created through ConvertStack2NC to the forcing data.

InsertRS("BCNED_LAI.nc", forcPath="FORCING", forcName="LDASIN_DOMAIN3", varName="LAI")
}
\concept{
aconcept
}
\keyword{IO}

