% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gage_gage_dist_gridded.R
\name{IterateFunction}
\alias{IterateFunction}
\title{Iterate a function}
\usage{
IterateFunction(f, args, resultArg, iterations = 1)
}
\description{
Iterate a function
}
\examples{
\dontrun{
ii <- GoToDownstreamJunct(downstream, upstream, 9296)
ii <- GoToDownstreamJunct(downstream, upstream, ii, skipFirst=TRUE)
IterateFunction(GoToDownstreamJunct,
               list(downstream=downstream, upstream=upstream, ind=9296, skip=TRUE),
               'ind', 3)
}
}

