% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Kge}
\alias{Kge}
\title{Calculate Kling-Gupta Efficiency.}
\usage{
Kge(m, o, na.rm = TRUE, s.r = 1, s.alpha = 1, s.beta = 1)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The Kling-Gupta Efficiency.
}
\description{
\code{Kge} calculates the Kling-Gupta Efficiency.
Calculate the Kling-Gupta Efficiency for vectors
of modelled and observed values: http://www.sciencedirect.com/science/article/pii/S0022169409004843
}
\keyword{internal}

