% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{LoadMetaDB}
\alias{LoadMetaDB}
\title{Load the metadata for the USGS streamflow database.
\code{LoadMetaDB} Load the metadata for the USGS streamflow database.}
\usage{
LoadMetaDB(path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData",
  envir = globalenv())
}
\arguments{
\item{path}{Character path to the meta DB.}

\item{metaDBFileName}{Character name of the meta DB.}

\item{envir}{Envrionment where it is to be loaded.}
}
\value{
Character the name of the variables loaded with the file (invisible).
}
\description{
Load the metadata for the USGS streamflow database.
\code{LoadMetaDB} Load the metadata for the USGS streamflow database.
}

