% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LsObjects}
\alias{LsObjects}
\title{List objects with more detailed metadata}
\usage{
LsObjects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
<<<<<<< HEAD
\item{pos}{pos}

\item{pattern}{pattern}

\item{order.by}{order.by}

\item{decreasing}{decreasing}
=======
\item{pos}{position}

\item{pattern}{pattern}

\item{order.by}{ordering variable}

\item{decreasing}{decreasing order}
>>>>>>> 5fd43922988a528979b189a4ea65340872c07f72

\item{head}{head}

\item{n}{n}
}
\value{
dataframe
}
\description{
\code{LsObjects} lists objects with more detailed info on type, size,
etc. Blatantly plagiarized from the following sources:
Petr Pikal, David Hinds, Dirk Eddelbuettel, Tony Breyal.
}
\keyword{internal}
\keyword{utilities}

