% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/streamflow_to_dart.R
\name{MkDischargeObsSeq}
\alias{MkDischargeObsSeq}
\title{Error (99.5% of all errors) background is always, fith quantile
flow dependent error saturates at 10% of the observed flow as one moves away from
median flow (assumed to be be flows at which rating curve is most accurate).}
\usage{
MkDischargeObsSeq(data, outPath, stationName)
}
\description{
Error (99.5% of all errors) background is always, fith quantile
flow dependent error saturates at 10% of the observed flow as one moves away from
median flow (assumed to be be flows at which rating curve is most accurate).
}

