% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dart_prep_streamflow.R
\name{MkDischargeVariance}
\alias{MkDischargeVariance}
\title{Make variances for prettyUsgs discharge observations.}
\usage{
MkDischargeVariance(prettyUsgs, error3SdFunc, retVariance = TRUE)
}
\arguments{
\item{prettyUsgs}{The prettyUsgs discharge observations to which variances
are to be added.}

\item{error3SdFunc}{Function which accepts the data and returns 3-sigma error
estimates.}

\item{retVariance}{Logical Returns variance if TRUE, else returns 1-sigma
error.}
}
\description{
\code{MkDischargeVariance} makes variances fr prettyUsgs discharge
observations. The formulation of the variances is subjective. Assuming
zero-mean Gaussian observation errors, the approach here is to supply a
function which estimates the 3-sigma (inner 99.5% error quantiles) around the
observations. This amount seems somewhat easier to conceptualize than
1-sigma, hence here we are. This function divides the error amounts by 3 and
either returns the standard deviation or squares the result to return the
variance (default).
}
\examples{
dbPath <- '~/usgsDb/'
prettyOrodell <-
 PrettySiteData(QuerySiteData(QuerySiteName("FOURMILE CREEK AT ORODELL, CO", path=dbPath),
                              product='00060', path=dbPath),metricOnly=FALSE, metric=TRUE)
prettyOro <- MkDischargeVariance(prettyOrodell, Model3SdErrClimTaper)
prettyO <- subset(prettyOro, dateTime < as.POSIXct('2012-01-01'))
oroPlot <- PlotPrettyData(prettyO)
}
\concept{
DART dataMgmt
}
\keyword{manip}

