% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_ncdf.R
\name{MkNcdf}
\alias{MkNcdf}
\title{List based creation of netcdf files.}
\usage{
MkNcdf(varList, globalAttList, filename)
}
\arguments{
\item{varList}{The variable list. See structure in examples.}

\item{globalAttList}{The global attribute list. See structure in examples.}

\item{filename}{The name of the netcdf file to be created.}
}
\value{
Returns the filename if successful.
}
\description{
\code{MkNcdf} List based creation of netCDF files. Works for most cases.
}
\examples{
# Example 1
varList = list()
varList[[1]] <- list( name='precipMult',
                      longname='Precipitation Multiplier',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList = list(scalar=list(name='scalar',values=1,
                                             units='-', unlimited=FALSE,
                                             create_dimvar=FALSE)),
                      data = 1:1 )
varList[[2]] <- list( name='precipMult22',
                      longname='Precipitation Multiplier22',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList =
                          list(  # n.b. the dimension order: z,y,x,t
                                y=list(name='y',values=2.5+((0:6)*5),
                                       units='lat', unlimited=FALSE,
                                       create_dimvar=TRUE),
                                x=list(name='x',values=c(10,20),
                                       units='lon', unlimited=FALSE,
                                       create_dimvar=TRUE)
                               ),
                     data = matrix( 1:7, nrow=7, ncol=2 ) )
globalAttList <- list()
globalAttList[[1]] <- list(name='Restart_Time',value='2012-07-05_00:00:00', precision="text")
globalAttList[[2]] <- list(name='Some reall atts',value='#$\%^!!', precision="text" )
dum <- MkNcdf( varList, globalAttList, 'test3.nc' )
}

