% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usgs_time_slice.R
\name{MkUsgsTimeSlice}
\alias{MkUsgsTimeSlice}
\title{Make timeslices from active USGS discharge data files.}
\usage{
MkUsgsTimeSlice(realTimeFiles, outPath, nearestMin = 5, oldestTime,
  qcFunction, varianceFunction, processedInputPath)
}
\arguments{
\item{realTimeFiles}{Character vector of active RData format files to be processed.}

\item{outPath}{Character, the directory path where ncdf files are to be written.}

\item{nearestMin}{Numeric, the time resolution to which the observation times are rounded
and the netcdf timeslice files are to be written. Must evenly divide 60.}

\item{oldestTime}{POSIXct, the date BEFORE which data will be ignored.}

\item{qcFunction}{Function, used to apply quality control procedures to the timeslice in
metric units. Note that this QC can only use information from the current time. QC
procedures involving the temporal domain will be applied elsewhere.}

\item{varianceFunction}{Function, used to derive the observation variance from
a dataframe with the following columns: \code{site_no}, \code{dateTime}, \code{code},}

\item{processedInputPath}{Character, a directory path to which proccessed input files are
moved upon successful processing to ncdf timeslice files.
\code{queryTime}, and \code{discharge.cms}. The function accepts the dataframe and
returns the data frame with the new \code{variance} column.}
}
\value{
A dataframe
}
\description{
Make timeslices from active USGS discharge data files.
}
\examples{
realTimeFiles <- list.files(pattern='huc.*.RData',
                            path='~/usgsStreamData/realTimeData',
                            full.names=TRUE)
outPath = '~/usgsStreamData/timeSliceData/'
library(doMC)
registerDoMC(4)
ret1 <- MkUsgsTimeSlice( realTimeFiles[1:21], outPath=outPath )
ncdump(ret1$V1[230])
ret2 <- MkUsgsTimeSlice( realTimeFiles[22:42], outPath=outPath )
ncdump(ret1$V1[230])
}

