<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dart_prep_streamflow.R
\name{Model3SdErrPctErrPlusQntlIncpt}
\alias{Model3SdErrPctErrPlusQntlIncpt}
\title{Three-sigma errors specified as percent of observed plus some quantile of
historical flows.}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dart_prep_streamflow.R
\name{Model3SdErrPctErrPlusQntlIncpt}
\alias{Model3SdErrPctErrPlusQntlIncpt}
\title{Three-sigma errors specified as percent of observed plus some quantile of historical flows.}
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs
\usage{
Model3SdErrPctErrPlusQntlIncpt(data, qntlIncpt = 0.005, pctErr = 0.1)
}
\arguments{
\item{data,}{Numeric the values for which errors are to be modeled.}

<<<<<<< HEAD
\item{qntlIncpt}{Numeric the quantile of historical observations to be used
as minimum error.}
=======
\item{qntlIncpt}{Numeric the quantile of historical observations to be used as minimum error.}
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs

\item{pctErr}{Numeric the percent error associated with the observations.}
}
\description{
<<<<<<< HEAD
\code{Model3SdErrPctErrPlusQntlIncpt} models three sigma (standard deviation)
errors (same units as input) as percent of observed plus some quantile of
historical record.
=======
\code{Model3SdErrPctErrPlusQntlIncpt} models three sigma (standard deviation) errors (same
units as input) as percent of observed plus some quantile of historical record.
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs
}

