% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{NamedList}
\alias{NamedList}
\title{Create and or name a list with its entries.}
\usage{
NamedList(theNames)
}
\arguments{
\item{theNames}{Vector to be coerced to character.}
}
\value{
List with names equal to entries.
}
\description{
\code{NamedList} creates a list with names equal to its entries.
}
\examples{
NamedList(1:5)
}
\keyword{internal}

