% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_get.R
\name{ncdump}
\alias{ncdump}
\title{Emulate ncdump -h and -v.}
\usage{
ncdump(file, variable, quiet = FALSE, collapse_degen = TRUE)
}
\arguments{
\item{file}{Character, the file to inspect.}

\item{variable}{Character, a variable to return.}

\item{quiet}{Logical, suppress the 'meta' dump?}

\item{collapse_degen}{If TRUE (the default), then degenerate (length==1) dimensions in the returned array are removed.}
}
\value{
If variable is not set, the meta object \code{ncdf4::nc_open(file)} is returned. If \code{variable}
is set, its values are returned.
}
\description{
I just hacked print.ncdf4 just to make it look more like unix output.
}
\seealso{
Other ncdf: \code{\link{GetNcdfFile}},
  \code{\link{MkFrcAdjFromLdasin}}, \code{\link{MkNcdf}}
}
\concept{
ncdf
}
