% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Nse}
\alias{Nse}
\title{Calculate Nash-Sutcliffe Efficiency.}
\usage{
Nse(m, o, nullModel = mean(o, na.rm = na.rm), na.rm = TRUE)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The Nash-Sutcliffe Efficiency.
}
\description{
\code{Nse} calculates the Nash-Sutcliffe Efficiency.
Calculate the Nash-Sutcliffe Efficiency for vectors
of modelled and observed values.
}
\keyword{internal}
