% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{PadRange}
\alias{PadRange}
\title{Expand limits by some amount or proportionally to their difference.}
\usage{
PadRange(limits, delta = diffMult * diff(limits), diffMult = 0.05)
}
\arguments{
\item{limits}{A vector of length 2, an initial range, to be expanded.}

\item{delta}{An amount to add(subtract) from the upper(lower) limit.}

\item{diffMult}{A fraction of the passed range (\code{limits}) to use as \code{delta}.}
}
\description{
\code{PadRange} Takes limits and expands them by some amount or proportionally to their difference.
}
\examples{
PadRange(c(0,100))
PadRange(c(0,100), delta=.1)
PadRange(c(0,100), diffMult=.1)
}
\keyword{internal}

