% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_fdc.R
\name{PlotFdc}
\alias{PlotFdc}
\title{Plot a flow duration curve for a single streamflow time series}
\usage{
PlotFdc(strDf, strCol = "q_cms", spline = TRUE, fdcProb = NULL)
}
\arguments{
\item{strDf}{The streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}} or \code{\link{ReadUsgsGage}})
already processed through \code{\link{CalcFdc}}. The data frame must contain a column of streamflow values and a
corresponding column (named "<strCol>.fdc") of flow exceedance probabilities.}

\item{strCol}{The name of the column containing the streamflow values (DEFAULT="q_cms").}

\item{spline}{(TRUE or FALSE) Option to add spline-fit curves to the plot (DEFAULT=TRUE).}

\item{fdcProb}{Optional exceedance threshold to calculate/plot (e.g., 0.2 for 20\% exceedance).}
}
\value{
A plot of flow (y) vs. flow exceedance probabilities (x).
}
\description{
\code{PlotFdc} plots a flow duration curve from a dataframe processed through \code{\link{CalcFdc}}
and optional spline function processed through \code{\link{CalcFdcSpline}}.
}
\details{
\code{PlotFdc} reads the dataframe generated by \code{\link{CalcFdc}} and outputs a plot of flow vs.
percent exceedances and, optionally the fitted spline curve.
}
\examples{
## Take a time series of observed 5-minute streamflow values for Fourmile Creek (stored in the dataframe
## obsStr5min.fc in the column "q_cms") that has already been processed through "CalcFdc" (so also contains
## a column named "q_cms.fdc") and plot with the 20\% exceedance threshold called out.

PlotFdc(obsStr5min.fc, fdcProb=0.2)
}
\seealso{
Other flowDurationCurves: \code{\link{CalcFdcSpline}};
  \code{\link{CalcFdc}}; \code{\link{PlotFdcCompare}}
}
\concept{
modelEval plot
}
\keyword{hplot}

