% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_fdc.R
\name{PlotFdcCompare}
\alias{PlotFdcCompare}
\title{Plots a flow duration curve for up to three streamflow time series (e.g., observed & two model outputs)}
\usage{
PlotFdcCompare(strflowDf.obs, strflowCol.obs = "q_cms", strflowDf.mod1,
  strflowCol.mod1 = "q_cms", strflowDf.mod2 = NULL,
  strflowCol.mod2 = "q_cms", spline = TRUE, logy = TRUE,
  labelObs = "Observed", labelMod1 = "Model 1", labelMod2 = "Model 2")
}
\arguments{
\item{strflowDf.obs}{The OBSERVED streamflow time series dataframe (e.g., output from \code{\link{ReadUsgsGage}})
already processed through \code{\link{CalcFdc}}. The dataframe must contain a column of streamflow values and a
corresponding column (named "<strflowCol>.fdc") of flow exceedance probabilities.}

\item{strflowCol.obs}{The name of the column containing the streamflow values for the OBSERVED
dataframe (DEFAULT="q_cms").}

\item{strflowDf.mod1}{The FIRST MODEL streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}})
already processed through \code{\link{CalcFdc}}. The dataframe must contain a column of streamflow values and a
corresponding column (named "<strflowCol>.fdc") of flow exceedance probabilities.}

\item{strflowCol.mod1}{The name of the column containing the FIRST MODEL streamflow values
(DEFAULT="q_cms").}

\item{strflowDf.mod2}{The SECOND MODEL streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}})
already processed through \code{\link{CalcFdc}}. The dataframe must contain a column of streamflow values and a
corresponding column (named "<strflowCol>.fdc") of flow exceedance probabilities.}

\item{strflowCol.mod2}{The name of the column containing the SECOND MODEL streamflow values
(DEFAULT="q_cms").}

\item{spline}{(TRUE or FALSE) Option to add spline-fit curves to the plot (DEFAULT=TRUE).}

\item{logy}{(TRUE or FALSE) Optional flag to set the y-axis to log-scale (DEFAULT=TRUE).}

\item{labelObs}{Optional label for the observed streamflow (DEFAULT="Observed")}

\item{labelMod1}{Optional label for the FIRST MODEL (DEFAULT="Model 1")}

\item{labelMod2}{Optional label for the SECOND MODEL (DEFAULT="Model 2")}
}
\value{
A plot of flow (y) vs. flow exceedance probabilities (x).
}
\description{
\code{PlotFdcCompare} plots up to three flow duration curves from dataframes processed through \code{\link{CalcFdc}}.
}
\details{
\code{PlotFdcCompare} reads up to three dataframes generated by \code{\link{CalcFdc}} and outputs a plot of
flow vs. percent exceedances and, optionally, the fitted spline curves. Intended to plot one observed flow duration
curve and one or two modelled flow duration curves for comparison.
}
\examples{
## Take a time series of observed 5-minute streamflow values for Fourmile Creek (obsStr5min.fc) and two model
## runs (mod1Str5min.fc, mod2Str5min.fc), all with streamflow columns named "q_cms" and already processed
## through "CalcFdc", and plot the flow duration curves for all three.

PlotFdcCompare(obsStr5min.fc, "q_cms", mod1Str5min.fc, "q_cms", strflowDf.mod2=mod2Str5min.fc, strflowCol.mod2="q_cms", labelObs="Observed Fourmile Creek (5-min)", labelMod1="1-D Flow Model", labelMod2="2-D Flow Model")
}

