% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_modeleval.R
\name{PlotFluxCompare}
\alias{PlotFluxCompare}
\title{Plot time series comparing modeled and observed fluxes}
\usage{
PlotFluxCompare(strDf.obs, strCol.obs = "q_cms", strDf.mod1,
  strCol.mod1 = "q_cms", strDf.mod2 = NULL, strCol.mod2 = "q_cms",
  stdate = NULL, enddate = NULL, logy = FALSE, labelObs = "Observed",
  labelMod1 = "Model 1", labelMod2 = "Model 2",
  title = "Observed and Modelled Fluxes", colorObs = "black",
  colorMod1 = "green2", colorMod2 = "blue")
}
\arguments{
\item{strDf.obs}{The OBSERVED flux time series dataframe (e.g., output from
\code{\link{ReadUsgsGage}}). The dataframe must contain a column of flux
values and a POSIXct or Date column.}

\item{strCol.obs}{The name of the column containing the flux values for the
OBSERVED dataframe (DEFAULT="q_cms").}

\item{strDf.mod1}{The FIRST MODEL flux time series dataframe (e.g., output
from \code{\link{ReadFrxstPts}}). The dataframe must contain a column of
flux values and a POSIXct or Date column.}

\item{strCol.mod1}{The name of the column containing the FIRST MODEL flux
values (DEFAULT="q_cms").}

\item{strDf.mod2}{The SECOND MODEL flux time series dataframe (e.g., output
from \code{\link{ReadFrxstPts}}). The dataframe must contain a column of
flux values and a POSIXct or Date column.}

\item{strCol.mod2}{The name of the column containing the SECOND MODEL flux
values (DEFAULT="q_cms").}

\item{stdate}{Start date for plot/statistics (DEFAULT=NULL, all records will
be used). Date MUST be specified in POSIXct format with appropriate
timezone (e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d
\%H:\%M:\%S", tz="UTC")) or Date (e.g., as.Date("2013-05-01", 
format="\%Y-\%m-\%d")) where Date is assumed to match a UTC date.}

\item{enddate}{End date for plot/statistics (DEFAULT=NULL, all records will
be used). Date MUST be specified in POSIXct format with appropriate
timezone (e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d
\%H:\%M:\%S", tz="UTC")) or Date (e.g., as.Date("2013-05-01", 
format="\%Y-\%m-\%d")) where Date is assumed to match a UTC date.}

\item{logy}{(TRUE or FALSE) Optional flag to set the y-axis to log-scale
(DEFAULT=FALSE).}

\item{labelObs}{Optional label for the observed streamflow
(DEFAULT="Observed")}

\item{labelMod1}{Optional label for the FIRST MODEL (DEFAULT="Model 1")}

\item{labelMod2}{Optional label for the SECOND MODEL (DEFAULT="Model 2")}

\item{title}{Optional for the plot (DEFAULT="Observed and Modelled Fluxes")}

\item{colorObs}{Optional color for the observed line (DEFAULT="black")}

\item{colorMod1}{Optional color for the FIRST MODEL line (DEFAULT="green2")}

\item{colorMod2}{Optional color for the FIRST MODEL line (DEFAULT="blue")}
}
\value{
A plot of the hydrographs.
}
\description{
\code{PlotFluxCompare} plots a time series of an observed flux (e.g.,
streamflow, ET) and up to 2 modelled fluxes.
}
\details{
\code{PlotFluxCompare} reads modelled and observed dataframes (e.g., as
generated from \code{\link{ReadFrxstPts}} and \code{\link{ReadUsgsGage}}) and
plot the time series and summary statistics. The tool will subset data to
matching time periods (e.g., if the observed data is at 5-min increments and
modelled data is at 1-hr increments, the tool will subset the observed data
to select only observations on the matching hour break).
}
\examples{
## Take a time series of observed 5-minute streamflow values for Fourmile
## Creek (obsStr5min.fc) and two model runs (mod1Str1h.fc, mod2Str1h.fc), 
## all with streamflow columns named "q_cms", and plot the the hydrographs 
## for all three over the May-June snowmelt period.

\dontrun{
PlotFluxCompare(obsStr5min.fc, "q_cms", modStrh.chrt.fc, "q_cms", 
                strDf.mod2=modStrh.allrt.fc, strCol.mod2="q_cms",
                labelObs="Observed Fourmile Creek at Orodell",
                labelMod1="Channel Routing Only", labelMod2="All Routing",
                title="Streamflow: Fourmile Creek",
                stdate=as.POSIXct("2013-05-01 00:00:00", 
                                  format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC"),
                enddate=as.POSIXct("2013-06-30 00:00:00",
                                   format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC"))
}
}
