% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{PlotPrettyData}
\alias{PlotPrettyData}
\title{Plot USGS site data which has been prettied with PrettySiteData.
\code{PlotPrettyData} plots USGS site data which has been prettied with PrettySiteData.}
\usage{
PlotPrettyData(prettyUsgs, plot = TRUE)
}
\arguments{
\item{prettyUsgs}{dataframe returned from PrettySiteData}

\item{plot}{Logical to plot before returning or not.}
}
\value{
A function(closure) with arguments controlling the look of its graphical output. It's actual
        return value is a list of ggplot2 object which can be custom manipulated.
}
\description{
Plot USGS site data which has been prettied with PrettySiteData.
\code{PlotPrettyData} plots USGS site data which has been prettied with PrettySiteData.
}
\examples{
# See vignette "Collect USGS stream observations and build a local database" for examples.
}

