% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{PlotPrettyUsgs}
\alias{PlotPrettyUsgs}
\title{Plot USGS site data which has been prettied with PrettyUsgs.}
\usage{
PlotPrettyUsgs(prettyUsgs, plot = TRUE, errInnerQntl = 0.995)
}
\arguments{
\item{prettyUsgs}{dataframe returned from PrettyUsgs}

\item{plot}{Logical to plot before returning or not.}

\item{errInnerQntl}{The inner quantile of the error estimate to be plotted
with error bars. For example, the "68-95-99.7 rule" where specifying these
as the errInnerQntl would display 1, 2, and 3 standard deviations,
respectively.}
}
\value{
A function(closure) with arguments controlling the look of its
  graphical output. It's actual return value is a list of ggplot2 object
  which can be custom manipulated.
}
\description{
\code{PlotPrettyUsgs} plots USGS site data which has been prettied with
\code{PrettyUsgs}.
}
\examples{
# See vignette "Collect USGS stream observations and build a local database" for examples.
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PrettyUsgs}}; \code{\link{QueryHaveSite}};
  \code{\link{QuerySiteData}}; \code{\link{QuerySiteInfo}};
  \code{\link{QuerySiteName}}; \code{\link{QuerySiteProd}};
  \code{\link{SaveHucData}}; \code{\link{[.prettyUsgs}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
plot usgsStreamObs
}
\keyword{hplot}

