% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize_modeleval.R
\name{PlotStationData}
\alias{PlotStationData}
\title{Plot point NetCDF station time series observations}
\usage{
PlotStationData(stnObs.list, stack = FALSE)
}
\arguments{
\item{stnObs.list}{Full list of data output from ReadNearStationData, including time series data, list of units and filename}

\item{stack}{(default=FALSE) if TRUE, instead of individual plots to click through, the output will be one page of stacked plots}
}
\value{
Plot comparison of model and station data
}
\description{
\code{PlotStationData} plots station observations from list containing data frame of time series
}
\details{
\code{PlotStationData} plots station observation time series for all recorded variables
}
\examples{
## Plot Station Observations at a given station
## (Upper Canejos River, URG1)
PlotStationData(URG1)
}
\seealso{
Other obsDataReads: \code{\link{ReadAmerifluxCSV}};
  \code{\link{ReadAmerifluxNC}};
  \code{\link{ReadCoDwrGage}};
  \code{\link{ReadNearStationData}};
  \code{\link{ReadUsgsGage}}
}
\concept{
dataGet
}
\keyword{IO}

