% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualize_modeleval.R
\name{PlotWatBudg}
\alias{PlotWatBudg}
\title{Plot water balance from WRF-Hydro (w/NoahMP) output}
\usage{
PlotWatBudg(wbDf, plottyp = "pie")
}
\arguments{
\item{wbDf}{The water budget dataframe (required)}

\item{plottyp}{The plot type (pie or bar) (default=pie)}
}
\value{
A plot of the water budget components in mm.
}
\description{
\code{PlotWatBudg} plot water budget components from WRF-Hydro (w/NoahMP) model output.
}
\details{
Read water budget dataframe (as generated from \code{\link{CalcNoahmpWatBudg}}) and plot water budget
components as a piechart or barchart.
NOTE: Currently only works for runs using NoahMP as the LSM.
}
\examples{
## Plot the water budget components from a water budget dataframe generated using
## CalcNoahmpWatBudg. Plot as a piechart.

PlotWatBudg(wb.allrt.fc)

## Plot the same as a barchart.

PlotWatBudg(wb.allrt.fc, "bar")
}

