% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PolyToRaster.R
\name{PolyToRaster}
\alias{PolyToRaster}
\title{Creates raster file from polygon matching the \code{geoFile}}
\usage{
PolyToRaster(geoFile = NULL, useRfc = FALSE, polygon = NULL,
  polygonAddress = NULL, polygonShapeFile = NULL, field = NULL,
  fun = "last", mask = FALSE, maskValue = 1, getCoverMask = FALSE,
  getCover = FALSE, plot.it = TRUE, parallel = FALSE)
}
\arguments{
\item{geoFile}{The geogrid NetCDF file.}

\item{useRfc}{Logical: if \code{TRUE} will use the loaded rfc polygon in rwrfhydro.
Default is \code{FALSE}}

\item{polygon}{SpatialPolygon* This is used if you want to use a
polygon already loaded in memory.}

\item{polygonAddress}{Character: Address to where polygonShapeFile is located.}

\item{polygonShapeFile}{Character: Name of a polygon shapefile.}

\item{field}{Numeric or character: The value(s) to be transferred. This can be
a single number, or a vector of numbers that has the same length as the
number of spatial features (polygons).}

\item{fun}{Function or character: To determine what values to assign to cells
that are covered by multiple spatial features. You can use functions such as
\code{min, max}, or \code{mean}, or one of the following character values:
\code{'first'}, \code{'last'}, \code{'count'}.}

\item{mask}{Logical: If \code{TRUE} the values of the input Raster object are
'masked' by the spatial features of polygon shapefile. That is, cells that spatially
overlap with the spatial features retain their values, the other cells
become \code{NA}. Default is \code{FALSE}.}

\item{maskValue}{Numeric: The value to be used for creating the mask. Default is \code{1}.}

\item{getCoverMask}{Logical: If \code{TRUE}, the fraction of each grid cell that
is covered by the polygons is returned (and the values of \code{field, fun,
mask}. The fraction covered is estimated by
dividing each cell into 100 subcells and determining presence/absence of
the polygon in the center of each subcell}

\item{getCover}{Logical: If \code{TRUE}, the fraction of each grid cell that
is covered by the each \code{field} is returned as a layer in the returned RasterStack.
Name of the RasetrLayers in the RasterStacK is obtained from unique values of \code{field}.
The fraction covered is estimated by
dividing each cell into 100 subcells and determining presence/absence of
the polygon in the center of each subcell}

\item{plot.it}{Logical: If \code{TRUE} will plot both the raster and the polygon on the same plot}

\item{parallel}{Logical: will be used only if \code{GetCover} is \code{TRUE}}
}
\value{
return a RasterLayer or a RasterStack.
}
\description{
\code{PolyToRaster} takes a NetCDF geogrid as well as a polygon
and return the attribute requested from polygon as a raster file.
}
\examples{
\dontrun{

##Example 1:
## To rasterize the rfc SpatialPolygonsDataFrame :

r <- PolyToRaster(geoFile = "/glade/scratch/arezoo/IOC/ESMF/geo_em.d01.nc.conus_1km",
                 useRfc = TRUE,
                 field ="BASIN_ID")

You can get the numbres assign to each BAISN_ID like :


## Example 2:
## To return a mask with default value of 1 inside the polygons and NA outside

r1 <- PolyToRaster(geoFile = "/glade/scratch/arezoo/IOC/ESMF/geo_em.d01.nc.conus_1km",
                  useRfc = TRUE,
                  field ="BASIN_ID",
                  mask = TRUE)

##Example 3:
## To return a mask with value of 5 inside the polygons and NA outside

r2 <- PolyToRaster(geoFile = "/glade/scratch/arezoo/IOC/ESMF/geo_em.d01.nc.conus_1km",
                              useRfc = TRUE,
                              field ="BASIN_ID",
                              mask = TRUE,
                              maskValue = 5)

##Example 4:
## To return a raster with values to be the BASIN_ID (convert character to integer numbers)
## with fraction of each grid cell that is covered by the polygons

r3 <- PolyToRaster(geoFile = "/glade/scratch/arezoo/IOC/ESMF/geo_em.d01.nc.conus_1km",
                              useRfc = TRUE,
                              field ="BASIN_ID",
                              getCover = TRUE,
                              parallel = TRUE))
plot(r3)

## Example 5:
## To read a shapefile (polygon) from disk and
## return a raster with values of field
## with fraction of each grid cell that is covered by the polygons

r4 <- PolyToRaster(geoFile = "/glade/scratch/arezoo/IOC/ESMF/geo_em.d01.nc.conus_1km",
                  polygonAddress= "/glade/scratch/arezoo/QPF_verification_rwrfhydro/gis",
                  polygonShapeFile= "HUC4",
                  field ="HUC4")
plot(r4)
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetProj}};
  \code{\link{GetRfc}}; \code{\link{GetTimeZone}}
}
\concept{
dataMgmt geospatial
}
\keyword{IO}

