% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PosixOrigin}
\alias{PosixOrigin}
\title{The standard posix origin 1970-01-01}
\usage{
PosixOrigin(as.POSIX = FALSE, tz = "UTC")
}
\arguments{
\item{as.POSIX}{Logical for return type.}

\item{tz}{Character string for time zone to use.}
}
\value{
The standard POSIX origin in requested format.
}
\description{
\code{PosixOrigin} as either a character string or a POSIXct.
}
\keyword{internal}
\keyword{utilities}

