% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{PrettySiteData}
\alias{PrettySiteData}
\title{Pretty USGS site data makes nice headers and optionally converts to metric.}
\usage{
PrettySiteData(data, tz = "UTC", metric = metricOnly, metricOnly = TRUE,
  na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe from QuerySiteData}

\item{tz}{Character The timezone for the POSIXct dataTime variable to be returned.}

\item{metric}{Logical}

\item{metricOnly}{Logical}

\item{na.rm}{Logical Remove all missing observations?}
}
\value{
dataframe similar to input with improved names and/or metric variables.
}
\description{
\code{PrettySiteData} beautifies the data frame
}
\examples{
p='~/streamflow/OBS/'
dataOrodell <- PrettySiteData(QuerySiteData(QuerySiteName("FOURMILE CREEK AT ORODELL, CO", p),
                                            '00060', p), metricOnly=TRUE)
}

