% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{PrettyUsgs}
\alias{PrettyUsgs}
\title{PrettyUsgs constructs the S3 class prettyUsgs.}
\usage{
PrettyUsgs(data, tz = "UTC", metric = TRUE, na.rm = TRUE,
  format = "long")
}
\arguments{
\item{data}{Dataframe from QuerySiteData}

\item{tz}{Character The timezone for the POSIXct dataTime variable to be returned.}

\item{metric}{Logical. Units are either metric or not (not both).}

\item{na.rm}{Logical Remove all missing observations?}

\item{format}{character, either 'wide' or 'long'. In wide format there are more columns, one
for each variable named \code{'variable (units)'}. In long format, there are columns \code{value},
\code{variable}, and \code{units}. Switching between long and wide formats can be done with
\code{Reformat()} (though I've yet to program this...).}
}
\value{
dataframe similar to input with improved names and/or metric variables.
}
\description{
\code{PrettyUsgs} beautifies data frames of USGS data. This is the S3 constructor for the
PrettyUsgs object.
}
\examples{
p='~/wrfHydroTestCases/usgsDb/'
dataOrodell <- PrettyUsgs(QuerySiteData(QuerySiteName("FOURMILE CREEK AT ORODELL, CO", p),
                                            '00060', p), metric=TRUE)
## multisite and multiproduct case
dataMultiHuc <- PrettyUsgs(QuerySiteData(c('06730500','02470072'),c('00065','00060'),p))
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{manip}

