% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{PutSnodasCoordsNcdf}
\alias{PutSnodasCoordsNcdf}
\title{PutSnodasCoordsNcdf}
\usage{
PutSnodasCoordsNcdf()
}
\value{
Success if the filename (which is SNODAS_Coordinates.nc), otherwise NULL.
}
\description{
\code{PutSnodasCoordsNcdf} Put the output of CalcSnodasCoords into a netcdf file.
}
\examples{
PutSnodasCoordsNcdf()
}

