% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{PutSnodasNcdf}
\alias{PutSnodasNcdf}
\title{PutSnodasNcdf}
\usage{
PutSnodasNcdf(snodasList)
}
\arguments{
\item{snodasList}{The output of readSnodasDepthSweDate.}
}
\value{
Success if the filename which is (SNODAS_YYYYMMDD.nc), otherwise NULL.
}
\description{
\code{PutSnodasNcdf} Put the output of readSnodasDepthSweDate into a netcdf file.
}
\examples{
snodasGot <- getSnodasDepthSweDate(as.POSIXct('2015-02-28'))
if(snodasGot) snodasList <- readSnodasDepthSweDate(as.POSIXct('2015-02-28'))
PutSnodasNcdf(snodasList)
}

