% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{QueryHaveSite}
\alias{QueryHaveSite}
\title{General purpose query/get for instantaneous USGS streamflow data.}
\usage{
QueryHaveSite(site, path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData", get = FALSE,
  retData = FALSE)
}
\arguments{
\item{site}{Character USGS site number.}

\item{path}{Character The path to the database.}

\item{metaDBFileName}{Character The name of the metadata file.}

\item{get}{Logical Get the data from NWIS if it is not local? Data are saved
to local database.}

\item{retData}{Logical OR Character If true return all products, otherwise
return the specified product. If get from NWIS, all products are retrieved
and saved locally but only specified products are returned.}
}
\value{
See retData argument.
}
\description{
\code{QueryHaveSite}
}
\examples{
\dontrun{
haveOro <- QueryHaveSite('06727500', path='~/wrfHydroTestCases/usgsDb', retData=TRUE)
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyUsgs}}; \code{\link{PrettyUsgs}};
  \code{\link{QuerySiteData}}; \code{\link{QuerySiteInfo}};
  \code{\link{QuerySiteName}}; \code{\link{QuerySiteProd}};
  \code{\link{SaveHucData}}; \code{\link{[.prettyUsgs}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{database}

