% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{QuerySiteData}
\alias{QuerySiteData}
\title{Returns the data for given sites from local database.}
\usage{
QuerySiteData(site, product, path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{site}{Character USGS site number.}

\item{product}{Character USGS product code number.}

\item{path}{Character path to the database.}

\item{metaDBFileName}{Character The name of the metadata file.}
}
\value{
dataframe of data with pertinent attributes.
}
\description{
\code{QuerySiteData} gets the specified data from the local database.
}
\examples{
p='~/streamflow/OBS/'
dataOrodell <- QuerySiteData(QuerySiteName("FOURMILE CREEK AT ORODELL, CO", p), '00060', p)
siteInfo<-QuerySiteInfo(c('station_nm','site_no','stateCd'), path=p)
dataCO <- QuerySiteData(subset(siteInfo, stateCd=='08' & product=='00060')$site_no, '00060', p)
}

