% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{QuerySiteInfo}
\alias{QuerySiteInfo}
\title{Returns the desired information from the database metadata file.}
\usage{
QuerySiteInfo(info = NULL, path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{info}{Character vector, information fields in
\code{HUC$prod$meta$SiteInfo$info}.}

\item{path}{Character The path to the database directory.}

\item{metaDBFileName}{Character The name of the metadata file.}
}
\value{
dataframe of requested info with all available HUC and product codes.
}
\description{
\code{QuerySiteInfo} gets the specified info from the local database.
}
\examples{
\dontrun{
QuerySiteInfo(c('station_nm','site_no'), path='~/streamflow/OBS/')
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteName}}; \code{\link{QuerySiteProd}};
  \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{database}

