% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{QuerySiteProd}
\alias{QuerySiteProd}
\title{Find the products available for a given site in the local database.}
\usage{
QuerySiteProd(site, path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{site}{Character USGS site number.}

\item{path}{Character The path to the database.}

\item{metaDBFileName}{Character The name of the metadata file.}
}
\description{
\code{QuerySiteProd} finds the products available for a given site in the local database.
}
\examples{
\dontrun{
QuerySiteProd('06730500', '~/streamflow/OBS/')
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{database}

