% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RBFlash}
\alias{RBFlash}
\title{Calculate Richards-Baker Flashiness Index.}
\usage{
RBFlash(m, na.rm = TRUE)
}
\arguments{
\item{m}{The vector of values.}
}
\value{
The Richards-Baker Flashiness Index.
}
\description{
\code{RBFlash} calculates the Richards-Baker Flashiness Index.
Calculate the Richards-Baker Flashiness Index for vectors
of modelled or observed values.
}
\keyword{internal}

