% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameriflux_get.R
\name{ReadAmeriflux}
\alias{ReadAmeriflux}
\title{Read Ameriflux data into R dataframe.}
\usage{
ReadAmeriflux(inFile, tz, useEndTime = FALSE)
}
\arguments{
\item{inFile}{Path to the CSV file (e.g., "AMF_US-NR1_BASE_HH_9-1.csv").}

\item{tz}{The time zone for the site (generally Ameriflux data is in local time).}

\item{useEndTime}{By default, a POSIXct column will be defined based on the 
"TIMESTAMP_START" column. If you prefer to use "TIMESTAMP_END", set this to TRUE.}
}
\value{
dataframe
}
\description{
\code{ReadAmeriflux} reads pre-downloaded Ameriflux BASE data table (BASE data 
product CSV file) and creates a dataframe.
}
\details{
\code{ReadAmeriflux} reads Ameriflux BASE data product CSV file and outputs a 
dataframe with consistent date and data columns for use with other rwrfhydro tools.
}
\examples{
\dontrun{
amf.NR <- ReadAmeriflux('~/wrfhydroTestCases/Fourmile_Creek_testcase_v2.0/OBS/AMF_US-NR1_BASE_HH_9-1.csv',
                        tz='America/Denver')
}
}
\seealso{
Other Ameriflux: \code{\link{amfMeta}},
  \code{\link{tzLookup}}
}
\concept{Ameriflux}
\keyword{IO}
