% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_observations.R
\name{ReadAmerifluxNC}
\alias{ReadAmerifluxNC}
\title{Read standard-format NetCDF data downloaded from Ameriflux}
\usage{
ReadAmerifluxNC(pathFluxData, timeZone)
}
\arguments{
\item{pathFluxData}{The full pathname to the flux time series NetCDF file as downloaded from an
Ameriflux data server.}

\item{timeZone}{The time zone for the flux data. Time zone name must be R-friendly for your
current OS. See: \url{http://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html}}
}
\value{
A dataframe containing the Ameriflux data.
}
\description{
\code{ReadAmerifluxNC} reads Ameriflux data table (Level 2 standardized NetCDF file) and creates
a dataframe.
}
\details{
\code{ReadAmerifluxNC} reads an Ameriflux Level 2 standardized NetCDF file and outputs a dataframe
with consistent date and data columns for use with other rwrfhydro tools.
}
\examples{
## Takes a NetCDF file downloaded from the ORNL Amerifux website for US-NC2 (North Carolina Loblolly Pine)
## and returns a dataframe.

obsFlux30min.usnc2 <- ReadAmerifluxNC("../OBS/FLUX/AMF_USNC2_2005_L2_WG_V003.nc", "America/New_York")
}
\seealso{
Other obsDataReads: \code{\link{ReadAmerifluxCSV}};
  \code{\link{ReadCoDwrGage}}; \code{\link{ReadUsgsGage}}
}
\concept{
dataGet
}
\keyword{IO}

