% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_chan_connectivity.R
\name{ReadChanConn}
\alias{ReadChanConn}
\title{Read channel connectivity files now output by WRF Hydro.}
\usage{
ReadChanConn(connFile)
}
\arguments{
\item{connFile,}{character. The path/file of the connectivity file of interest.}
}
\description{
Read channel connectivity files now output by WRF Hydro.
}
\examples{
connFile4Mile <- '~/wrfHydroTestCases/Fourmile_Creek/CHANNEL_CONNECTIVITY.nc'
conn4Mile <- ReadChanConn(connFile4Mile)
}

