% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadChrtgrid}
\alias{ReadChrtgrid}
\title{Read WRF-Hydro CHRTOUT data files for gridded routing.}
\usage{
ReadChrtgrid(pathOutdir = NULL, gaugeFile = NULL, gaugePtlist = NULL,
  pattern = glob2rx("*.CHRTOUT_DOMAIN*"))
}
\arguments{
\item{pathOutdir}{The full pathname to the output directory containing the 
CHRTOUT files.}

\item{gaugeFile}{Optional list of gaugePts or frxstPts to import (must be consistent
with the format of the frxstpts.txt file specified for creating frxstpts on the grid).}

\item{gaugePtlist}{Optional list of gage or frxst locations to import. Must provide the LAT,LON,and SITE_NO 
for each location specified.}

\item{pathOutdir}{The full pathname to the output directory containing the 
CHRTOUT files.}

\item{gaugeFile}{Optional list of gaugePts or frxstPts to import (must be consistent
with the format of the frxstpts.txt file specified for creating frxstpts on the grid).}

\item{gaugePtlist}{Optional list of gage or frxst locations to import. Must provide the LAT,LON,and SITE_NO 
for each location specified.}
}
\value{
A datatable containing a time series of channel fluxes. Note that datatable is REQUIRED.

A datatable containing a time series of channel fluxes. Note that datatable is REQUIRED.
}
\description{
\code{ReadChrtgrid} reads in WRF-Hydro CHRTOUT files for gridded routing and outputs a time series of
channel fluxes at specified lat/long locations.
}
\details{
\code{ReadChrtgrid} reads standard-format WRF-Hydro CHRTOUT NetCDF files with gridded routing and 
outputs a time series of channel fluxes.
}
\examples{
## Take an OUTPUT directory for a model run with gridded routing (channel_routing option =3) 
## and find the streamflow at specific lat/longs (e.g. forecast points and gages) that corresponds to the location on 
## the channel grid; output is a data table of streamflow for all pts provided.

\dontrun{
ReadChrtgrid('~/wrfHydroTestCases/FRN.REACH/OUTPUT', 
     gaugeFile='~/wrfHydroTestCases/FRN.REACH/frxstpts_frntrng.txt')

gaugePtlist=list(BigTOMP=data.frame(lon=-105.5836, lat=40.3534, id='402114105350101'),
                 Fountain_Cr_nr_Col_Spgs=data.frame(lon=-104.8782, lat=38.85422, id='07105500')) 
  ReadChrtgrid(('~/wrfHydroTestCases/FRN.REACH/OUTPUT',gaugePtlist=gaugePtlist) 
}
## Take an OUTPUT directory for a model run with gridded routing (channel_routing option =3) 
## and find the streamflow at specific lat/longs (e.g. forecast points and gages) that corresponds to the location on 
## the channel grid; output is a data table of streamflow for all pts provided.

\dontrun{
ReadChrtgrid('~/wrfHydroTestCases/FRN.REACH/OUTPUT', 
     gaugeFile='~/wrfHydroTestCases/FRN.REACH/frxstpts_frntrng.txt')

gaugePtlist=list(BigTOMP=data.frame(lon=-105.5836, lat=40.3534, id='402114105350101'),
                 Fountain_Cr_nr_Col_Spgs=data.frame(lon=-104.8782, lat=38.85422, id='07105500')) 
  ReadChrtgrid(('~/wrfHydroTestCases/FRN.REACH/OUTPUT',gaugePtlist=gaugePtlist) 
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadChrtout}}, \code{\link{ReadFrxstPts}},
  \code{\link{ReadGwOut}}, \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRouteLink}},
  \code{\link{ReadRtout}}

Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadChrtout}}, \code{\link{ReadFrxstPts}},
  \code{\link{ReadGwOut}}, \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRouteLink}},
  \code{\link{ReadRtout}}
}
\concept{dataGet}
\concept{modelDataReads}
\keyword{IO}
\keyword{ts}
\keyword{univar}
