% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadChrtout}
\alias{ReadChrtout}
\title{Read WRF-Hydro RTOUT data files and generate basin-wide mean water fluxes.}
\usage{
ReadChrtout(pathOutdir, idList = NULL, ncores = 1)
}
\arguments{
\item{pathOutdir}{The full pathname to the output directory containing the 
RTOUT files.}

\item{idList}{Optional list of station IDs to import (must be consistent
with IDs as used in the "station_id" variable.}

\item{ncores}{If multi-core processing is available, the number of cores to 
use (DEFAULT=1). Must have doMC installed if ncores is more than 1.}
}
\value{
A dataframe containing a time series of channel variables.
}
\description{
\code{ReadChrtout} reads in WRF-Hydro CHRTOUT files and outputs a time series of
fluxes.
}
\details{
\code{ReadChrtout} reads standard-format WRF-Hydro CHRTOUT NetCDF files and 
outputs a time series of fluxes.

OUTPUT CHRTOUT water budget variables:
\itemize{
   \item QSTRMVOLRT: Mean accumulated depth of stream channel inflow (mm)
   \item SFCHEADSUBRT: Mean depth of ponded water (mm)
   \item QBDRYRT: Mean accumulated flow volume routed outside of the domain 
             from the boundary cells (mm)
}
}
\examples{
## Take an OUTPUT directory for an hourly routing timestep model run of 
## Fourmile Creek (Basin ID = 1) and create a new dataframe containing the 
## basin-wide mean values for the major water budget components over the 
## time series.

\dontrun{
modRtout1h.mod1.fc <- 
  ReadRtout("../RUN.MOD1/OUTPUT", "../DOMAIN/Fulldom_hires_hydrofile_4mile.nc", 
            basid=1, ncores=16)
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadFrxstPts}}, \code{\link{ReadGwOut}},
  \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}
\keyword{ts}
\keyword{univar}

