% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_observations.R
\name{ReadCoDwrGage}
\alias{ReadCoDwrGage}
\title{Read standard-format text data table downloaded from CO DWR}
\usage{
ReadCoDwrGage(pathGageData, returnMetric = TRUE, returnEnglish = TRUE,
  timeZone = "America/Denver")
}
\arguments{
\item{pathGageData}{The full pathname to the streamflow/stage time series
text file as downloaded from CO DWR. The file should include the standard
CO DWR header info and the data time series with standard CO DWR columns.}

\item{returnMetric}{Logical: Return variables in these units (cms and m)}

\item{returnEnglish}{Logical: Return variables in english units (cfs and ft)}

\item{timeZone}{(OPTIONAL) The time zone for the gage data
(DEFAULT="America/Denver", which is MST/MDT). Time zone name must be
R-friendly for your current OS. See:
\url{http://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html}}
}
\value{
A dataframe containing the flow and/or stage data.
}
\description{
\code{ReadCoDwrGage} reads CO DWR gage data table (streamflow and/or stage)
and puts into a dataframe.
}
\details{
\code{ReadCoDwrGage} reads a streamflow or stage time series data table
(standard CO DWR data format) and outputs a dataframe with consistent date
and data columns for use with other rwrfhydro tools.
}
\examples{
## Take a text file downloaded from the CO DWR website for 15-minute flow 
## at Alamosa River and create a dataframe called "obsStr15min.alaterco".

\dontrun{
obsStr15min.alaterco <- 
  ReadCoDwrGage("../OBS/STRFLOW/ALATERCO_41715103512.txt")
obsStr15min.alaterco <- 
  ReadCoDwrGage("../OBS/STRFLOW/ALATERCO_41715103512.txt", returnEnglish=FALSE)
}
}
\seealso{
Other obsDataReads: \code{\link{GetStage4Files}},
  \code{\link{ReadAmerifluxCSVOldFormat}},
  \code{\link{ReadAmerifluxNCOldFormat}},
  \code{\link{ReadUsgsGage}}
}
\concept{
dataGet
}
\keyword{IO}
