% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadFrxstPts}
\alias{ReadFrxstPts}
\title{Read WRF-Hydro standard-format forecast points output text file.}
\usage{
ReadFrxstPts(pathOutfile)
}
\arguments{
\item{pathOutfile}{The full pathname to the WRF-Hydro forecast points text file
(frxst_pts_out.txt).}
}
\value{
A dataframe containing the forecast points output flow data.
}
\description{
\code{ReadFrxstPts} reads in WRF-Hydro forecast points output text file.
}
\details{
\code{ReadFrxstPts} reads a standard-format WRF-Hydro forecast points output text
file and creates a dataframe with consistent date and data columns for use with other
rwrfhydro tools.
}
\examples{
## Take a forecast point output text file for an hourly model run of Fourmile Creek
## and return a dataframe.

modStr1h.mod1.fc <- ReadFrxstPts("../OUTPUT/frxst_pts_out.txt")
}
\seealso{
Other modelDataReads: \code{\link{ReadGwOut}};
  \code{\link{ReadLdasoutWb}}; \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}

