% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{ReadGwbuckFile}
\alias{ReadGwbuckFile}
\title{Read groundwater bucket netcdf parameter file into dataframe.}
\usage{
ReadGwbuckFile(gwbuckFile)
}
\arguments{
\item{gwbuckFile}{Path to groundwater bucket parameter netcdf file}
}
\value{
Dataframe
}
\description{
\code{ReadGwbuckFile} reads in a netcdf groundwater bucket parameter file 
and generates a dataframe
}
\keyword{utilities}
