% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadGwOut}
\alias{ReadGwOut}
\title{Read WRF-Hydro standard-format groundwater output text file.}
\usage{
ReadGwOut(pathOutfile)
}
\arguments{
\item{pathOutfile}{The full pathname to the WRF-Hydro groundwater text file
(GW_inflow.txt, GW_outflow.txt, or GW_zlev.txt).}
}
\value{
A dataframe containing the groundwater data.
}
\description{
\code{ReadGwOut} reads in WRF-Hydro groundwater output text file.
}
\details{
\code{ReadGwOut} reads a standard-format WRF-Hydro groundwater output text file
(GW_inflow.txt, GW_outflow.txt, or GW_zlev.txt) and creates a dataframe with consistent
date and data columns for use with other rwrfhydro tools.
}
\examples{
## Take a groundwater outflow text file for an hourly model run of Fourmile Creek
## and return a dataframe.
\dontrun{
modGWout1h.mod1.fc <- ReadGwOut("../OUTPUT/GW_outflow.txt")
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadFrxstPts}}, \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}

