% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadLakeout}
\alias{ReadLakeout}
\title{Read LAKEOUT files from gridded lake routing option and create data frame/table for selected lakes.
\code{ReadLakeout} reads in the WRF-Hydro LAKEOUT files from gridded lake routing: Community Release version.}
\usage{
ReadLakeout(pathOutdir = NULL, lakeidList = NULL, parallel = FALSE,
  useDatatable = TRUE, pattern = glob2rx("*.LAKEOUT_DOMAIN*"))
}
\arguments{
\item{pathOutdir}{The full pathname to the WRF-Hydro lake output files.}

\item{lakeidList}{The list of lakeids to put in the output data table. These lakeids correspond to the lakes in the LAKEGRID, LAKEPARM and lakes.shp files that are
created in the routing files from the ArcGIS pre-processing tools.}

\item{parallel}{Logical for running in parallel mode (must have a parallel
backend installed and registered (e.g., doMC or doParallel) (DEFAULT=FALSE)}

\item{useDatatable}{Logical for utilizing the data.table package and
outputting in data.table format (DEFAULT=TRUE)}

\item{pattern}{Pattern to match in the model output (e.g. *LAKEOUT_DOMAIN1*)}
}
\description{
Read LAKEOUT files from gridded lake routing option and create data frame/table for selected lakes.
\code{ReadLakeout} reads in the WRF-Hydro LAKEOUT files from gridded lake routing: Community Release version.
}
\examples{
# This function loops through LAKEOUT files in the pathOutdir and creates a data frame or 
# data table to store inflow, outflow, elevation and the station_id (lakeid).  The lakeids 
# correspond to the lakes in the LAKEGRID, LAKEPARM and lakes.shp files that are
# created in the routing files from the ArcGIS pre-processing tools.
\dontrun{
ReadLakeout('~/wrfHydroTestCases/FRN.REACH/OUTPUT',
     lakeidList=c("1", "5"))
}
}

