% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadLdasoutWb}
\alias{ReadLdasoutWb}
\title{Read WRF-Hydro (w/NoahMP) LDASOUT data files and generate basin-wide mean water 
budget variables.}
\usage{
ReadLdasoutWb(pathOutdir, pathDomfile, mskvar = "basn_msk", basid = 1,
  aggfact = 10, pattern = glob2rx("*LDASOUT_DOMAIN*"), parallel = FALSE)
}
\arguments{
\item{pathOutdir}{The full pathname to the output directory containing the 
LDASOUT files.}

\item{pathDomfile}{The full pathname to the high-res hydro domain NetCDF file 
used in the model run (for grabbing the basin mask).}

\item{mskvar}{The variable name in pathDomfile to use for the mask 
(DEFAULT="basn_msk").}

\item{basid}{The basin ID to use (DEFAULT=1)}

\item{aggfact}{The high-res (hydro) to low-res (LSM) aggregation factor 
(e.g., for a 100-m routing grid and a 1-km LSM grid, aggfact = 10)}

\item{pattern}{Pattern to match in the model output 
(DEFAULT=glob2rx('*LDASOUT_DOMAIN*'))}

\item{parallel}{Logical for running in parallel mode (must have a parallel
backend installed and registered (e.g., doMC or doParallel) (DEFAULT=FALSE)}
}
\value{
A dataframe containing a time series of basin-wide mean water 
budget variables.
}
\description{
\code{ReadLdasoutWb} reads in WRF-Hydro (w/NoahMP) LDASOUT files and outputs a 
time series of basin-wide mean variables for water budget calculations.
}
\details{
\code{ReadLdasoutWb} reads standard-format WRF-Hydro (w/NoahMP) LDASOUT NetCDF 
files and calculates basin-wide mean values for each time step suitable for running 
basin water budget calculations.

OUTPUT NoahMP LDASOUT water budget variables:
\itemize{
   \item ACCECAN: Mean accumulated canopy evaporation (mm)
   \item ACCEDIR: Mean accumulated surface evaporation (mm)
   \item ACCETRAN: Mean accumulated transpiration (mm)
   \item ACCPRCP: Mean accumulated precipitation (mm)
   \item CANICE: Mean canopy ice storage (mm)
   \item CANLIQ: Mean canopy liquid water storage (mm)
   \item SFCRNOFF: Mean surface runoff from LSM \emph{(meaningful for an 
             LSM-only run)} (mm)
   \item SNEQV: Mean snowpack snow water equivalent (mm)
   \item UGDRNOFF: Mean subsurface runoff from LSM \cr \emph{(meaningful 
             for an LSM-only run)} (mm)
   \item SOIL_M1: Mean soil moisture storage in soil layer 1 (top) (mm)
   \item SOIL_M2: Mean soil moisture storage in soil layer 2 (mm)
   \item SOIL_M3: Mean soil moisture storage in soil layer 3 (mm)
   \item SOIL_M4: Mean soil moisture storage in soil layer 4 (bottom) (mm)
}
}
\examples{
## Take an OUTPUT directory for a daily LSM timestep model run of Fourmile Creek and
## create a new dataframe containing the basin-wide mean values for the major water
## budget components over the time series.

\dontrun{
library(doMC)
registerDoMC(3)
modLdasoutWb1d.mod1.fc <- 
  ReadLdasoutWb("../RUN.MOD1/OUTPUT", "../DOMAIN/Fulldom_hires_hydrofile_4mile.nc", 
                parallel=TRUE)
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadChrtout}}, \code{\link{ReadFrxstPts}},
  \code{\link{ReadGwOut}}, \code{\link{ReadLdasoutAll}},
  \code{\link{ReadRouteLink}}, \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}
\keyword{ts}
\keyword{univar}

