% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{ReadLinkFile}
\alias{ReadLinkFile}
\title{Read route link file into dataframe.}
\usage{
ReadLinkFile(linkFile)
}
\arguments{
\item{linkFile}{Path to route link file}
}
\value{
Dataframe
}
\description{
\code{ReadLinkFile} reads in a route link file and generates a dataframe
}
\keyword{utilities}
