% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_to_ncdf.R
\name{ReadNcTimeSlice}
\alias{ReadNcTimeSlice}
\title{Read a USGS discharge data timeslice from a netcdf file.}
\usage{
ReadNcTimeSlice(file)
}
\arguments{
\item{file, }{the ncdf file to read.}
}
\description{
This is kind of the inverse of WriteNcTimeSlice to be used in extending existing
timeslices on file.
}
\examples{
\dontrun{
sliceFiles <- list.files('~/usgsStreamData/timeSliceData/','.*', full.names=TRUE)
ReadNcTimeSlice(tail(sliceFiles,1))
}
}
\keyword{internal}
