% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_observations.R
\name{ReadNearStationData}
\alias{ReadNearStationData}
\title{Read point NetCDF station time series observations}
\usage{
ReadNearStationData(pathData)
}
\arguments{
\item{pathData}{The full pathname to the point NetCDF file
\url{http://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html}}
}
\value{
A list containing station observations.
}
\description{
\code{ReadNearStationData} reads station observation data table (point NetCDF file) and creates
 a list containing dataframe of time series and metadata.
}
\details{
\code{ReadNearStationData} reads station observation  NetCDF file
and outputs a dataframe with consistent date and data columns for use with
other rwrfhydro tools. Includes observation metadata useful for plotting.
}
\examples{
## Takes a point NetCDF file of station observations
## (Upper Canejos River) and returns a dataframe.
URG1 <- ReadNearStationData("URG1_Upper_Conejos_River_thru_Jul_1_2015.csv_filled.csv.nc")
URG1[[1]] #data frame with observations
}
\seealso{
Other obsDataReads: \code{\link{PlotStationData}};
  \code{\link{ReadAmerifluxCSV}};
  \code{\link{ReadAmerifluxNC}};
  \code{\link{ReadCoDwrGage}}; \code{\link{ReadUsgsGage}}
}
\concept{
dataGet
}
\keyword{IO}

