% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadRouteLink}
\alias{ReadRouteLink}
\title{Read in WRF-Hydro route link file}
\usage{
ReadRouteLink(linkFile)
}
\arguments{
\item{linkFile}{Full path to route link file}
}
\value{
Dataframe of route link data
}
\description{
\code{ReadRouteLink} is simply a usage of format.
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadChrtgrid}}, \code{\link{ReadChrtout}},
  \code{\link{ReadFrxstPts}}, \code{\link{ReadGwOut}},
  \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}
