% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadRtout}
\alias{ReadRtout}
\title{Read WRF-Hydro RTOUT data files and generate basin-wide mean water fluxes.}
\usage{
ReadRtout(pathOutdir, pathDomfile, mskvar = "basn_msk", basid = 1,
  ncores = 1)
}
\arguments{
\item{pathOutdir}{The full pathname to the output directory containing the RTOUT files.}

\item{pathDomfile}{The full pathname to the high-res hydro domain NetCDF file used in
the model run (for grabbing the basin mask).}

\item{mskvar}{The variable name in pathDomfile to use for the mask (DEFAULT="basn_msk").}

\item{basid}{The basin ID to use (DEFAULT=1)}

\item{ncores}{If multi-core processing is available, the number of cores to use (DEFAULT=1).
Must have doMC installed if ncores is more than 1.}
}
\value{
A dataframe containing a time series of basin-wide mean water budget variables.
}
\description{
\code{ReadRtout} reads in WRF-Hydro RTOUT files and outputs a time series of
basin-wide mean water fluxes for water budget.
}
\details{
\code{ReadRtout} reads standard-format WRF-Hydro RTOUT NetCDF files and calculates
basin-wide mean values for each time step for water budget terms.

OUTPUT RTOUT water budget variables:
\itemize{
   \item QSTRMVOLRT: Mean accumulated depth of stream channel inflow (mm)
   \item SFCHEADSUBRT: Mean depth of ponded water (mm)
   \item QBDRYRT: Mean accumulated flow volume routed outside of the domain from the boundary cells (mm)
}
}
\examples{
## Take an OUTPUT directory for an hourly routing timestep model run of Fourmile Creek (Basin ID = 1)
## and create a new dataframe containing the basin-wide mean values for the major water budget
## components over the time series.

\dontrun{
modRtout1h.mod1.fc <-
  ReadRtout("../RUN.MOD1/OUTPUT", "../DOMAIN/Fulldom_hires_hydrofile_4mile.nc",
            basid=1, ncores=16)
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}};
  \code{\link{ReadFrxstPts}}; \code{\link{ReadGwOut}};
  \code{\link{ReadLdasoutWb}}
}
\concept{
dataGet
}
\keyword{IO}
\keyword{ts}
\keyword{univar}

