% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{readSnodasDepthSweDate}
\alias{readSnodasDepthSweDate}
\title{readSnodasDepthSweDate}
\usage{
readSnodasDepthSweDate(datePOSIXct, outputDir = ".")
}
\arguments{
\item{datePOSIXct}{The date in POSIXct format for which data is desired. Only whole days matter.}

\item{outputDir}{The directory where the data are archived.}
}
\value{
A list with a SWE and a depth matrix.
}
\description{
\code{GetSnodasDepthSweDate} Read (from local disk into memory) the SNODAS snow depth and SWE for a given date.
}
\examples{
snodasGot <- getSnodasDepthSweDate(as.POSIXct('2015-02-28'))
if(snodasGot) snodasList <- readSnodasDepthSweDate(as.POSIXct('2015-02-28'))
}

