% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observations.R
\name{ReadUsgsGage}
\alias{ReadUsgsGage}
\title{Read standard-format text data table downloaded from USGS Water Data}
\usage{
ReadUsgsGage(pathGageData, timeZone)
}
\arguments{
\item{pathGageData}{The full pathname to the streamflow/stage time series text file as
downloaded from USGS Water Data. The file should include the standard USGS header info and
the data time series with standard USGS columns.}

\item{timeZone}{The time zone for the USGS gage data. Time zone name must be R-friendly for your
current OS. See: \url{http://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html}}
}
\value{
A dataframe containing the USGS flow and/or stage data.
}
\description{
\code{ReadUsgsGage} reads USGS data table (streamflow and/or stage) and puts into a dataframe.
}
\details{
\code{ReadUsgsGage} reads a streamflow or stage time series data table (standard USGS Water
Data format) and outputs a dataframe with consistent date and data columns for use with other
rwrfhydro tools.
}
\examples{
## Take a text file downloaded from the USGS Water Data website for 5-minute flow at Fourmile Creek
## and create a dataframe called "obsStr5min.fc".

obsStr5min.fc <- ReadUsgsGage("../OBS/STRFLOW/5min_str_06727500_110401_140810.txt", "America/Denver")
}

