% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_mod.R
\name{readWghtFile}
\alias{readWghtFile}
\title{Open weight file and extract weight arrays used for ESMF regridding.}
\usage{
readWghtFile(wghtFile, nxSrc, nySrc, nxDst, nyDst)
}
\arguments{
\item{wghtFile}{Weight file opened for extract.}

\item{nxSrc}{Integer number of columns of source data.}

\item{nySrc}{Integer number of rows of source data.}

\item{nxDst}{Integer number of columns of destination data.}

\item{nyDst}{Integer number of rows of destination data.}
}
\value{
list with regridding weight arrays
}
\description{
\code{readWghtFile} Open weight file and extract important weight arrays.
}
\examples{
\dontrun{
list <- readWghtFile('./hrrr_wghts_rio_grande.nc',
nxIn,nyIn,nxGeo,nyGeo)
}
}
\concept{
regrid, ESMF
}
\keyword{ESMF}
\keyword{geospatial,}
\keyword{regrid,}

