% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{ReadWtFile}
\alias{ReadWtFile}
\title{Read weight file into list of dataframes.}
\usage{
ReadWtFile(wtFile)
}
\arguments{
\item{wtFile}{Path to weight file}
}
\value{
List of 2 dataframes
}
\description{
\code{ReadWtFile} reads in a spatial weight file and generates a list of two 
dataframes: 1 for the gridded info (data) and 2 for the poly info (poly)
}
\keyword{utilities}
