% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{ReExpNetwork}
\alias{ReExpNetwork}
\title{ReExpress stream networks indexed network traversal.}
\usage{
ReExpNetwork(routeLinkReInd, upstream = TRUE)
}
\arguments{
\item{routeLinkReInd}{The netcdf routelink file to process.}

\item{upstream}{Logical, re-express connectivity upstream (TRUE) or downstream (FALSE).}
}
\value{
The resulting file which was written to disk, of the form the "infile.reExpTo.nc"
        (downstream) or "infile.reExpFrom.nc" (upstream).
}
\description{
\code{ReExpNetwork} re-expresses topological relationships between three variables,
up- and down- stream traversal, depending on the upstream argument.
}
\examples{
\dontrun{
  library(rwrfhydro)
  doMC::registerDoMC(16)
  ReExpNetwork("/home/jamesmcc/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink.reInd.Rdb")
  ReExpNetwork("/home/jamesmcc/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink.reInd.Rdb", up=FALSE)

  nContrib<-from$end-from$start
  nContrib[which(from$start>0)] <-nContrib[which(from$start>0)] +1
  table(nContrib)

 nOut<-to$end-to$start
 nOut[which(to$start>0)] <-nOut[which(to$start>0)] +1
 table(nOut)

}
}
\seealso{
Other networkExpression: \code{\link{CheckConn}};
  \code{\link{ReExpressGridChanConn}};
  \code{\link{ReIndexRouteLink}}
}
\concept{
dataMgmt
}
\keyword{manip}

