% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_expression.R
\name{ReExpressRouteLink}
\alias{ReExpressRouteLink}
\title{ReExpress the Route_Link.nc file.}
\usage{
ReExpressRouteLink(routeLink.nc = "Route_Link.nc", parallel = FALSE)
}
\arguments{
\item{routeLink.nc}{Character path/file to the desired Route_link.nc netcdf file for the link/reach-based routing.}

\item{parallel}{Logical use a registered backend for plyr?}
}
\value{
Named Character vector for each of the 4 files created as outputs with full paths.
}
\description{
A wrapper on the individual functions which perform this full reexpression and generate the netcdf file.
}
\examples{
\dontrun{
ReExpressRouteLink("~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.nc")
ReExpressRouteLink() # if there is a Route_Link.nc file in getwd() (the current directory).
}
}
\seealso{
Other networkExpression nudging: \code{\link{GatherStreamInds}},
  \code{\link{NtwKReExToNcdf}}, \code{\link{ReExpNetwork}},
  \code{\link{ReIndexRouteLink}},
  \code{\link{VisualizeRouteLink}},
  \code{\link{VisualizeSubsetStream}}
}
\concept{
dataMgmt nudging
}
\keyword{manip}

