% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_mod.R
\name{regrid}
\alias{regrid}
\title{High level call to regrid data to WRF-Hydro domain.}
\usage{
regrid(dataIn, latIn, lonIn, geoFile, method, wghtFile, ndvSrc)
}
\arguments{
\item{dataIn}{the stack of matrices of data to be regridded.}

\item{latIn}{2D array of latitude values of the center 
of each pixel cell of the data being regridded.}

\item{lonIn}{2D array of longitude values of the center
of each pixel cell of the data being regridded.}

\item{geoFile}{geogrid file of WRF-Hydro domain.}

\item{method}{integer of regridding. Options are 2 - "conserve",
1 - "bilinear", 3 - "nneighbor".}

\item{wghtFile}{NetCDF weight file that contains regridding 
weight arrays to be passed to ESMF for sparse matrix multiplication.}

\item{ndvSrc}{Missing value of source data being regridded. 
used as a mask value during regridding calls.}
}
\value{
dataOut the stack of regridded matrices.
}
\description{
\code{regrid} High level call to regrid data.
}
\examples{
\dontrun{
dataOut <- regrid(hrrrDSWRF,latIn,lonIn,'geo_em.d01.nc',1,'./wghts_hrrr.Rdata',
                  9999)
}

}
\seealso{
Other regrid: \code{\link{genWghtFile}}
}
\concept{
regrid geospatial
}
\keyword{ESMF}
\keyword{geospatial}
\keyword{io}
\keyword{regrid}

