% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_multi_GRIB.R
\name{RegridGRIB}
\alias{RegridGRIB}
\title{Extraction and regridding of GRIB data for multiple files for a given variable.}
\usage{
RegridGRIB(files, var, levType, lev, wghtFile, geoFile, method)
}
\arguments{
\item{files}{The files vector list}

\item{var}{The GRIB variable being regridded}

\item{levType}{Level type to extract GRIB data. Examples of levType would
be "surface", "cloudTop", "heightAboveGround"}

\item{lev}{The vertical level being regridded}

\item{wghtFile}{The weight regridding file. If it does not exist, it will 
be created.}

\item{geoFile}{The Geogrid LSM file needed for regridding.}

\item{method}{Method of regridding. Current acceptable values are 
"bilinear"}
}
\value{
A dataframe containing the regridded stack.
}
\description{
\code{RegridGRIB}: For a list of GRIB files, extract and regrid a variable
of a given vertical level to a WRF-Hydro domain.
}
\seealso{
Other regridMultiGRIB: \code{\link{RegridMultiGRIBVar}}
}
\concept{
dataRegrid
}
\keyword{internal}

