% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_multi_GRIB.R
\name{RegridMultiGRIBFile}
\alias{RegridMultiGRIBFile}
\title{The RegridMultiGRIB for individual file groups.}
\usage{
RegridMultiGRIBFile(fileInd, fileList, varList, levTypeList, levList, wghtList,
  geoFile, method)
}
\arguments{
\item{fileInd}{The index of the file group.}

\item{fileList}{The list of the file group.}

\item{varList}{The variable list.}

\item{levTypeList}{The level type list.}

\item{levList}{The vertical level list.}

\item{wghtList}{The weight file list.}

\item{geoFile}{The geofile needed for regridding.}

\item{method}{Method of regridding. Current acceptable values are 
"bilinear".}
}
\value{
A dataframe
}
\description{
\code{RegridMultiGRIBFile} : For this file group, regrid 
Regrid GRIB data spread over multiple files. Typically called
by RegridMultiGRIB.
}
\keyword{dataRegrid}
\keyword{internal}
\keyword{regridMultiGRIB}

