% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_multi_GRIB.R
\name{RegridMultiGRIBVar}
\alias{RegridMultiGRIBVar}
\title{The RegridMultiGRIB for getting variables out of individual files.}
\usage{
RegridMultiGRIBVar(varInd, varList, levTypeList, levList, files, geoFile,
  method)
}
\arguments{
\item{varInd}{The variable index.}

\item{varList}{The variable list.}

\item{levTypeList}{The list of level types for each variable.}

\item{levList}{The vertical level list.}

\item{files}{The files vector.}

\item{geoFile}{The Geogrid LSM file for regridding.}

\item{method}{Method of regridding. Current acceptable values are
"bilinear", and "conserve"}
}
\value{
A dataframe
}
\description{
\code{RegridMultiGRIBVar}: For a group of GRIB files, extract
and regrid a list of variables (for a specified level) to a WRF-Hydro domain.
}
\seealso{
Other regridMultiGRIB: \code{\link{RegridGRIB}}
}
\keyword{dataRegrid}
\keyword{internal}

