% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_multi_ncdf.R
\name{regridMultiNcdf}
\alias{regridMultiNcdf}
\title{Regrid multiple variables across various NetCDF file groups to the
WRF-Hydro domain.}
\usage{
regridMultiNcdf(fileList, varList, latList, lonList, wghtList, geoFile, method)
}
\arguments{
\item{fileList}{The list of file groups. Names must match those in the
other lists.}

\item{varList}{The list of NetCDF variables for each file group.
Names must match fileList.}

\item{latList}{The list of latitude variables for each file group.}

\item{lonList}{The list of longitude variables for each file group.}

\item{wghtList}{The list of weight files for each file group.}

\item{geoFile}{The path to the Geo LSM file used for regridding.}

\item{method}{Method of regridding. Current acceptable values are
"bilinear"}
}
\value{
A list containing regridded data along with meta-data.
}
\description{
\code{RegridMultiNcdf} is designed to regrid multiple variables across
multiple NetCDF files. Two collated lists specify 1) file groups,
and 2) variables for each file group. The names of the lists must match.
In addition, a Geo LSM file must be presented for regridding to the WRF-Hydro
domain. NetCDF input files must have center lat/lon values for the entire
grid. Also, a regridding method must be provided to specify how ESMF will regrid
the data. Valid options currently include "bilinear". See examples for details.
}
\examples{
##This example shows opening multiple SNODAS NetCDF conus files containing
##snow depth and SWE for regridding.
\dontrun{
snodasPath <- '/d4/karsten/data/SNODAS'
NCFiles <- list.files(path=snodasPath, pattern='SNODAS_CONUS_', full.names=TRUE)
#fileList <- these are the groups of files.
fileList <- list(SNODASList1 = NCFiles)
#varList <- Define which variables are desired for each file group.
NCVars <- list(SWE='SNEQV', DEPTH='SNOWH')
varList <- list(SNODASList1 = NCVars)
latVars <- list(lat='latitude')
latList <- list(SNODASList1 = latVars)
lonVars <- list(lon='longitude')
lonList <- list(SNODASList1 = lonVars)
wghtFiles <- list(wght1 = "./wght_snodas_ioc_1km.nc")
wghtList <- list(SNODASList1 = wghtFiles
geoFile <- '/d4/karsten/geospatial/geo_em.d01.nc'
regridData <- regridMultiNcdf(fileList=fileList,varList=varList,latList=latList,lonList=lonList,
              wghtList=wghtList,geoFile=geoFile,'bilinear')

}
}

