% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_multi_ncdf.R
\name{regridMultiNcdfFile}
\alias{regridMultiNcdfFile}
\title{The regridMultiNcdf for individual file groups.}
\usage{
regridMultiNcdfFile(fileInd, fileList, varList, latList, lonList, wghtList,
  geoFile, method)
}
\arguments{
\item{fileInd}{The index of the file group.}

\item{fileList}{The list of the file group.}

\item{varList}{The variable list.}

\item{latList}{The latitude variable list.}

\item{lonList}{The longitude variable list.}

\item{wghtList}{The weight file list.}

\item{geoFile}{The geofile needed for regridding.}

\item{method}{Method of regridding. Current acceptable values are 
'bilinear'.}
}
\value{
A list
}
\description{
\code{regridMultiNcdfFile} : For this file group, regrid
NetCDF data spread over multiple files. Typically called by 
regridMultiNcdf.
}
\keyword{dataRegrid}
\keyword{internal}
\keyword{regridMultiNcdf}

