% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_multi_ncdf.R
\name{regridMultiNcdfVar}
\alias{regridMultiNcdfVar}
\title{The regridMultiNcdf for getting variables out of individual files.}
\usage{
regridMultiNcdfVar(varInd, varList, files, latVar, lonVar, wghtFile, geoFile,
  method)
}
\arguments{
\item{varInd}{The variable index.}

\item{varList}{The variable list.}

\item{files}{The files vector.}

\item{latVar}{The latitude variable}

\item{lonVar}{The longitude variable}

\item{wghtFile}{The weight file used for regridding.}

\item{geoFile}{The Geogrid LSM file for regridding.}

\item{method}{Method of regridding. Current acceptable values are
"bilinear"}
}
\value{
A list
}
\description{
\code{regridMultiNcdfVar}: For a group of NetCDF files, extract
and regrid a list of variables to a WRF-Hydro domain.
}
\seealso{
Other regridMultiNcdf: \code{\link{regridNcdf}}
}
\keyword{dataRegrid}
\keyword{internal}

