% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{ReIndexRouteLink}
\alias{ReIndexRouteLink}
\title{ReIndex RouteLink.nc (netcdf) files for indexed network traversal.}
\usage{
ReIndexRouteLink(routeLinkFile)
}
\arguments{
\item{routeLinkFile}{The netcdf routelink file to process.}
}
\value{
The resulting file which was written to disk, of the form "infile.reInd.nc"
}
\description{
\code{ReIndexRouteLink} reIndexes by order in the RouteLink file,
replacing ComID with this index.
}
\examples{
\dontrun{
  ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink3.nc')
  ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink.nc')
}
}
\seealso{
Other networkExpression: \code{\link{CheckConn}};
  \code{\link{ReExpNetwork}};
  \code{\link{ReExpressGridChanConn}}
}
\concept{
dataMgmt
}
\keyword{manip}

