% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReshapeMultiNcdf}
\alias{ReshapeMultiNcdf}
\title{"Flatten" the output from GetMultiNcdf}
\usage{
ReshapeMultiNcdf(myDf)
}
\arguments{
\item{myDf}{The output dataframe from GetMultiNcdf.}
}
\value{
The reshaped output dataframe(s) (a single dataframe if
only one fileGroup, a list of multiple dataframes if multiple
fileGroups.
}
\description{
\code{ReshapeMultiNcdf} flattens the output from GetMultiNcdf.
Take the output dataframe from GetMultiNcdf and reshape the dataframe
for ease of use in other functions.
}
\keyword{internal}
\keyword{utilities}
