% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_poly.R
\docType{data}
\name{rfc}
\alias{rfc}
\title{Polygon shapefile of River Forecast Center boundaries.}
\format{SpatialPolygonsDataFrame}
\usage{
rfc
}
\description{
Polygon shapefile of River Forecast Center boundaries.
}
\concept{
data
}
\keyword{data}

