% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{Rmse}
\alias{Rmse}
\title{Calculate root mean squared error.
\code{Rmse} calculates the root mean squared error.
Calculate the root mean squared error for vectors
of modelled and observed values.}
\usage{
Rmse(m, o)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The root mean squared error.
}
\description{
Calculate root mean squared error.
\code{Rmse} calculates the root mean squared error.
Calculate the root mean squared error for vectors
of modelled and observed values.
}

