% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{RotateCcw}
\alias{RotateCcw}
\title{Rotate a matrix counter-clock-wise.}
\usage{
RotateCcw(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
\code{RotateCcw} Rotates a matrix counter-clock-wise.
}
\examples{
x <- matrix(1:9, 3)
x
RotateCcw(x)
RotateCcw(RotateCcw(x))
}

