<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs
% Please edit documentation in R/usgs_time_slice.R
\name{RoundMinutes}
\alias{RoundMinutes}
\title{Round a POSIXct time to the nearest mth minute.}
\usage{
RoundMinutes(POSIXct, nearest = 5)
}
\arguments{
\item{POSIXct,}{POSIXct class vector.}

<<<<<<< HEAD
\item{nearest,}{Numeric, the time resolution to which the passed times are
rounded. Must evenly divide 60.}
}
\value{
A vector of POSIXct class corresponding to the input, rounded to the
  nearest \code{nearest} minutes.
=======
\item{nearest,}{Numeric, the time resolution to which the passed times are rounded.
Must evenly divide 60.}
}
\value{
A vector of POSIXct class corresponding to the input, rounded to the nearest
\code{nearest} minutes.
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs
}
\description{
Round a POSIXct time to the nearest mth minute.
}
\examples{
print(time <- Sys.time()+lubridate::minutes(1:10))
RoundMinutes(time, nearest=5)
}
<<<<<<< HEAD
\keyword{internal}
=======
>>>>>>> parent of 3e9dc0d... Merge branch 'master' of github.com:mccreigh/rwrfhydro into procUsgs

