\name{rwrfhydro}
\alias{rwrfhydro}
\alias{rwrfhydro}
\docType{package}
\title{ rwrfydro}
\description{
Tools for working with WRF Hydro data.
}
\details{
\tabular{ll}{
Package: \tab rwrfhydro\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-13\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
James McCreight, Aubrey Dugger, ... 
Maintainer: James McCreight <jamesmcc@ucar.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ WRF-Hydro wrfHydro hydrology }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
