% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{SaveHucData}
\alias{SaveHucData}
\title{Save the output of GetUsgsHucData to an archive.}
\usage{
SaveHucData(hucData, outPath,
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData",
  overwriteHucDataFile = FALSE)
}
\arguments{
\item{hucData}{List Returned from GetUsgsHucData.}

\item{outPath}{Character The path to the database directory}

\item{metaDBFileName}{Character The name of the database metadata file.}

\item{overwriteHucDataFile}{Logical Replace/overwrite the existing data file for the HUC with the current data.}
}
\description{
\code{SaveHucData} take a list returned by \code{GetUsgsHucData}, create or append to an existing archive, and update
metadata file. Input list is grouped by HUC codes where each HUC has data and meta lists. The output files
are written to outPath with the following format HHHHHHHH.data.RData where HHHHHHHH is the HUC8 code. The
name of the metadata file is configurable. The metadata file contains metadata for all files in the outPath.
}
\examples{
files <- SaveHucData(GetUsgsHucData(huc='10190005'), outPath='/Users/jamesmcc/streamflow/OBS') ##boulder, CO
files <- SaveHucData(GetUsgsHucData(huc='03160203'), outPath='/Users/jamesmcc/streamflow/OBS') ##satilpa, AL
files <- SaveHucData(GetUsgsHucData(huc='03020104'), outPath='/Users/jamesmcc/streamflow/OBS') ##vans, NC
}

