% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RS_get.R
\name{SmoothStack}
\alias{SmoothStack}
\title{Run MODIS-R Whittaker smoothing over pre-processed raster stack.}
\usage{
SmoothStack(inStack, w = NULL, t = NULL, lambda = 5000, nIter = 3,
  outputAs = "one", collapse = FALSE, outDirPath = "./",
  removeOutlier = FALSE, outlierThreshold = NULL, mergeDoyFun = "max",
  ...)
}
\arguments{
\item{inStack}{The name of the raster stack to smooth.}

\item{w}{FROM \code{\link[MODIS]{whittaker.raster}}: In case of MODIS
composite the 'VI_Quality' raster-Brick, Stack or filenames. Use preStack
functionality to ensure the right input.}

\item{t}{FROM \code{\link[MODIS]{whittaker.raster}}: In case of MODIS
composite the 'composite_day_of_the_year' raster-Brick, Stack or filenames.
Use preStack functionality to ensure the right input.}

\item{lambda}{FROM \code{\link[MODIS]{whittaker.raster}}: _Yearly_ lambda
value passed to ?ptw:::wit2. If set as character (i.e. lambda="600"), it is
not adapted to the time serie length but used as a fixed value (see
details). High values = stiff/rigid spline}

\item{nIter}{FROM \code{\link[MODIS]{whittaker.raster}}: Number of iteration
for the upper envelope fitting.}

\item{outputAs}{FROM \code{\link[MODIS]{whittaker.raster}}: Character.
Organisation of output files: single each date one RasterLayer, yearly a
RasterBrick for each year, one one RasterBrick for the entire time-serie.}

\item{collapse}{FROM \code{\link[MODIS]{whittaker.raster}}: logical, if TRUE
the input data is treated as _1_single_Year_ collapsing the data using the
Julian date information without the year.}

\item{outDirPath}{FROM \code{\link[MODIS]{whittaker.raster}}: Output path
default is the current directory.}

\item{removeOutlier}{FROM \code{\link[MODIS]{whittaker.raster}}: Logical. See
details}

\item{outlierThreshold}{FROM \code{\link[MODIS]{whittaker.raster}}: Numerical
in the same unit as vi, used for outliers removal. See details}

\item{mergeDoyFun}{FROM \code{\link[MODIS]{whittaker.raster}}: Especially
when using argument collapse=TRUE, multiple measurements for one day can be
present, here you can choose how those values are merged to one single
value: "max" use the highest value, "mean" or "weighted.mean" use the mean
if no weighting "w" is available and weighted.mean if it is.}

\item{...}{nFROM \code{\link[MODIS]{whittaker.raster}}: Arguments passed to
?writeRaster (except filename is automatic), NAflag, datatype,
overwrite,...}
}
\value{
raster brick of smoothed images
}
\description{
\code{SmoothStack} takes a raster stack of RS images and outputs a smoothed
raster brick over the same time period.
}
\details{
\code{SmoothStack} converts a raster stack of RS images (as processed through
\code{\link{ConvertRS2Stack}}) and calls the MODIS-R
\code{\link[MODIS]{whittaker.raster}} smoothing function to generate a
smoothed raster brick over the same time series as the input. All function
parameters are per the \code{\link[MODIS]{whittaker.raster}} function except
we force the timeInfo to be derived from the input raster stack so the names
match and the smoothed brick can be used in other tools. The
\code{\link[MODIS]{whittaker.raster}} tool also exports a set of smoothed
TIFs, so also specify an output file directory.
}
\examples{
## Take the raster stack of LAI images created through ConvertRS2Stack and
## apply a smoothing filter that also removes outliers, which we specify to
## be more than 0.5 LAI from the smoothed value.
\dontrun{
lai.b.sm <- SmoothStack(lai.b,
         outDirPath="/Volumes/d1/adugger/RS/MODIS_ARC/PROCESSED/FRNTRNG_LAI_SMOOTHED",
         outputAs="one", removeOutlier=TRUE, outlierThreshold=0.5, lambda=1000,
         overwrite=TRUE)
}
}
\seealso{
Other MODIS: \code{\link{CalcStatsRS}};
  \code{\link{ConvertRS2Stack}};
  \code{\link{ConvertStack2NC}}; \code{\link{GapFillRS}};
  \code{\link{GetMODIS}}; \code{\link{InsertRS}}
}
\concept{
MODIS dataAnalysis
}
\keyword{smooth}

