% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snotel_get.R
\docType{data}
\name{snotelMeta}
\alias{snotelMeta}
\title{SNOTEL site metadata}
\format{data.frame:  854 obs. of  11 variables:
\describe{
 \item{ntwk}{Observation network}
 \item{state}{State where site is located}
 \item{site_name}{Site name}
 \item{site_id}{Site integer ID}
 \item{start}{Start date for the observation record at the site}
 \item{lat}{Site latitude}
 \item{lon}{Site longitude}
 \item{elev}{Site elevation in feet}
 \item{county}{County where site is located}
 \item{huc}{HUC name where site is locted}
 \item{huc_id}{HUC ID where site is located}
}}
\usage{
snotelMeta
}
\description{
Metadata for the SNOTEL sites (as of June 2015) including lat/lon.
}
\seealso{
Other SNOTEL: \code{\link{GetSnotel}}
}
\concept{
data
}
\keyword{data}

