% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{StdLon}
\alias{StdLon}
\title{Standardize lon to (-180,180].}
\usage{
StdLon(x)
}
\arguments{
\item{x}{The numeric objeect to be standardized.}
}
\value{
The standardized object.
}
\description{
\code{StdLon} Standardizes longitude to (-180,180]
}
\examples{
StdLon(0:360)
}

