% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_real_time.R
\name{stnTsDiff}
\alias{stnTsDiff}
\title{Get variables from timeslice files.}
\usage{
stnTsDiff(df)
}
\description{
Get variables from timeslice files.
}
\examples{
doMC::registerDoMC(8)
si <- GetTimeSliceVars( c('stationId'), lastNDays=21)
asi<-unique(unlist(plyr::llply(si, function(tt) tt$vars$stationId)))
uniqueStIds <- asi
save(uniqueStIds, file='~/realTimeUniqueStns.RData')
#want to know if data are changing more than hourly
qq<-GetTimeSliceVars(c('discharge', 'stationId', 'time'), lastNDays=1, pattern='2015-06-06')
qDf <- plyr::ldply(qq, function(ll) {df <- as.data.frame(ll$vars[c("discharge","stationId")]) ;
                                    df$time<-ll$vars$time; df }, .parallel=TRUE )
qDf2 <- plyr::ddply(qDf, plyr::.(stationId), stnTsDiff)
obsTab <- table(qDf2$tDiff)
names(obsTab) <- as.numeric(names(obsTab))/60
cumsum(obsTab)/sum(obsTab)
minDiff <- plyr::daply(qDf2, plyr::.(stationId), function(dd) min(dd$tDiff))
minDiffTab <- table(minDiff)
names(minDiffTab) <- as.numeric(names(minDiffTab))/60
minDiffTab
cumsum(minDiffTab)/sum(minDiffTab)
}

