% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{[.prettyUsgs}
\alias{[.prettyUsgs}
\title{Extract or replace parts of prettyUsgs objects.}
\usage{
\method{[}{prettyUsgs}(x, i, ...)
}
\arguments{
\item{prettyUsgs}{A dataframe of class \code{c("prettyUsgs", "data.frame")}
returned from \code{PrettyUsgs}}
}
\value{
A dataframe of class c("prettyUsgs", "data.frame") or numeric
}
\description{
\code{`[.prettyUsgs`} extracts or replaces parts prettyUsgs objects and retains appropriate
attributes.
}
\examples{
\dontrun{
## these should be made formal tests
dbPath <- '~/wrfHydroTestCases/usgsDb/'
threeSites <- QuerySiteInfo(path = dbPath, info = c("site_no"))[1:3,]$site_no
obs <- PrettyUsgs(QueryHaveSite(threeSites, path=dbPath, ret='00060', get=TRUE))
length(unique(obs$site_no))
atomicV <- obs[,'value']
str(atomicV)
atomicS <- obs[1,'value']
str(atomicS)
df0 <- obs[,c('value','POSIXct','site_no')]
df1 <- obs[,c('value','POSIXct')]
df2 <- obs[1,c('value','POSIXct','site_no')]
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyUsgs}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{internal}
\keyword{manip}

