% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{subset.prettyUsgs}
\alias{subset.prettyUsgs}
\title{subset prettyUsgs objects.
\code{subset.prettyUsgs} subsets prettyUsgs objects and retains their attributes.}
\usage{
\method{subset}{prettyUsgs}(prettyUsgs, ...)
}
\arguments{
\item{prettyUsgs}{A dataframe of class c("prettyUsgs", "data.frame") returned from PrettySiteData}

\item{...}{additional arguments to subset.data.frame}
}
\value{
A dataframe of class c("prettyUsgs", "data.frame")
# See vignette "Collect USGS stream observations and build a local database" for examples.
}
\description{
subset prettyUsgs objects.
\code{subset.prettyUsgs} subsets prettyUsgs objects and retains their attributes.
}

