% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{SubsetWts}
\alias{SubsetWts}
\title{Subset spatial weight file based on combination of link IDs and i/j coords.}
\usage{
SubsetWts(wts, rlids, istart, iend, jstart, jend)
}
\arguments{
\item{wts}{Weight file list object (output from ReadWtFile)}

\item{rlids}{Vector of link IDs to keep in the subsetted output}

\item{istart}{Start index for the grid in the i (x) direction}

\item{iend}{End index for the grid in the i (x) direction}

\item{jstart}{Start index for the grid in the j (y) direction}

\item{jend}{End index for the grid in the j (y) direction}
}
\value{
List of 2 dataframes
}
\description{
\code{SubsetWts} takes a weight file object (as read in from ReadWtFile) and
subsets the data and poly dataframes based on a user-defined set of link IDs
and user-defined start and end indices. Returns a list of 2 dataframes for
the subsetted data and poly objects.
}
\keyword{utilities}

