<<<<<<< HEAD
<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_zone.R
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> 69bfbb5a27d96178965b8e8d863b6464ef1a29e8
% Please edit documentation in R/get_poly.R
>>>>>>> upstream/devBranch
\docType{data}
\name{timeZone}
\alias{timeZone}
\title{Time zone polygon}
\format{data.frame}
\usage{
timeZone
}
\description{
Polygon shapefile of world time zone classification.
}
\section{Citation}{

Online_Linkage: \link[http://efele.net/maps/tz/world/]{http://efele.net/maps/tz/world/} \cr
}
\concept{
data
}
\keyword{data}

