% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{TransUsgsProdStat}
\alias{TransUsgsProdStat}
\title{Translate USGS product/stat codes to something readable (and vice versa).
\code{TransUsgsProdStat} Translate USGS product/stat codes to something readable (and vice versa)
using a lookup table.}
\usage{
TransUsgsProdStat(names, whichIn = FALSE)
}
\arguments{
\item{names}{Character product_stat codes (e.g. 'X_00060_00011') or their translation
(e.g. 'Discharge (cfs)') separated with an underscore.}

\item{whichIn}{Logical}
}
\value{
if whichIn==FALSE : Character of the translation.
        if whichIn==TRUE  : Integer index which passed names are in the table.
}
\description{
Translate USGS product/stat codes to something readable (and vice versa).
\code{TransUsgsProdStat} Translate USGS product/stat codes to something readable (and vice versa)
using a lookup table.
}

