% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observations.R
\name{TransUsgsTz}
\alias{TransUsgsTz}
\title{Translate USGS timezone codes to the so-called "Olson names" used by POSIXct.
\code{TransUsgsTz} translates the USGS time codes to Olson Names.}
\usage{
TransUsgsTz(usgsTz)
}
\arguments{
\item{usgsTz}{\code{TransUsgsTz}}
}
\description{
Translate USGS timezone codes to the so-called "Olson names" used by POSIXct.
\code{TransUsgsTz} translates the USGS time codes to Olson Names.
}

