% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{UpdateGwbuckFile}
\alias{UpdateGwbuckFile}
\title{Update a groundwater bucket parameter netcdf file based on an R dataframe.}
\usage{
UpdateGwbuckFile(gwbuckFile, gwbuckDf, subDim = TRUE)
}
\arguments{
\item{gwbuckFile}{Path to new groundwater bucket parameter file to update. 
This should be a copy of an existing groundwater parameter file, as it will 
be overwritten}

\item{gwbuckDf}{Dataframe of the new groundwater bucket parameter information 
to write to the new file}

\item{subDim}{Boolean whether the dimensions in the new file need to be subsetted
to match the new dataframe (DEFAULT=TRUE)}
}
\description{
\code{UpdateGwbuckFile} takes a groundwater bucket parameter dataframe
(e.g., as generated by ReadGwbuckFile then subsetted) and exports it to a 
new model-ready groundwater bucket parameter file.
}
\keyword{utilities}

