% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{UpdateLinkFile}
\alias{UpdateLinkFile}
\title{Update a route link netcdf file based on an R dataframe.}
\usage{
UpdateLinkFile(linkFile, linkDf, subDim = TRUE)
}
\arguments{
\item{linkFile}{Path to new route link file to update. This should be a copy
of an existing route link file, as it will be overwritten}

\item{linkDf}{Dataframe of the new route link information to write to the
new file}

\item{subDim}{Boolean whether the dimensions in the new file need to be subsetted
to match the new dataframe (DEFAULT=TRUE)}
}
\description{
\code{UpdateLinkFile} takes a route link dataframe (e.g., as generated
by ReadLinkFile then subsetted) and exports it to a new model-ready
route link file.
}
\keyword{utilities}

