% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmap_utils.R
\name{UpdateWtFile}
\alias{UpdateWtFile}
\title{Update a spatial weight netcdf file based on an R dataframe.}
\usage{
UpdateWtFile(wtFile, wtdataDf, wtpolyDf, subDim = TRUE)
}
\arguments{
\item{wtFile}{Path to new spatial weight file to update. This should be a copy
of an existing spatial weight file, as it will be overwritten}

\item{wtdataDf}{Dataframe of the new grid-based (data) information to write to the
new file}

\item{wtpolyDf}{Dataframe of the new poly-based (poly) information to write to the
new file}

\item{subDim}{Boolean whether the dimensions in the new file need to be subsetted
to match the new dataframe (DEFAULT=TRUE)}
}
\description{
\code{UpdateWtFile} takes a pair of spatial weight dataframes (e.g., data aand poly 
dataframes as generated by ReadWtFile then subsetted using SubsetWts) and exports 
them to a new model-ready spatial weight file.
}
\keyword{utilities}

