% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualize_channtwk.R
\name{VisualizeChanNtwk}
\alias{VisualizeChanNtwk}
\title{Visualize WRF Hydro channel link indices and get coordinates.}
\usage{
VisualizeChanNtwk(file)
}
\arguments{
\item{file}{A path/name to an output YYYMMDDHHmm.CHRTOUT_DOMAIN* file or a hydroDART Posterior_Diag.nc file.}
}
\value{
A function which allows the plot to be interactively queried once each time it is run and returns
the coordinates of the selected location.
}
\description{
\code{VisualizeChanNtwk} shows the channel network indices and provides
interactive selection to retrieve link coordinates.
}
\details{
A fairly crude represenation of the
}
\examples{
filePost <- "/home/jamesmcc/WRF_Hydro/Fourmile_test_case_v2/DART.noah/Posterior_Diag.nc"
post <- VisualizeChanNtwk(filePost)
fileCh <- "/home/jamesmcc/WRF_Hydro/Fourmile_test_case_v2/DART.noah/RUNS.initialEnsembleWFlow.20120426-20120726.Global/ensemble.0001/201207072200.CHRTOUT_DOMAIN1"
ch <- visChanNtwk(fileCh)
ch()
getChanPoint <- visChanNtwk(filePost)
dev.new()
getChanPoint <- visChanNtwk(fileCh)
getChanPoint()
}

