% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualize_domain.R
\name{VisualizeDomain}
\alias{VisualizeDomain}
\title{Visualize WRF Hydro domain files.}
\usage{
VisualizeDomain(file, variable = NULL, plot = TRUE)
}
\arguments{
\item{file}{A path/name to a geo or hydro domain file.}

\item{variable}{An optional variable name of interest within the file.}

\item{plot}{Logical: plot or not?}
}
\value{
A function which can be called to plot the data and allow adjustment of its arguments, the plotting parameters.
}
\description{
\code{VisualizeDomain} creates basic plots of WRF Hydro domains.
}
\details{
Crude plots of the WRF Hydro domain files. The routine accepts a file path/name for either the geo
(coarse resolution) or hydro (fine resolution) files. Spatial variables of interest are listed
if none is provided. The return is a function (a closure which encapuslates the domain data) which
creates a plot when called. The arguments to the function can be changed to tailor the plot
(arguments are passed to ggmap and ggplot inside the function). This function (the closure) returns
a ggplot object whose data can be accessed.
}
\examples{
file<-'~/WRF_Hydro//DOMAIN_library/Fourmile_test_case_AD/hydro_OrodellBasin_100m_8.nc'
ggMapFunction <- VisualizeDomain(file, "CHANNELGRID")
ggMap1 <- ggMapFunction(zoom=11, pointshape=15, pointsize=7,
                        source="google", maptype="terrain")
# Add a streamflow gauge point; compare reality and the model.
orodellLonLat <- data.frame(lon=c(254.6722259521484375, 254.67374999999998408)-360,
                           lat=c(40.019321441650390625, 40.018666670000001773),
                           gauge=c('model','USGS'))
ggMap2 <- ggMapFunction(location=c(lon=orodellLonLat$lon[1], lat=orodellLonLat$lat[1]),
                        zoom=14, pointshape=15, pointsize=7,
                        source="google", maptype="terrain", plot=FALSE)
ggMap2 + geom_point(data=orodellLonLat, aes(x=lon,y=lat, shape=gauge)) +
         scale_x_continuous(limits=rev(orodellLonLat$lon+c( .01,-.01))) +
         scale_y_continuous(limits=rev(orodellLonLat$lat+c( .005,-.005)))
}

