% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize_routelink.R
\name{VisualizeSubsetStream}
\alias{VisualizeSubsetStream}
\title{Visualize upstream or downstream links determined from GatherStream}
\usage{
VisualizeSubsetStream(indDist, ncFile)
}
\arguments{
\item{indDist}{List containing indies and accumulated distance from start, obtained from GatherStream}

\item{ncFile}{Route Link file read in/initially processed with VisualizeRouteLink()}
}
\value{
Map of Route Links with selected upstream/downstream links highlighted in red, starting location in black
}
\description{
Visualize upstream or downstream links determined from GatherStream
}

