% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WhichMulti}
\alias{WhichMulti}
\title{Return indices in native matrix dims}
\usage{
WhichMulti(A)
}
\arguments{
\item{A}{Array of booleans}
}
\value{
a sum(A) x length(dim(A)) array of multi-indices where A == TRUE
}
\description{
\code{WhichMulti} is same as which but returns indices
in native input dimensions. Blatantly copied from
Mark van der Loo 16.09.2011.
}
\keyword{internal}
\keyword{utilities}

