% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_prep_streamflow.R
\name{WriteDischargeObsSeq}
\alias{WriteDischargeObsSeq}
\title{Writes an ascii obs_seq.in file for DART binary create_obs_sequence}
\usage{
WriteDischargeObsSeq(pretty, outPath = ".", stationName, errorId,
  typeQ = 20)
}
\arguments{
\item{pretty}{The prettyUsgs discharge observations to which variances are to be added.}

\item{outPath}{Character The directory where the file is to be written.}

\item{stationName}{Character To help identify the outputfile and the station data in the file.}

\item{errorId}{Character To help identify the outputfile and the error function used.}

\item{typeQ}{Numeric   Stream discharge obs type index in my pre-release version of DART. It could change.
DART/obs_kind/DEFAULT_obs_kind_mod.F90}
}
\description{
\code{WriteDischargeObsSeq} makes variances for prettyUsgs discharge observations. The formulation 
of the variances is subjective. Assuming zero-mean Gaussian observation errors, the approach 
here is to supply a function which estimates the 3-sigma (inner 99.5% error quantiles) around 
the observations. This amount seems somewhat easier to conceptualize than 1-sigma, hence here
we are. This function divides the error amounts by 3 and either returns the standard deviation
or squares the result to return the variance (default).
}
\examples{
#Following on examples for MkDischargeVariance
\dontrun{
WriteDischargeObsSeq(prettyOro, '~/.', 'orodell', 'climTaperDefault')
}
}
\concept{
DART dataMgmt
}
\keyword{manip}

