% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_to_ncdf.R
\name{WriteNcPrettyUsgs}
\alias{WriteNcPrettyUsgs}
\title{Write a pretty usgs format object to file
For now there variable length (ragged) arrays are not supported in the ncdf4 package. After  
emailing Pierce, who welcomded the addition, I sketched how this might be done and we may 
do it.
For now, we will have to create one file per station*product. If returning the file name
as success, that's a plyr::daply. 
keyrwords}
\usage{
WriteNcPrettyUsgs(prettyDf, outPath = ".")
}
\description{
Write a pretty usgs format object to file
For now there variable length (ragged) arrays are not supported in the ncdf4 package. After  
emailing Pierce, who welcomded the addition, I sketched how this might be done and we may 
do it.
For now, we will have to create one file per station*product. If returning the file name
as success, that's a plyr::daply. 
keyrwords
}

