% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_aggregations.R
\name{CalcDailyGhcn}
\alias{CalcDailyGhcn}
\title{To aggregate model output hourly precipitation data into daily values
comparable to GHCN daily data.}
\usage{
CalcDailyGhcn(sg, prcp, reportTime = 700, parallel = FALSE)
}
\arguments{
\item{sg}{A dataframe of the selectedGauges, it should have at least one column
with name of siteIds.
siteIds should match the the standardized GHCN IDs (for example : ACW00011604).
See \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}
for a list of siteIds.}

\item{prcp}{A dataframe of hourly data. It should have at least three columns:
POSIXct, statArg, and DEL_ACCPRCP. POSIXct has the time in UTC, statArg has the
site ID, and DEL_ACCPRCP is the depth of rainfall for that hour. The DEL_ACCPRCP
column will be with the same name if you use GetMultiNcdf, ReshapeMultiNcdf and
CalcNoahmpFluxes to get the precipitation from WRF-Hydro output files.}

\item{reportTime}{A vector of reportTime matching with sg$siteIds.
reportTime is a unique reporting time for each elelment of each gauge.
It may vary for different elemenets for a single gauge.
The report time values are retrived when reading the daily GHCN data
using  \code{GetGhcn2}. Note, use 700 wherever the report time is missing.
It will be overwritten if sg$reportTime exits.}

\item{parallel}{Logical (DEFAULT=FALSE)}
}
\value{
Daily precipitation comparable with daily GHCN data.
}
\description{
This function inputs the model output hourly data and aggregate the
hourly data into daily info. GHCN gauges report data usually at 7:00 AM,
however, that is not the case for all gauges. The reporting time may
even differ for the same gauge but different elements. One can obtain the
report time using \code{GetGhcn2}.
}
\examples{
\dontrun{
#ADD EXAMPLE HERE
}
}
\seealso{
Other GHCN: \code{\link{GetGhcn2}}; \code{\link{GetGhcn}};
  \code{\link{SelectGhcnGauges}}
}
\concept{
GHCN
}
\keyword{manip}

