% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{CalcFdcPerf}
\alias{CalcFdcPerf}
\title{Computes flow duration curve statistics for WRF-Hydro streamflow output}
\usage{
CalcFdcPerf(strDf.mod, strDf.obs, strCol.mod = "q_cms",
  strCol.obs = "q_cms", stdate = NULL, enddate = NULL)
}
\arguments{
\item{strDf.mod}{The forecast point output dataframe (required). Assumes only
one forecast point per file, so if you have multiple forecast points in
your output dataframe, use subset to isolate a single forecast point's
data. Also assumes model output and observation both contain POSIXct fields
(called "POSIXct").}

\item{strDf.obs}{The observed streamflow dataframe. Assumes only one gage per
file, so if you have multiple gages in your dataframe, use subset to
isolate a single gage's data. Also assumes model output and observation
both contain POSIXct fields (called "POSIXct").}

\item{strCol.mod}{The column name for the streamflow time series for the
MODEL data (default="q_cms")}

\item{strCol.obs}{The column name for the streamflow time series for the
OBSERVED data (default="q_cms")}

\item{stdate}{Start date for statistics (DEFAULT=NULL, all records will be
used). Date MUST be specified in POSIXct format with appropriate timezone 
(e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d \%H:\%M:\%S",
tz="UTC"))}

\item{enddate}{End date for statistics (DEFAULT=NULL, all records will be
used). Date MUST be specified in POSIXct format with appropriate timezone 
(e.g., as.POSIXct("2013-05-01 00:00:00", format="\%Y-\%m-\%d \%H:\%M:\%S",
tz="UTC"))}
}
\value{
A new dataframe containing the flow duration curve statistics.
}
\description{
\code{CalcFdcPerf} calculates flow duration curve statistics for streamflow
output.
}
\details{
\code{CalcFdcPerf} reads a model forecast point streamflow timeseries (i.e.,
created using \code{\link{ReadFrxstPts}}) and a streamflow observation
timeseries (i.e., created using \code{\link{ReadUsgsGage}}) and calculates
flow duration curve statistics at various exceedance thresholds (e.g., 10\%,
20\%, etc.). The tool will subset data to matching time periods (e.g., if the
observed data is at 5-min increments and modelled data is at 1-hr increments,
the tool will subset the observed data to select only observations on the
matching hour break).

Flow Duration Curve Statistics: \cr (mod = model output, obs = observations) 
\itemize{ \item p.exceed: exceedance threshold (e.g., 0.2 means a flow value
that is exceeded 20\% of the time) \item q.mod: MODEL flow value at specified
exceedance threshold (in native flow units) \item q.obs: OBSERVED flow value
at specified exceedance threshold (in native flow units) \item q.err:
difference between model and observed flow values [mod-obs] (in native flow
units) \item q.perr: percent error in model flow [(mod-obs)/obs] }
}
\examples{
## Take forecast point model output for Fourmile Creek (modStrh.mod1.fc) 
## and a corresponding USGS gage observation file (obsStrh.fc), both at an 
## hourly time step, and calculate flow duration curve statistics. The 
## model forecast point data was imported using ReadFrxstPts and the gage
## observation data was imported using ReadUsgsGage.

\dontrun{
CalcFdcPerf(modStr1h.allrt.fc, obsStr5min.fc)

Output:
 p.exceed    q.mod   q.obs
 0.1         3.07    5.25
 0.2         1.35    2.31
 0.3         0.82    1.06
 0.4         0.48    0.65
 0.5         0.29    0.45
 0.6         0.18    0.34
 0.7         0.14    0.25
 0.8         0.11    0.19
 0.9         0.08    0.16
}
}
\concept{
modelEval
}
\keyword{ts}
\keyword{univar}
