% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcMeanMinrm}
\alias{CalcMeanMinrm}
\title{Calculate mean with enforced minimum valid value.}
\usage{
CalcMeanMinrm(x, minValid = -1e+30)
}
\arguments{
\item{x}{The vector of values.}

\item{minValid}{The minimum valid value.}
}
\value{
The mean.
}
\description{
\code{CalcMeanMinrm} calculates a mean with an enforced minimum valid value.
Read a vector and calculate the mean with all values below
a specified minimum value set to NA (and therefore ignored).
}
\examples{
x <- c(1,2,-1e+20,3)
mean(x) # yields -2.5e+19
CalcMeanMinrm(x, 0) # yields 2
}
\keyword{internal}
