% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_MET_measures.R
\name{CalcMetCont}
\alias{CalcMetCont}
\title{Calculate the MET verification measures for continuous variables.}
\usage{
CalcMetCont(obs, mod, conRange = NULL)
}
\arguments{
\item{obs}{A numeric vector of observation time series.}

\item{mod}{A numeric vector of model simulation or forecast time series.}

\item{conRange}{A numeric vector containing two elements.
Values are used as the lower and upper boundary for conditional statistics.
If conditioning only at one tail, leave the second value as NA.}
}
\value{
A matrix containing most of the MET verification measures
for continuous variable. It includes \itemize{
\item NO: number of datapoint
\item minObs: min of observation time series}
\item minMod: min of model simulation/forecast time series
\item maxObs: max of observation time series
\item maxMod: Max of model simulation/forecast time series
\item meanObs: Mean of observation time series
\item meanMod: Mean of model simulation/forecast time series
\item stdObs: Standard deviation of observation time series
\item stdMod: Standard deviation of
\item pearsonCor: Peason correlation coefficient
\item SpearmanCor: Spearman correlation coefficient
\item KendallCor: Kendall tau correlation coefficient (tau-b)
\item ME:Mean Error
\item MultiBias: Mutliplicative bias
\item MSE: Mean Square Error
\item RMSE: Root Mean Square Error
\item MAE: Mean Absolute Error
\item MAD: Median Absolute Deviation
\item IQR: Inter Quantile Range of the Errors
\item E10: 10 percen of Errors
\item E25: 25 percent of Errors
\item E50  50 percent of Errors
\item E75: 75 percent of Errors
\item E90: 90 percent of Errors
}
\description{
It inputs two times series of observation and model/forecast and
computes the MET verification measures for continous variables.
Pair matching is not done in this function. It assumes the two time series are
comparable one by obe.
}
\examples{
DF <- data.frame(obs=seq(1,5), mod=seq(6,10), stringsAsFactors=FALSE)
stat<-CalcMetCont(DF$obs,DF$mod)
}

