% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MET_measures.R
\name{CalcMetCont}
\alias{CalcMetCont}
\title{Calculate MET verification measures for continuous variables.}
\usage{
CalcMetCont(obs, mod, conRange = Inf)
}
\arguments{
\item{obs}{Numeric vector of observation time series.}

\item{mod}{Numeric vector of model simulation or forecast time series.}

\item{conRange}{Numeric vector containing two elements. 
Values are used as the lower and upper boundary for conditional statistics. 
If conditioning only at one tail, leave the second value as Inf.}
}
\value{
A matrix containing most of the MET verification measures
for continuous variable.
}
\description{
\code{CalcMetCont} inputs two times series of observation and model/forecast and 
computes MET verification measures for continous variables.
Pair matching is not done in this function. It assumes the two time series are 
comparable one by one.
}
\details{
The calculated statistics are the following: \itemize{
\item numPaired: number of paired datapoint
\item minObs: min of observation time series
\item minMod: min of model simulation/forecast time series
\item maxObs: max of observation time series
\item maxMod: max of model simulation/forecast time series
\item meanObs: mean of observation time series
\item meanMod: mean of model simulation/forecast time series
\item stdObs: standard deviation of observation time series
\item stdMod: standard deviation of model simulation/forecast time series
\item pearsonCor: Pearson correlation coefficient
\item spearmanCor: spearman correlation coefficient
\item kendallCor: kendall tau correlation coefficient (tau-b)
\item ME:mean error
\item multiBias: mutliplicative bias
\item MSE: mean square error
\item RMSE: root mean square error
\item MAE: mean absolute error
\item MAD: median absolute deviation
\item IQR: inter quantile range of the errors
\item E10: 10 percent of errors 
\item E25: 25 percent of errors 
\item E50:  50 percent of errors 
\item E75: 75 percent of errors 
\item E90: 90 percent of errors
} For more information refer to Model Evaluation Tool (MET) documentation
}
\examples{
DF <- data.frame(obs=seq(1,5), mod=seq(6,10), stringsAsFactors=FALSE)
stat<-CalcMetCont(DF$obs,DF$mod)
}
\seealso{
Other modelEvaluation: \code{\link{CalcMetContGrid}},
  \code{\link{CalcModPerfMulti}},
  \code{\link{CalcModPerf}},
  \code{\link{CalcNoahmpFluxes}},
  \code{\link{CalcNoahmpWatBudg}}
}
\concept{
modelEval
}
\keyword{ts}
\keyword{univar}

