% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{CalcNoahmpFluxes}
\alias{CalcNoahmpFluxes}
\title{Calculate water fluxes from NoahMP output}
\usage{
CalcNoahmpFluxes(ldasoutDf, idCol = NULL)
}
\arguments{
\item{ldasoutDf}{The LDASOUT dataframe}

\item{idCol}{The optional ID column to use to parse the output;
for example, a basin ID column or other index ID. (DEFAULT=NULL)}
}
\value{
The input dataframe with new water flux columns added.
}
\description{
\code{CalcNoahmpFluxes} calculates water balance fluxes from accumulated
water terms.
}
\details{
Read a dataframe derived from NoahMP LDASOUT output (i.e., using
\code{\link{GetMultiNcdf}}) and calculate water budget component fluxes from
accumulated water variables. NOTE: Currently only works for runs using NoahMP
as the LSM.
}
\examples{
## Take a NoahMP LDASOUT dataframe for a model run of Fourmile Creek
## and generate a dataframe with water fluxes added.
\dontrun{
modLDASOUT1d.nort.fc <- CalcNoahmpFluxes(modLDASOUT1d.nort.fc)
}
}
\seealso{
Other modelEvaluation: \code{\link{CalcModPerfMulti}},
  \code{\link{CalcModPerf}},
  \code{\link{CalcNoahmpWatBudg}},
  \code{\link{CalcStatCategorical}},
  \code{\link{CalcStatCont}}
}
\concept{
dataMgmt
}
\keyword{manip}

