% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snodas.R
\name{CalcSnodasCoords}
\alias{CalcSnodasCoords}
\title{Calculate the SNODAS coordinates.}
\usage{
CalcSnodasCoords()
}
\value{
A list of lon and lat
}
\description{
\code{CalcSnodasCoords} Calculate the SNODAS coordinates.
}
\examples{
snodasCoords <- CalcSnodasCoords()
}
\seealso{
Other SNODAS: \code{\link{GetSnodasDepthSweDate}};
  \code{\link{GetSnodasPointTs}};
  \code{\link{PutSnodasCoordsNcdf}};
  \code{\link{PutSnodasNcdf}};
  \code{\link{ReadSnodasDepthSweDate}}
}
\concept{
SNODAS
}
\keyword{manip}

