% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_RS.R
\name{CalcStatsRS}
\alias{CalcStatsRS}
\title{Calculates summary statistics from a remote sensing time series}
\usage{
CalcStatsRS(inBrick)
}
\arguments{
\item{inBrick}{The name of the raster brick to calculate statistics on.}
}
\value{
A dataframe of statistics by time period (date).
}
\description{
\code{CalcStatsRS} takes a raster brick of RS images and generates a time series
dataframe of summary statistics across the domain for each time step.
}
\details{
\code{CalcStatsRS} takes a raster brick of remote sensing images over a time period
(as created by \code{\link{ConvertRS2Brick}} or \code{\link{SmoothBrick}})
and generates a dataframe object that summarizes all cells in the RS image at each
time step. The statistics calculated are mean, min, max, and standard deviation.
This tool is useful for evaluating how a smoothing function is impacting the
time series of images.
}
\examples{
## Calculate domain statistics and plot the mean.

stats.lai.b <- CalcStatsRS(lai.b)
with(stats.lai.b, plot(POSIXlt, mean, typ='l'))
}

