% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression_gridded.R
\name{CheckReGridConnDown}
\alias{CheckReGridConnDown}
\title{Check for re-expressed grid connectivity downstream.}
\usage{
CheckReGridConnDown(ind, printInds = FALSE)
}
\description{
Check for re-expressed grid connectivity downstream.
}

