% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_RS.R
\name{ConvertStack2NC}
\alias{ConvertStack2NC}
\title{Convert a raster stack to a NetCDF file.}
\usage{
ConvertStack2NC(inStack, outFile = NULL, varName = NULL, varUnit = NULL,
  varLong = NULL, varNA = -1e+36)
}
\arguments{
\item{inStack}{The name of the raster stack to export.}

\item{outFile}{Name for an output NetCDF file.}

\item{varName}{Name for the NetCDF export variable.}

\item{varUnit}{Units for the NetCDF export variable.}

\item{varLong}{Long name for the NetCDF export variable.}

\item{varNA}{Value to set for "NA" or "no data". Default is -1.e+36.}
}
\description{
\code{ConvertStack2NC} takes a raster stack of RS images and outputs a NetCDF file
with a time dimension and specified variable name & metadata.
}
\details{
\code{ConvertStack2NC} converts a raster stack to an output NetCDF file. The raster stack
should be already processed through the \code{\link{ConvertRS2Stack}} tool or follow the
same layer (date) naming convention.
}
\examples{
## Export the raster stack of LAI images created through ConvertRS2Stack to a NetCDF file. Use the full time series of images.

ConvertStack2NC(lai.b, outFile="BCNED_LAI.nc", varName="LAI", varUnit="(m^2)/(m^2)", varLong="Leaf area index")
}
\concept{
aconcept
}
\keyword{IO}

