% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{ExportGeogrid}
\alias{ExportGeogrid}
\title{Creates a georeferenced TIF from a geogrid variable}
\usage{
ExportGeogrid(inFile, inVar, outFile, inCoordFile = NA, inLyr = NA,
  inLyrPos = 2, inProj4 = NA)
}
\arguments{
\item{inFile}{A netcdf file containing the variable to be converted to a TIF.}

\item{inVar}{The name of the variable to be converted, should exist in inFile.}

\item{outFile}{The geoTiff filename to create.}

\item{inCoordFile}{(OPTIONAL) The netcdf file containing the lat/lon
coordinates if they are not included in the inFile. This is useful, for
example, when creating a geotiff from an LDASOUT output file which does not
contain lat/lon coordinates but matches the spatial coordinate system of
the geogrid input file.}

\item{inLyr}{(OPTIONAL, default=NA) The layer number to export if the variable 
has more than 2 dimensions, e.g., for soil or snow layer variables.}

\item{inLyrPos}{(OPTIONAL, default=2) The position of the dimension to export
if the variable has more than 2 dimensions, e.g., for soil or snow layer variables.}

\item{inProj4}{(OPTIONAL, default=NA) If your data does not follow default WRF 
lambert conformal conic specifications, specify the proj4 string.}
}
\description{
\code{ExportGeogrid} takes a NetCDF geogrid and converts the specified
variable into a georeferenced TIF file.
}
\details{
\code{ExportGeogrid} takes a standard geogrid in NetCDF format and
converts the specified variable to a georeferenced geoTiff for use
in standard GIS tools.
}
\examples{
## Export the HGT_M field from the geogrid file geo_em.d01_1km.nc
## to a geoTiff called geogrid_hgt.tif.

\dontrun{
ExportGeogrid("~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/geo_em.d01_1km_nlcd11.nc",
             "HGT_M", "geogrid_hgt.tif")
ExportGeogrid("~/wrfHydroTestCases/Fourmile_Creek/RUN.RTTESTS/OUTPUT_ALLRT_DAILY/2013031500.LDASOUT_DOMAIN1",
             inVar="SOIL_M",
             outFile="20130315_soilm3.tif",
             inCoordFile="~/wrfHydroTestCases/Fourmile_Creek/DOMAIN/geo_em.d01_1km_nlcd11.nc",
             inLyr=3)
}
}
\seealso{
Other geospatial: \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}},
  \code{\link{GetPoly}}, \code{\link{GetProj}},
  \code{\link{GetRfc}}, \code{\link{GetTimeZone}},
  \code{\link{PolyToRaster}}
}
\concept{dataMgmt geospatial}
\concept{geospatial}
\keyword{IO}
