% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_RS.R
\name{ExportGeogrid}
\alias{ExportGeogrid}
\title{Creates a georeferenced TIF from a geogrid variable}
\usage{
ExportGeogrid(inFile, inVar, outFile)
}
\arguments{
\item{inFile}{The geogrid NetCDF file.}

\item{inVar}{The name of the variable to be converted.}

\item{outFile}{The geoTiff filename to create.}
}
\description{
\code{ExportGeogrid} takes a NetCDF geogrid and converts the specified
variable into a georeferenced TIF file.
}
\details{
\code{ExportGeogrid} takes a standard geogrid in NetCDF format and
converts the specified variable to a georeferenced geoTiff for use
in standard GIS tools.
}
\examples{
## Export the HGT_M field from the geogrid file geo_em.d01_1km.nc
## to a geoTiff called geogrid_hgt.tif.

ExportGeogrid("geo_em.d01_1km.nc", "HGT_M", "geogrid_hgt.tif")
}
\concept{
dataMgmt
}
\keyword{IO}

