% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grib_mod.R
\name{extractGRIBGrid}
\alias{extractGRIBGrid}
\title{Extract grid of GRIB data given a variable name and level.}
\usage{
extractGRIBGrid(fileIn, var, levType, level, nx, ny)
}
\arguments{
\item{fileIn}{GRIB file to be opened for reading.}

\item{var}{GRIB variable to be extracted.}

\item{levType}{Level type to extract GRIB data. Examples of
levType are "surface", "heightAboveGround","cloudTop"}

\item{level}{GRIB level to be extracted.}

\item{nx}{Integer number of GRIB data columns.}

\item{ny}{Integer number of GRIB data rows.}
}
\value{
gridOut Float matrix of GRIB data
}
\description{
\code{extractGRIBGrid} Extract grid of GRIB data.
}
\examples{
gridOut <- extractGRIBGrid('rap.t01z.awp130bgrbf16.grib2','acpcp',
           'surface',0,451,337)
}
\concept{
GRIB
}
\keyword{GRIB}
\keyword{IO}

