% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{FindUsgsStns}
\alias{FindUsgsStns}
\title{Discover USGS stations using huc8 code or lat/lon/radius.}
\usage{
FindUsgsStns(stnLon = NULL, stnLat = NULL, within = NULL, huc8 = NULL,
  siteType = "ST", hasDataTypeCd = "iv")
}
\arguments{
\item{stnLon}{optional}

\item{stnLat}{optional}

\item{within}{optional, goes with stnLon and stnLat and specifies a search
radius in decimal degrees.}

\item{huc8}{optional EIGHT digit HUC code.}

\item{siteType}{the type of USGS site to look for.}

\item{hasDataTypeCd}{the kind of data of interest (iv=instantaneous,
dv=daily, etc.)}
}
\description{
\code{FindUsgsStns} wraps \code{dataRetrieval::whatNWISsites} with common
options for finding USGS gauges. See the dataRetrieval package for more
deails on what else could be passed or how this could be modified. One
improvement here would be to order results by proximity to supplied lat/lon
and filter to nClosest.
}
\examples{
stnDf <- FindUsgsStns(huc8='10190005')
stnDf <- FindUsgsStns(huc8=c('10190005','03160203'))
stnDf <- FindUsgsStns(stnLon=254.67374999999998408,stnLat=40.018666670000001773,within=.001)
stnDf <- FindUsgsStns(stnLon=c(254.67374999999998408,-87.747224700000004),
                      stnLat=c(40.018666670000001773, 31.864042489999999),within=.001)
}
\seealso{
Other streamObs: \code{\link{GetSiteHuc}};
  \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyUsgs}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{[.prettyUsgs}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataGet usgsStreamObs
}
\keyword{IO}

