% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flipLR}
\alias{flipLR}
\title{Flip a matrix from left to right.}
\usage{
flipLR(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
\code{flipLR} Flips a matrix from left to to right.
}
\examples{
x <- matrix[1:9,3]
x
flipLR(x)
}
\keyword{internal}

