% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{flipUD}
\alias{flipUD}
\title{Flip a matrix upside down.}
\usage{
flipUD(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
\code{flipUD} Flips a matrix upside down.
}
\examples{
x <- matrix[1:9, 3]
x
flipUD(x)
}
\keyword{internal}

