% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{FlipUD}
\alias{FlipUD}
\title{Flip a matrix upside down.}
\usage{
FlipUD(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
\code{FlipUD} Flips a matrix upside down.
}
\examples{
x <- matrix(1:9, 3)
x
FlipUD(x)
}
\keyword{internal}

