% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gages_2.R
\docType{data}
\name{gages2Attr}
\alias{gages2Attr}
\title{Gages-II attributes}
\format{data.frame:  9322 obs. of  16 variables:
\describe{
 \item{STAID}{Station ID from USGS NWIS}
 \item{STANAME}{Station Name from USGS NWIS}
 \item{CLASS}{Classification (Ref or Non-ref)}
 \item{AGGECOREGI}{Aggregated ecoregion}
 \item{DRAIN_SQKM}{Drainage area, sq km}
 \item{HUC02}{Hydrologic Unit Code, 2-digit}
 \item{HUC08}{ONLY in gages2AttrPlus: Hydrologic Unit Code, 8-digit}
 \item{LAT_GAGE}{Latitude, decimal degrees}
 \item{LNG_GAGE}{Longitude, decimal degrees}
 \item{STATE}{State at gage location}
 \item{HCDN_2009}{If gage is part of HCDN-2009}
 \item{ACTIVE09}{If gage active in water year 2009}
 \item{FLYRS1900}{Number of complete years of flow data between 1900 and 2009}
 \item{FLYRS1950}{Number of complete years of flow data between 1950 and 2009}
 \item{FLYRS1990}{Number of complete years of flow data between 1990 and 2009}
}}
\usage{
gages2Attr
}
\description{
These are the attributes taken from the USGS Gages-II data set. \cr
\link[http://water.usgs.gov/GIS/metadata/usgswrd/XML/gagesII_Sept2011.xml]{http://water.usgs.gov/GIS/metadata/usgswrd/XML/gagesII_Sept2011.xml} \cr
The original gages-II atributes are in \code{gages2Attr} additional attributes may be placed in \code{gages2AttrPlus}.
}
\section{Citation}{

Originator: James Falcone \cr
Publication_Date: 20111012 \cr
Title: GAGES-II: Geospatial Attributes of Gages for Evaluating Streamflow \cr
Geospatial_Data_Presentation_Form: vector digital data
Publication_Information: \cr
Publication_Place: Reston, Virginia \cr
Publisher: U.S. Geological Survey \cr
Online_Linkage: \link[http://water.usgs.gov/lookup/getspatial?gagesII_Sept2011]{http://water.usgs.gov/lookup/getspatial?gagesII_Sept2011} \cr
}
\seealso{
Other gages2: \code{\link{gages2AttrPlus}}
}
\concept{
data
}
\keyword{data}

