% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_expression.R
\name{GatherStreamInds}
\alias{GatherStreamInds}
\title{Gather upstream or downstream distance from a given starting location}
\usage{
GatherStreamInds(stream, start, linkLengths)
}
\description{
A non-recursive function (recusive version runs in to stack overflow problems for large
 domains. Both are available internally).
}
\details{
@param stream List of stream information containing either from/to and start and end positions, returned from ReExpNetwork.
 @param start Indexed location (NOT comID) of where stream starts
 @param length Vector of link lengths for each re-indexed reach, contained in reExp.nc.
 @param indDist Optional list of indices and distance to that index. Typically not used externally to the recursion.
 
 @return List containing indices and accumulated distance from start
 @examples 
 \dontrun{
     PlotRouteLink <- 
         VisualizeRouteLink(file='~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.nc')
     PlotRouteLink()
     PlotRouteLink(comId=TRUE)
     PlotRouteLink(indices=TRUE)
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reInd.Rdb')
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reExpFrom.Rdb')
     upstreamInds <- GatherStreamInds(from, 379, linkLengths=reInd$length)
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reExpTo.Rdb')
     downstreamInds <- GatherStreamInds(to, 91, length=reInd$length)
 }
}
\concept{
nudging dataMgmt
}
\keyword{manip}

