% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{GatherStreamInds}
\alias{GatherStreamInds}
\title{\code{GatherStreamInds} gathers upstream or downstream indices and distances from a given starting location.}
\usage{
GatherStreamInds(stream, start, linkLengths)
}
\arguments{
\item{stream}{List of stream information containing either from/to and start and end positions, returned from ReExpNetwork.}

\item{start}{Indexed location (NOT comID) of where stream starts}

\item{linkLengths}{Vector of link lengths for each re-indexed reach, contained in reExp.nc.}
}
\value{
List containing indices and accumulated distance from start
}
\description{
A non-recursive function (recusive version runs in to stack overflow problems for large domains. Both are available internally).
}
\examples{
\dontrun{
     PlotRouteLink <-
         VisualizeRouteLink(file='~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.nc')
     PlotRouteLink()
     PlotRouteLink(comId=TRUE)
     PlotRouteLink(indices=TRUE)
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reInd.Rdb')
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reExpFrom.Rdb')
     upstreamInds <- GatherStreamInds(from, 379, linkLengths=reInd$length)
     load('~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.reExpTo.Rdb')
     downstreamInds <- GatherStreamInds(to, 91, length=reInd$length)
}
}
\concept{
nudging dataMgmt
}
\keyword{manip}

