% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_domain.R
\name{GetDomainCoordsProj}
\alias{GetDomainCoordsProj}
\title{\code{GetDomainCoordsProj} collects and reprojects spatial data from geo and hydro files.}
\usage{
GetDomainCoordsProj(file, proj4Str = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{file}{A path/name to a geo or hydro domain file.}

\item{proj4Str}{the desired output projection.}
}
\value{
An sp::SpatialPoints object.
}
\description{
Handle the projection and datum iformation associated with WRF-Hydro geo and hydro grid files.
}
\examples{
## See the vignette "WRF Hydro Domain and Channel Visualization", for details. 
## set your test case path
\dontrun{
hydroFile<- '~/WRF_Hydro/TESTING/TEST_FILES/FRNG/2015-12-04_20:08:06.b8e1e01c4cc2/STD/DOMAIN/Fulldom_hires_netcdf_file_2015_11_20.nc'
hydroCoords <- GetDomainCoordsProj(hydroFile)
str(hydroCoords)
hydroCoords <- as.data.frame(hydroCoords)
Calculate/Plot the error of the longitude and latitidue coords in the Fulldom file
hydroCoords$latFile <- as.vector(ncdump(hydroFile,'LATITUDE', q=TRUE))
hydroCoords$lonFile <- as.vector(ncdump(hydroFile,'LONGITUDE', q=TRUE))
hydroCoords$latErr <- lat-hydroCoords$lat
hydroCoords$lonErr <- lon-hydroCoords$long
summary(hydroCoords$latErr)
summary(hydroCoords$lonErr)
theMap <- ggmap::get_map(location=c(lon=mean(hydroCoords$long)-.5,
                                    lat=mean(hydroCoords$lat))+.2, zoom = 13)
ggmap::ggmap(theMap) + geom_point(data=hydroCoords, aes(x=lonFile, y=latFile),size=1)
ggmap::ggmap(theMap) + geom_point(data=hydroCoords, aes(x=long, y=lat),size=.5)
## test for unique pairs in lat/long
testLon <- hydroCoords$long
testLat <- hydroCoords$lat
tally=list()
for(i in 1:300) {
  matches <- (testLon[(i+1):length(testLat)] == testLon[i] &
              testLat[(i+1):length(testLat)] == testLat[i]   )
  whMatch <- which(matches)+i  
  if(length(whMatch)) {
    for(j in 1:length(whMatch)) tally[[length(tally)+1]] <- c(i,whMatch)
    print(i)
    print(length(tally))
  }
}
} #dontrun
}
\seealso{
Other domain: \code{\link{VisualizeChanNtwk}},
  \code{\link{VisualizeDomain}},
  \code{\link{VisualizeSpatial}}
}
\concept{
plot
}
\keyword{hplot}

