% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_multi_ncdf.R
\name{GetFileStat}
\alias{GetFileStat}
\title{Return data from a netcdf file.}
\usage{
GetFileStat(theFile, variable, index, ...)
}
\arguments{
\item{theFile}{The file to open.}

\item{variable}{TODO}

\item{index}{TODO}

\item{...}{Further arguments to be passsed to a statistic.}
}
\value{
A dataframe with columns TODO
}
\description{
\code{GetFileStat} opens a netcdf file, extracts specified indices for
a variable, and may apply a specified statistic.
}
\examples{
This is to do.
}

