% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geospatial.R
\name{GetGeogridSpatialInfo}
\alias{GetGeogridSpatialInfo}
\title{Pull necessary geospatial information from geogrid file used for
regridding and deprojection.}
\usage{
GetGeogridSpatialInfo(geoFile)
}
\arguments{
\item{geoFile}{The geogrid file information is being pulled from}
}
\value{
spatialDF data frame containing geospatial information
}
\description{
\code{GetGeogridSpatialInfo} Pull geospatial information about WRF-Hydro
 modeling domain from geogrid file.
}
\examples{
\dontrun{
lccGeoDF <- GetGeogridSpatialInfo('./geo_em.d02.nc')
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}}; \code{\link{GetPoly}};
  \code{\link{GetProj}}; \code{\link{GetRfc}};
  \code{\link{GetTimeZone}}; \code{\link{PolyToRaster}}
}
\concept{
geospatial getData
}
\keyword{IO}

