% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn_get.R
\name{GetGhcn}
\alias{GetGhcn}
\title{Get GHCN-D data for specified siteIds.}
\usage{
GetGhcn(siteIds, elements, startDate = NULL, endDate = NULL,
  parallel = FALSE, fileAdd = NULL)
}
\arguments{
\item{siteIds}{A single siteId or vector of siteIds. SiteIds should match the
standardized GHCN-D IDs (for example : ACW00011604). See
\url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt} for
a list of siteIds.}

\item{elements}{A character vector defining what type of observation you are
interested in. There are five core elements as well as a number of 
additional elements. The five core elements are: \describe{ \item{PRCP}{ =
Precipitation (tenths of mm), the precipitation will be converted to mm
when calling \code{\link{GetGhcn}} or \code{\link{GetGhcn2}} } \item{SNOW}{
= Snowfall (mm)} \item{SNWD}{ = Snow depth (mm)} \item{TMAX}{ = Maximum
temperature (tenths of degrees C)} \item{TMIN}{ = Minimum temperature
(tenths of degrees C)} } For the full list of elemenst refer to 
\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}}

\item{startDate, endDate}{Date.}

\item{parallel}{Logical (DEFAULT=FALSE)}

\item{fileAdd}{Address to the url containg all the daily GHCN data, default
is \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/all/}}
}
\value{
A dataframe containing the date, corresponding daily GHCN-D value, 
  the qFlag and the element.
}
\description{
\code{\link{GetGhcn}} downloads the daily GHCN (Global Historic Climatology
Network) data for each site in siteIds and creates a dataframe containing
five fields: siteIds, date, daily GHCN value, the qFlag and the element. Data
has become available in two formats either categorized by gauge station (use
\code{\link{GetGhcn}}) or categorized by year (use \code{\link{GetGhcn2}}). 
If there are many gauges, then \code{\link{GetGhcn2}} would be much faster.
}
\examples{
\dontrun{
siteIds=c("ACW00011604","AJ000037579","AJ000037883","ASN00005095")
startDate="1949/02/01"
endDate="1949/10/01"
element="PRCP"
obsPrcp<-GetGhcn(siteIds,element,startDate,endDate,parallel=FALSE)
}

# Or you could use the results of SelectGhcnGauges:
\dontrun{
countryCodeList=c("US")
networkCodeList=c("1")
statesList=c("WY")
selectedGauges<-SelectGhcnGauges(countryCode=countryCodeList,
                                 networkCode=networkCodeList,
                                 states=statesList)
obsPrcp<-GetGhcn(selectedGauges$siteIds,element,startDate,endDate,parallel=FALSE)
}
}
\seealso{
Other GHCN: \code{\link{CalcDailyGhcn}},
  \code{\link{GetGhcn2}}, \code{\link{SelectGhcnGauges}}
}
\concept{GHCN}
\keyword{IO}
