% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncdf_get_multi.R
\name{GetMultiNcdfVariable}
\alias{GetMultiNcdfVariable}
\title{The GetMultiNcdf for getting variables out of individual files.}
\usage{
GetMultiNcdfVariable(varInd, indexList, variableList, files,
  env = parent.frame(), parallel = FALSE)
}
\arguments{
\item{varInd}{The variable index.}

\item{indexList}{The variable list.}

\item{variableList}{The variable list.}

\item{files}{The files vector.}

\item{env}{The environment where the stat function lives}

\item{parallel}{Logical, this is the .parallel argument of plyr functions.}
}
\value{
A dataframe
}
\description{
\code{GetMultiNcdfVariable}: For a file, get all variables.
Typically calls GetFileStat to get the individual variables.
Typically called by GetMultiNcdfFile.
}
\seealso{
Other getMultiNcdf: \code{\link{GetFileStat}};
  \code{\link{GetMultiNcdfFile}}
}
\concept{
dataGet
}
\keyword{internal}

