% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetPkgMeta}
\alias{GetPkgMeta}
\title{Get a package's metadata fields and associated entries.}
\usage{
GetPkgMeta(meta = "concept", package = "rwrfhydro")
}
\arguments{
\item{meta}{Character the metadata field.}

\item{package}{Character The package to query for metadata.}
}
\value{
List of metadata fields in alphabetical order with corresponding entries.
}
\description{
\code{GetPkgMeta} Get metadata fields and associated entries from a package's documentation (e.g. "keyword" or "concepts".)
}
\examples{
GetPkgMeta()
GetPkgMeta('keyword', package='ggplot2')
}
\keyword{utilities}

