% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_poly.R
\name{GetRfc}
\alias{GetRfc}
\title{Return the RFC name for gauge locations.}
\usage{
GetRfc(points)
}
\arguments{
\item{points}{A dataframe of the gauges. The dataframe should contain at least two
fields: "latitude" and "longitude".}
}
\value{
Point dataframe with an extra column containing the RFC name for each gauge.
}
\description{
\code{GetRfc} takes a dataframe of gauge locations containing at least
two fields of latitude and longitude, overlays the points with RFC
shapefile, and adds the RFC name as a field to the input dataframe.
}
\examples{
\dontrun{
sg <- SelectGhcnGauges(countryCode="US",networkCode="C")[1:100,]
sg <- GetRfc(sg)
}
}
\seealso{
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GRIBLatLon}}; \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetTimeZone}}
}
\concept{
dataMgmt
}
\keyword{IO}

