% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{GetSiteHuc}
\alias{GetSiteHuc}
\title{Given a USGS site code, return its HUC8.}
\usage{
GetSiteHuc(site)
}
\arguments{
\item{site}{Character USGS site number.}
}
\value{
character HUC8
}
\description{
\code{GetSiteHuc} returns a HUC8 given a USGS site name.
}
\examples{
GetSiteHuc(FindUsgsStns(stnLon=254.67374999999998408,stnLat=40.018666670000001773,within=.001)$site_no)
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}};
  \code{\link{PrettySiteData}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataGet usgsStreamObs
}
\keyword{IO}

