<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_zone.R
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> 69bfbb5a27d96178965b8e8d863b6464ef1a29e8
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
>>>>>>> 45890bd465c2954b8ce326b6b129d8807788d3f0
% Please edit documentation in R/get_poly.R
>>>>>>> upstream/devBranch
\name{GetTimeZone}
\alias{GetTimeZone}
\title{Return the time zone for any point.}
\usage{
GetTimeZone(points,
  proj4 = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  parallel = FALSE)
}
\arguments{
\item{points}{A dataframe of the points The dataframe should contain at least two
fields: "latitude" and "longitude".}
<<<<<<< HEAD
<<<<<<< HEAD

\item{address}{Address to where polygonShapeFile is located.}

\item{polygonShapeFile}{Polygon shapefile containing the boundaries of
timezones (can be downloaded from \url{http://efele.net/maps/tz/world/}.
The TimeZone polygon has been included in rwrfhydro package. Therefore, location of 
polygonShapeFile does not have to be specified unless user has a different shapefile.}
=======
>>>>>>> upstream/devBranch
=======

\item{proj4}{Character: Projection to be used for points's dataframe.
(DEFAULT = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")}

\item{parallel}{Logical(DEFAULT = FALSE)}
>>>>>>> 69bfbb5a27d96178965b8e8d863b6464ef1a29e8
}
\value{
Dataframe with an extra column containing the time zone.
}
\description{
\code{GetTimeZone} takes a dataframe of locations containing at least
two fields of latitude and longitude, overlays the points with a timezone
shapefile (can be downloded from \url{http://efele.net/maps/tz/world/}),
and adds the timeZone as a field to the input dataframe.
}
\examples{
\dontrun{
sg <- SelectGhcnGauges(countryCode="US",networkCode="C")[1:10,]
sg <- GetTimeZone(sg)
}
}
\seealso{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
Other geospatial: \code{\link{ExportGeogrid}},
  \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}}
=======
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetRfc}}
>>>>>>> upstream/devBranch
=======
Other geospatial: \code{\link{ExportGeogrid}},
  \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}},
  \code{\link{GetPoly}}, \code{\link{GetProj}},
  \code{\link{GetRfc}}, \code{\link{PolyToRaster}}
>>>>>>> 69bfbb5a27d96178965b8e8d863b6464ef1a29e8
=======
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetProj}};
  \code{\link{GetRfc}}; \code{\link{PolyToRaster}}
>>>>>>> 45890bd465c2954b8ce326b6b129d8807788d3f0
}
\concept{
dataMgmt
}
\keyword{IO}

