<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_zone.R
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_poly.R
>>>>>>> upstream/devBranch
\name{GetTimeZone}
\alias{GetTimeZone}
\title{Return the time zone for gauge locations.}
\usage{
GetTimeZone(points)
}
\arguments{
\item{points}{A dataframe of the gauges. The dataframe should contain at least two
fields: "latitude" and "longitude".}
<<<<<<< HEAD

\item{address}{Address to where polygonShapeFile is located.}

\item{polygonShapeFile}{Polygon shapefile containing the boundaries of
timezones (can be downloaded from \url{http://efele.net/maps/tz/world/}.
The TimeZone polygon has been included in rwrfhydro package. Therefore, location of 
polygonShapeFile does not have to be specified unless user has a different shapefile.}
=======
>>>>>>> upstream/devBranch
}
\value{
Point dataframe with an extra column containing the time zone for each gauge.
}
\description{
\code{GetTimeZone} takes a dataframe of locations containing at least
two fields of latitude and longitude, overlays the points with a timezone
shapefile (can be downloded from \url{http://efele.net/maps/tz/world/}),
and adds the timezone as a field to the input dataframe.
}
\examples{
\dontrun{
sg <- SelectGhcnGauges(countryCode="US",networkCode="C")[1:100,]
sg <- GetTimeZone(sg)
}
}
\seealso{
<<<<<<< HEAD
Other geospatial: \code{\link{ExportGeogrid}},
  \code{\link{GetGeogridIndex}},
  \code{\link{GetGeogridSpatialInfo}}
=======
Other geospatial: \code{\link{ExportGeogrid}};
  \code{\link{GetGeogridIndex}};
  \code{\link{GetGeogridSpatialInfo}};
  \code{\link{GetPoly}}; \code{\link{GetRfc}}
>>>>>>> upstream/devBranch
}
\concept{
dataMgmt
}
\keyword{IO}

