% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{LoadMetaDB}
\alias{LoadMetaDB}
\title{Load the metadata for the USGS streamflow database.}
\usage{
LoadMetaDB(path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData",
  envir = globalenv())
}
\arguments{
\item{path}{Character path to the meta DB.}

\item{metaDBFileName}{Character name of the meta DB.}

\item{envir}{Envrionment where it is to be loaded.}
}
\value{
Character the name of the variables loaded with the file (invisible).
}
\description{
\code{LoadMetaDB} Load the metadata for the USGS streamflow database.
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}};
  \code{\link{PlotPrettyData}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}};
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{database}
\keyword{internal}

