% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{LsObjects}
\alias{LsObjects}
\title{List objects with more detailed metadata}
\usage{
LsObjects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{pos}{}

\item{pattern}{}

\item{order.by}{}

\item{decreasing}{}

\item{head}{}

\item{n}{}
}
\value{
dataframe
}
\description{
\code{LsObjects} lists objects with more detailed info on type, size,
etc. Blatantly plagiarized from the following sources:
Petr Pikal, David Hinds, Dirk Eddelbuettel, Tony Breyal.
}
\keyword{internal}
\keyword{utilities}

