% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{LsObjects}
\alias{LsObjects}
\title{List objects with more detailed metadata}
\usage{
LsObjects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{pos}{position}

\item{pattern}{pattern}

\item{order.by}{ordering variable}

\item{decreasing}{decreasing order}

\item{head}{head}

\item{n}{n}
}
\value{
dataframe
}
\description{
\code{LsObjects} lists objects with more detailed info on type, size,
etc. Taken from
Petr Pikal, David Hinds, Dirk Eddelbuettel, Tony Breyal.
}
\keyword{internal}
\keyword{utilities}

