% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lsOS}
\alias{lsOS}
\title{Shorthand call for LsObjects.}
\usage{
lsOS(..., n = 10)
}
\arguments{
\item{n}{n}
}
\value{
dataframe
}
\description{
\code{lsOS} Shorthand call for LsObjects
}
\keyword{internal}
\keyword{utilities}

