% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{lsOS}
\alias{lsOS}
\title{Shorthand call for LsObjects.}
\usage{
lsOS(..., n = 10)
}
\arguments{
\item{n}{n}
}
\value{
dataframe
}
\description{
\code{lsOS} Shorthand call for LsObjects
Blatantly plagiarized from the following sources:
Petr Pikal, David Hinds, Dirk Eddelbuettel, Tony Breyal.
}
\keyword{internal}
\keyword{utilities}

