% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf_make.R
\name{MkNcdf}
\alias{MkNcdf}
\title{List based creation of netcdf files.}
\usage{
MkNcdf(varList, filename, globalAttList = NULL, overwrite = FALSE,
  force_v4 = TRUE)
}
\arguments{
\item{varList}{List, the variable list. See structure in examples.}

\item{filename}{Character, the name of the netcdf file to be created.}

\item{globalAttList}{List, the global attribute list. See structure in examples.}

\item{overwrite}{Logical, overwrite (i.e. clobber in netcdf parlance) existing
file? Otherwise, the file will be appended if existing and correct (unlimited) dims
are extended.}

\item{force_v4}{Logical, make netcdf version 4 files.}
}
\value{
Returns the filename invisibly, if successful.
}
\description{
\code{MkNcdf} List based creation of netCDF files. Works for most cases.
Creates netcdf version 4 files by default. Appends to unlimited dimensions
in existing files.
}
\examples{
# Example 1 - Basic write.
varList = list()
varList[[1]] <- list( name='precipMult',
                      longname='Precipitation Multiplier',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList = list(scalar=list(name='scalar',values=1,
                                             units='-', unlimited=FALSE,
                                             create_dimvar=FALSE)),
                      data = 1:1 )
varList[[2]] <- list( name='precipMult22',
                      longname='Precipitation Multiplier22',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList =
                          list(  # n.b. the dimension order: charlen,z,y,x,t
                                y=list(name='y',values=2.5+((0:6)*5),
                                       units='lat', unlimited=FALSE,
                                       create_dimvar=TRUE),
                                x=list(name='x',values=c(10,20),
                                       units='lon', unlimited=FALSE,
                                       create_dimvar=TRUE)
                               ),
                     data = matrix( 1:7, nrow=7, ncol=2 ) )
globalAttList <- list()
globalAttList[[1]] <- list(name='Restart_Time',value='2012-07-05_00:00:00', precision="text")
globalAttList[[2]] <- list(name='Some reall atts',value='#$\%^!!', precision="text" )

outFile1 <- path.expand('~/test1.nc')
dum <- MkNcdf( varList, globalAttList=globalAttList, filename=outFile1 )
ncdump(outFile1)
unlink(outFile1)

#Example 2 - append to an existing file's variable.
varList1 = list()
varList1[[1]] <- list( name='precipMult22',
                      longname='Precipitation Multiplier22',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList =
                          list(  # n.b. the dimension order: charlen,z,y,x,t
                                y=list(name='y',values=2.5+((0:6)*5),
                                       units='lat', unlimited=TRUE,
                                       create_dimvar=TRUE),
                                x=list(name='x',values=c(10,20),
                                       units='lon', unlimited=FALSE,
                                       create_dimvar=TRUE)
                               ),
                     data = matrix( 1:7, nrow=7, ncol=2 ) )

varList2 = list()
varList2[[1]] <- list( name='precipMult22',
                      longname='Precipitation Multiplier22',
                      units='-',
                      precision = 'double',
                      missing = -9999,
                      dimensionList =
                          list(  # n.b. the dimension order: charlen,z,y,x,t
                                y=list(name='y',values=99,
                                       units='lat', unlimited=TRUE,
                                       create_dimvar=TRUE),
                                x=list(name='x',values=c(10,20),
                                       units='lon', unlimited=FALSE,
                                       create_dimvar=TRUE)
                               ),
                     data =  c(1:2)+10 )

globalAttList <- list()
globalAttList[[1]] <- list(name='Some reall atts',value='#$\%^!!', precision="text" )

outFile2 <- path.expand('~/test2.nc')
MkNcdf( varList1, globalAttList=globalAttList,
               filename='~/test2.nc', over=TRUE )
ncdump(outFile2)
unlink(outFile2)
}
\seealso{
Other ncdf: \code{\link{GetNcdfFile}}; \code{\link{ncdump}}
}
\concept{
ncdf
}

