% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_nudging_params.R
\name{MkNudgingParams}
\alias{MkNudgingParams}
\title{Create a parameter file for nudging.}
\usage{
MkNudgingParams(gageId, R, G, tau, outFile, overwrite = FALSE)
}
\arguments{
\item{gageId}{Character vector of gage identifiers also used in the timeslice files.}

\item{R}{Numeric vector same length as gageId which describes the radius of influence at each gage.}

\item{G}{Numeric vector same length as gageId which describes the nudging amplitude at each gage.}

\item{tau}{Numeric vector same length as gageId which describes the size of the temporal half-window in minutes.}

\item{outFile}{Character path/file to the desired output file}

\item{overwrite}{Logical, overwrite outFile if it already exists?}
}
\description{
Create a parameter file for nudging.
}
\examples{
\dontrun{
  ## once and future CONUS example snippett
  gageParams <- read.csv('~/ncar/WRF_Hydro/DOMAIN_library/CONUS/nhdRtIntersect.csv',
                         colClasses = 'character' )

  ## Boulder creek domain PMO example
  gageParams <- data.frame( gageId = paste0('g',formatC(1:403,width=3,flag='0')), stringsAsFactors=FALSE)
  gageParams$R=0
  gageParams$G=1
  gageParams$tau=20
  MkNudgingParams(gageId=gageParams$gageId, R=gageParams$R,
                  G=gageParams$G, tau=gageParams$tau,
                  outFile='~/WRF_Hydro/Col_Bldr_Creek/PMO/nudgingParams.PMOallGages.nc',
                  overwrite=TRUE)
}
}
\seealso{
Other nudging: \code{\link{AddRouteLinkGage}};
  \code{\link{ChanObsToTimeSlice}};
  \code{\link{EditFrxstPts}}
}
\concept{
dataMgmt nudging
}
\keyword{IO}
\keyword{manip}

