% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_nudging_params.R
\name{MkNudgingParams}
\alias{MkNudgingParams}
\title{Create a parameter file for nudging.}
\usage{
MkNudgingParams(gageId, R, G, tau, qThresh = NULL, acf = NULL, outFile,
  overwrite = FALSE, rmBlankGages = TRUE)
}
\arguments{
\item{gageId}{Character vector of gage identifiers also used in the timeslice files.}

\item{R}{Numeric vector same length as gageId which describes the radius of influence at each gage.}

\item{G}{Numeric vector same length as gageId which describes the nudging amplitude at each gage.}

\item{tau}{Numeric vector same length as gageId which describes the size of the temporal half-window in minutes.}

\item{qThresh}{Numeric [length(gageId), 12, nThresh] threshhold of flow for acf selection.}

\item{acf}{Numeric [length(gageId), 12, nThresh+1] actually the denominator in exp(-y/b).}

\item{outFile}{Character path/file to the desired output file}

\item{overwrite}{Logical, overwrite outFile if it already exists?}

\item{rmBlankGages}{Take out gages where the name is blank?}
}
\description{
Create a parameter file for nudging.
}
\examples{
\dontrun{
  ## Once and future CONUS example snippett
  gageParams <- read.csv('~/ncar/WRF_Hydro/DOMAIN_library/CONUS/nhdRtIntersect.csv',
                         colClasses = 'character' )
                         
  ## Boulder creek domain PMO example
  gageParams <- data.frame( gageId = paste0('g',formatC(1:403,width=3,flag='0')), stringsAsFactors=FALSE)
  gageParams$R=0
  gageParams$G=1
  gageParams$tau=20
  MkNudgingParams(gageId=gageParams$gageId, R=gageParams$R, 
                  G=gageParams$G, tau=gageParams$tau, 
                  outFile='~/WRF_Hydro/Col_Bldr_Creek/PMO/nudgingParams.PMOallGages.nc', 
                  overwrite=TRUE)
## FRNG example
## The existing params file
## existing parameter file to change/extend
paramPath <- '~/WRF_Hydro/FRNG_NHD/RUN/prstTestFeb2016/DOMAIN/'
frnParams <- paste0(paramPath,'nudgingParams.rwValid5.nc')
gageParamId <- ncdump(frnParams,'stationId',q=TRUE)
gageParamR <- ncdump(frnParams,'R',q=TRUE)
gageParamG <- ncdump(frnParams,'G',q=TRUE)
gageParamTau <- ncdump(frnParams,'tau',q=TRUE)
nGages=length(gageParamId)

## the existing parameter file had blanks, remove them.
options(warn=1)
MkNudgingParams(gageId=gageParamId, R=gageParamR, 
                G=gageParamG, tau=gageParamTau, 
                outFile=paste0(paramPath,'nudgingParams.rwValid5.rmBlanks.nc'),
                overwrite=TRUE, rmBlankGages=TRUE)


## In this example the exponent only depends on threshold, not location nor month.
gageParamqThresh1 = array(rep(c(.5,.8),each=4*12),dim=c(4,12,2))

MkNudgingParams(gageId=gageParams$gageId, R=gageParams$R, 
                G=gageParams$G, tau=gageParams$tau, 
                outFile=paste0(paramPath,'nudgingParams.rwValid5.persistence.nc'),
                overwrite=TRUE, rmBlankGages=TRUE)
}  #dontrun
}
\seealso{
Other nudging: \code{\link{AddRouteLinkGage}},
  \code{\link{ChanObsToTimeSlice}},
  \code{\link{EditFrxstPts}}
}
\concept{
dataMgmt nudging
}
\keyword{IO}
\keyword{manip}

