% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncdf_make.R
\name{ncdump}
\alias{ncdump}
\title{Emulate ncdump -h on OSX where ncdump might not be availabe.}
\usage{
ncdump(file, variable, quiet = FALSE)
}
\arguments{
\item{file}{Character, the file to inspect.}

\item{variable}{Character, a variable to return.}

\item{quiet}{Logical, suppress the 'meta' dump?}
}
\value{
If variable is not set, the meta object \code{ncdf4::nc_open(file)} is returned. If \code{variable}
is set, its values are returned.
}
\description{
I just hacked print.ncdf4 just to make it look more like unix output.
}
\examples{
# See \\code{\\link{MkNcdf}}.
}
\seealso{
Other ncdf: \code{\link{MkNcdf}}
}
\concept{
ncdf
}

