% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{NseLog}
\alias{NseLog}
\title{Calculate Log Nash-Sutcliffe Efficiency.}
\usage{
NseLog(m, o)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The Log Nash-Sutcliffe Efficiency.
}
\description{
\code{NseLog} calculates the Log Nash-Sutcliffe Efficiency.
Calculate the Nash-Sutcliffe Efficiency for vectors
of log-transformed modelled and observed values.
}

