% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{NtwKReExToNcdf}
\alias{NtwKReExToNcdf}
\title{Output the network reexpression to netcdf.}
\usage{
NtwKReExToNcdf(toFile, fromFile)
}
\arguments{
\item{toFile}{The Rdb (r binary file) for the downstream connectivity created by ReExpNetwork.}

\item{fromFile}{The Rdb (r binary file) for the upstream connectivity created by ReExpNetwork.}
}
\value{
Character path/file for the resulting netcdf file.
}
\description{
Output the network reexpression to netcdf.
}
\examples{
\dontrun{
reExp.nc    <- NtwKReExToNcdf(downstream.Rdb, upstream.Rdb)
}
}
\seealso{
Other networkExpression nudging: \code{\link{GatherStreamInds}};
  \code{\link{ReExpNetwork}};
  \code{\link{ReExpressRouteLink}};
  \code{\link{ReIndexRouteLink}};
  \code{\link{VisualizeRouteLink}};
  \code{\link{VisualizeSubsetStream}}
}
\concept{
dataMgmt nudging
}
\keyword{manip}

