% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_fdc.R
\name{PlotFdcCompare}
\alias{PlotFdcCompare}
\title{Plots a flow duration curve for up to three streamflow time series (e.g., observed & two model outputs)}
\usage{
PlotFdcCompare(strDf.obs, strCol.obs = "q_cms", strDf.mod1,
  strCol.mod1 = "q_cms", strDf.mod2 = NULL, strCol.mod2 = "q_cms",
  spline = TRUE, logy = TRUE, labelObs = "Observed",
  labelMod1 = "Model 1", labelMod2 = "Model 2")
}
\arguments{
\item{strDf.obs}{The OBSERVED streamflow time series dataframe (e.g., output from \code{\link{ReadUsgsGage}}).
The dataframe must contain a column of streamflow values and a POSIXct column.}

\item{strCol.obs}{The name of the column containing the streamflow values for the OBSERVED
dataframe (DEFAULT="q_cms").}

\item{strDf.mod1}{The FIRST MODEL streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}}).
The dataframe must contain a column of streamflow values and a POSIXct column.}

\item{strCol.mod1}{The name of the column containing the FIRST MODEL streamflow values
(DEFAULT="q_cms").}

\item{strDf.mod2}{The SECOND MODEL streamflow time series dataframe (e.g., output from \code{\link{ReadFrxstPts}}).
The dataframe must contain a column of streamflow values and a POSIXct column.}

\item{strCol.mod2}{The name of the column containing the SECOND MODEL streamflow values
(DEFAULT="q_cms").}

\item{spline}{(TRUE or FALSE) Option to add spline-fit curves to the plot (DEFAULT=TRUE).}

\item{logy}{(TRUE or FALSE) Optional flag to set the y-axis to log-scale (DEFAULT=TRUE).}

\item{labelObs}{Optional label for the observed streamflow (DEFAULT="Observed")}

\item{labelMod1}{Optional label for the FIRST MODEL (DEFAULT="Model 1")}

\item{labelMod2}{Optional label for the SECOND MODEL (DEFAULT="Model 2")}
}
\value{
A plot of flow (y) vs. flow exceedance probabilities (x).
}
\description{
\code{PlotFdcCompare} plots up to three flow duration curves for comparison.
}
\details{
\code{PlotFdcCompare} reads up to three dataframes (e.g., generated by \code{\link{ReadFrxstPts}} or \code{\link{ReadUsgsGage}}) and outputs a plot of
flow vs. percent exceedances and, optionally, the fitted spline curves. Intended to plot one observed flow duration
curve and one or two modelled flow duration curves for comparison. The tool will subset data to matching time
periods (e.g., if the observed data is at 5-min increments and modelled data is at 1-hr increments, the tool
will subset the observed data to select only observations on the matching hour break). Therefore, the tool will
RECALCULATE the flow exceedance probabilities on the fly.
}
\examples{
## Take a time series of observed 5-minute streamflow values for Fourmile Creek (obsStr5min.fc) and two hourly
## model runs (modStrh.allrt.fc, modStrh.chrt.fc), all with streamflow columns named "q_cms", and plot the
## flow duration curves for all three.

PlotFdcCompare(obsStr5min.fc, "q_cms", modStrh.allrt.fc, "q_cms", strDf.mod2=modStrh.chrt.fc, strCol.mod2="q_cms", labelObs="Observed Fourmile Creek", labelMod1="All Routing", labelMod2="Channel Routing Only")
}

