% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{PlotPrettyData}
\alias{PlotPrettyData}
\title{Plot USGS site data which has been prettied with PrettySiteData.}
\usage{
PlotPrettyData(prettyUsgs, plot = TRUE, errInnerQntl = 0.995)
}
\arguments{
\item{prettyUsgs}{dataframe returned from PrettySiteData}

\item{plot}{Logical to plot before returning or not.}

\item{errInnerQntl}{The inner quantile of the error estimate to be plotted with error bars. For example,
the "68-95-99.7 rule" where specifying these as the errInnerQntl would display 1, 2, and 3
standard deviations, respectively.}
}
\value{
A function(closure) with arguments controlling the look of its graphical output. It's actual
        return value is a list of ggplot2 object which can be custom manipulated.
}
\description{
\code{PlotPrettyData} plots USGS site data which has been prettied with \code{PrettySiteData}.
}
\examples{
# See vignette "Collect USGS stream observations and build a local database" for examples.
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetUsgsHucData}};
  \code{\link{ImproveHucMeta}}; \code{\link{SaveHucData}}
}
\concept{
plot
}
\keyword{hplot}

