% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snodas.R
\name{PutSnodasCoordsNcdf}
\alias{PutSnodasCoordsNcdf}
\title{Put the SNODAS coordinates into a netcdf file.}
\usage{
PutSnodasCoordsNcdf()
}
\value{
Success if the filename (which is SNODAS_Coordinates.nc), otherwise NULL.
}
\description{
\code{PutSnodasCoordsNcdf} Put the output of CalcSnodasCoords into a netcdf file.
}
\examples{
outFile <- PutSnodasCoordsNcdf()
ncdump(outFile)
unlink(outFile)
}
\seealso{
Other SNODAS: \code{\link{CalcSnodasCoords}};
  \code{\link{GetSnodasDepthSweDate}};
  \code{\link{PutSnodasNcdf}};
  \code{\link{ReadSnodasDepthSweDate}};
  \code{\link{getSnodasPointTs}}
}
\concept{
SNODAS
}
\keyword{IO}

