% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snodas.R
\name{PutSnodasNcdf}
\alias{PutSnodasNcdf}
\title{Write output of ReadSnodasDepthSweDate to netcdf.}
\usage{
PutSnodasNcdf(snodasList, outputDir = ".")
}
\arguments{
\item{snodasList}{The output of ReadSnodasDepthSweDate.}

\item{outputDir}{Character. The directory path where the output files are to be written.}
}
\value{
Success if the filename which is (SNODAS_YYYYMMDD.nc), otherwise NULL.
}
\description{
\code{PutSnodasNcdf} Put the output of ReadSnodasDepthSweDate into a netcdf file.
}
\examples{
\dontrun{
snodasGot <- GetSnodasDepthSweDate(as.POSIXct('2015-02-28'))
if(snodasGot) snodasList <- ReadSnodasDepthSweDate(as.POSIXct('2015-02-28'))
PutSnodasNcdf(snodasList)
}
}
\seealso{
Other SNODAS: \code{\link{CalcSnodasCoords}};
  \code{\link{GetSnodasDepthSweDate}};
  \code{\link{GetSnodasPointTs}};
  \code{\link{PutSnodasCoordsNcdf}};
  \code{\link{ReadSnodasDepthSweDate}}
}
\concept{
SNODAS
}
\keyword{IO}

