% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_usgs.R
\name{QuerySiteInfo}
\alias{QuerySiteInfo}
\title{Returns the desired information from the database file.}
\usage{
QuerySiteInfo(info, path = ".",
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{info}{Character vector, information fields in \code{HUC$prod$meta$SiteInfo$info}.}

\item{path}{Character The path to the database directory.}

\item{metaDBFileName}{Character The name of the metadata file.}
}
\value{
dataframe of requested info with all available HUC and product codes.
}
\description{
\code{QuerySiteInfo} gets the specified info from the local database.
}
\examples{
QuerySiteInfo(c('station_nm','site_no'), path='~/streamflow/OBS/')
}

