% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{QuerySiteName}
\alias{QuerySiteName}
\title{Find the name (site id) for a given site (name)  in the local database.}
\usage{
QuerySiteName(site, path = ".",
  retSiteId = tryCatch(is.na(as.numeric(site)), warning = function(w) TRUE,
  error = function(e) TRUE),
  metaDBFileName = "usgsDataRetrieval.metaDatabase.RData")
}
\arguments{
\item{site}{Character USGS site number or name.}

\item{path}{Character The path to the database.}

\item{retSiteId}{Logical return ID (name if FALSE) This is only exposed in
case there are issues, should work by default.}

\item{metaDBFileName}{Character The name of the metadata file.}
}
\value{
Character Site name or number.
}
\description{
\code{QuerySiteName} returns the name (site id) for a given site ID (name) in
the local database.
}
\examples{
\dontrun{
QuerySiteName('06730500', '~/streamflow/OBS/')
QuerySiteName('BOULDER CREEK AT MOUTH NEAR LONGMONT, CO', '~/streamflow/OBS/')
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}},
  \code{\link{GetSiteHuc}}, \code{\link{GetUsgsHucData}},
  \code{\link{GetUsgsIvProduct}},
  \code{\link{ImproveHucMeta}}, \code{\link{LoadMetaDB}},
  \code{\link{PlotPrettyUsgs}}, \code{\link{PrettyUsgs}},
  \code{\link{QueryHaveSite}}, \code{\link{QuerySiteData}},
  \code{\link{QuerySiteInfo}}, \code{\link{QuerySiteProd}},
  \code{\link{SaveHucData}}, \code{\link{[.prettyUsgs}},
  \code{\link{subset.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{database}

