% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadChrtout}
\alias{ReadChrtout}
\title{Read WRF-Hydro CHRTOUT data files.}
\usage{
ReadChrtout(pathOutdir, idList = NULL, gageList = NULL, rtlinkFile = NULL,
  parallel = FALSE, useDatatable = TRUE, gageOnly = TRUE,
  pattern = glob2rx("*.CHRTOUT_DOMAIN*"), idvar = "feature_id")
}
\arguments{
\item{pathOutdir}{The full pathname to the output directory containing the 
RTOUT files.}

\item{idList}{Optional list of station IDs to import (must be consistent
with IDs as used in the specified idvar variable).}

\item{gageList}{Optional list of gage IDs to import. Must provide a corresponding
route link file (used to map gage IDs to link IDs). Available only for reach-based
channel routing model runs.}

\item{rtlinkFile}{Optional path to the route link file. Available only for 
reach-based channel routing model runs.}

\item{parallel}{Logical for running in parallel mode (must have a parallel
backend installed and registered (e.g., doMC or doParallel) (DEFAULT=FALSE)}

\item{useDatatable}{Logical for utilizing the data.table package and 
outputting in data.table format (DEFAULT=TRUE)}

\item{gageOnly}{Logical for whether to bring in reaches with associated 
gage IDs only (vs. all reaches) (DEFAULT=TRUE)}

\item{pattern}{The pattern to match for file ingest
(DEFAULT=glob2rx('*.CHRTOUT_DOMAIN*'))}

\item{idvar}{The unique ID variable (DEFAULT="feature_id")}
}
\value{
A datatable containing a time series of channel fluxes.
}
\description{
\code{ReadChrtout} reads in WRF-Hydro CHRTOUT files and outputs a time series of
channel fluxes.
}
\details{
\code{ReadChrtout} reads standard-format WRF-Hydro CHRTOUT NetCDF files and 
outputs a time series of channel fluxes.
}
\examples{
## Take an OUTPUT directory for an hourly routing timestep model run of 
## the Front Range domain and create a new dataframe containing the channel
## fluxes for two USGS gages on Fourmile Creek.

\dontrun{
ReadChrtout('~/wrfHydroTestCases/FRN.REACH/OUTPUT', 
     gageList=c("06727500", "06730160"), 
     rtlinkFile="~/wrfHydroTestCases/FRN.REACH/DOMAIN/RouteLink.nc")
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadFrxstPts}}, \code{\link{ReadGwOut}},
  \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRouteLink}},
  \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}
\keyword{ts}
\keyword{univar}
