% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modelout.R
\name{ReadFrxstPts}
\alias{ReadFrxstPts}
\title{\code{ReadFrxstPts} reads in WRF-Hydro forecast points output text file.}
\usage{
ReadFrxstPts(pathOutfile, stIdType = "character", adjPosix = FALSE)
}
\arguments{
\item{pathOutfile}{The full pathname to the WRF-Hydro forecast points text file
(frxst_pts_out.txt).}

\item{stIdType}{Character describing the variable type desired for the stn_id variable, 
defaults to "character" but can also be "integer".}

\item{adjPosix}{If the first timest in the file matches the model init time, then set this to TRUE.}
}
\value{
A dataframe containing the forecast points output flow data. Note that POSIXct is the valid
time for the flow but timest is not, it is the LSM time prior to the flow at POSIXct.
}
\description{
\code{ReadFrxstPts} reads a standard-format WRF-Hydro forecast points output text
file and creates a dataframe with consistent date and data columns for use with other
rwrfhydro tools.
}
\examples{
## Take a forecast point output text file for an hourly model run of Fourmile Creek
## and return a dataframe.
\dontrun{
modStr1h.mod1.fc <- ReadFrxstPts("../OUTPUT/frxst_pts_out.txt")
}
}
\seealso{
Other modelDataReads: \code{\link{CreateBasinMask}},
  \code{\link{ReadChrtgrid}}, \code{\link{ReadChrtout}},
  \code{\link{ReadGwOut}}, \code{\link{ReadLdasoutAll}},
  \code{\link{ReadLdasoutWb}}, \code{\link{ReadRouteLink}},
  \code{\link{ReadRtout}}
}
\concept{
dataGet
}
\keyword{IO}
