% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_observations.R
\name{ReadNearStationData}
\alias{ReadNearStationData}
\title{Read point NetCDF station time series observations}
\usage{
ReadNearStationData(pathData)
}
\arguments{
\item{pathData}{The full pathname to the point NetCDF file.
\url{http://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html}}
}
\value{
A dataframe containing station observations.
}
\description{
\code{ReadNearStationData} reads station observation data table (point NetCDF file) and creates a dataframe.
}
\details{
\code{ReadNearStationData} reads a station observation  point NetCDF file
and outputs a dataframe with consistent date and data columns for use with
other rwrfhydro tools.
}
\examples{
## Takes a point NetCDF file of station observations
## (Rocky Mtn. Lodge) and returns a dataframe.
ReadNearStationData("URG1_Upper_Conejos_River_thru_Jul_1_2015.csv_filled.csv.nc")
}
\seealso{
Other obsDataReads: \code{\link{ReadAmerifluxCSV}};
  \code{\link{ReadAmerifluxNC}};
  \code{\link{ReadCoDwrGage}}; \code{\link{ReadUsgsGage}}
}
\concept{
dataGet
}
\keyword{IO}

