% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snodas.R
\name{ReadSnodasDepthSweDate}
\alias{ReadSnodasDepthSweDate}
\title{Read snodas snow depth and SWE into memory for a given date.}
\usage{
ReadSnodasDepthSweDate(datePOSIXct, outputDir = ".")
}
\arguments{
\item{datePOSIXct}{The date in POSIXct format for which data is desired. Only whole days matter.}

\item{outputDir}{The directory where the data are archived.}
}
\value{
A list with a SWE and a depth matrix.
}
\description{
\code{ReadSnodasDepthSweDate} Read (from local disk into memory) the SNODAS snow depth and SWE for a given date.
}
\examples{
snodasGot <- GetSnodasDepthSweDate(as.POSIXct('2015-02-28'),
                                   outputDir = path.expand('~'))
if(snodasGot) snodasList <- ReadSnodasDepthSweDate(as.POSIXct('2015-02-28'),
                                   outputDir = path.expand('~'))
unlink(path.expand('~/us_ssmv1103*.dat.gz'))
}
\seealso{
Other SNODAS: \code{\link{CalcSnodasCoords}};
  \code{\link{GetSnodasDepthSweDate}};
  \code{\link{PutSnodasCoordsNcdf}};
  \code{\link{PutSnodasNcdf}};
  \code{\link{getSnodasPointTs}}
}
\concept{
SNODAS
}
\keyword{manip}

