% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gage_gage_dist_gridded.R
\name{ReExpressGridChanConn}
\alias{ReExpressGridChanConn}
\title{gage-gage distances on gridded channel network}
\usage{
ReExpressGridChanConn(upstream, downstream)
}
\description{
gage-gage distances on gridded channel network
}
\examples{
library(rwrfhydro)
ncdump('~/ncar/WRF_Hydro/DOMAIN_library/Col_Bldr_Creek/CHANNEL_CONNECTIVITY.nc')
TO_NODE <-
ncdump('~/ncar/WRF_Hydro/DOMAIN_library/Col_Bldr_Creek/CHANNEL_CONNECTIVITY.nc',
       'TO_NODE', quiet=TRUE)
FROM_NODE <-
ncdump('~/ncar/WRF_Hydro/DOMAIN_library/Col_Bldr_Creek/CHANNEL_CONNECTIVITY.nc',
       'FROM_NODE', quiet=TRUE)
doMC::registerDoMC(4)
newCon <- ReExpressGridChanConn(FROM_NODE, TO_NODE)
upstream <- newCon$upstream
downstream <- newCon$downstream
TO_NODE[which(is.na(TO_NODE))] <- 0
done <- for(ii in 1:length(TO_NODE)) CheckReGridConnUp(ii)
done <- for(ii in 1:length(TO_NODE)) CheckReGridConnDown(ii)
save(upstream,downstream,
    file='~/ncar/WRF_Hydro/DOMAIN_library/Col_Bldr_Creek/newConnectivity.Rdb')
}

