% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid_multi_ncdf.R
\name{RegridMultiNcdfVar}
\alias{RegridMultiNcdfVar}
\title{The RegridMultiNcdf for getting variables out of individual files.}
\usage{
RegridMultiNcdfVar(varInd, varList, files, latVar, lonVar, wghtFile, geoFile,
  method)
}
\arguments{
\item{varInd}{The variable index.}

\item{varList}{The variable list.}

\item{files}{The files vector.}

\item{latVar}{The latitude variable}

\item{lonVar}{The longitude variable}

\item{wghtFile}{The weight file used for regridding.}

\item{geoFile}{The Geogrid LSM file for regridding.}

\item{method}{Method of regridding. Current acceptable values are
"bilinear"}
}
\value{
A list
}
\description{
\code{RegridMultiNcdfVar}: For a group of NetCDF files, extract
and regrid a list of variables to a WRF-Hydro domain.
}
\seealso{
Other RegridMultiNcdf: \code{\link{RegridNcdf}}
}
\keyword{DataRegrid}
\keyword{internal}

