% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_multi_ncdf.R
\name{regridNcdf}
\alias{regridNcdf}
\title{Extraction and regridding of NetCDF data for multiple files for a 
given variable.}
\usage{
regridNcdf(var, files, latVar, lonVar, wghtFile, geoFile, method)
}
\arguments{
\item{var}{The GRIB variable being regridded}

\item{files}{The files vector list}

\item{latVar}{The latitude variable from the source files.}

\item{lonVar}{The longitude variable from the source files.}

\item{wghtFile}{The weight regridding file. If it does not exist,
it will be created.}

\item{geoFile}{The Geogrid LSM file needed for regridding.}

\item{method}{Method of regridding. Current acceptable values are
"bilinear"}
}
\value{
A list containing the regridded stack.
}
\description{
\code{regridNcdf}: For a list of NetCDF files, extract and regrid
a variable WRF-Hydro domain.
}
\seealso{
Other regridMultiNcdf: \code{\link{regridMultiNcdfVar}}
}
\concept{
dataRegrid
}
\keyword{internal}

