% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/network_expression.R
\name{ReIndexRouteLink}
\alias{ReIndexRouteLink}
\title{ReIndex RouteLink.nc (netcdf) files for indexed network traversal.}
\usage{
ReIndexRouteLink(routeLinkFile)
}
\arguments{
\item{routeLinkFile}{The netcdf routelink file to process.}
}
\value{
The resulting file which was written to disk, of the form "infile.reInd.nc"
}
\description{
\code{ReIndexRouteLink} reIndexes by order in the RouteLink file,
replacing ComID with this index.
}
\examples{
\dontrun{
reIndFile <-
 ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink_2015_07_31.nc')
reIndFile <-
 ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink4.nc')
reIndFile <-
 ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/CONUS_IOC/DOMAIN/RouteLink3.nc')
reIndFile <-
 ReIndexRouteLink(routeLinkFile <- '~/WRF_Hydro/DOMAIN_library/BoCr_100m_1km_NHDPlus_2015_08_11/Route_Link.nc')
}
}
\seealso{
Other networkExpression nudging: \code{\link{NtwKReExToNcdf}};
  \code{\link{ReExpNetwork}};
  \code{\link{ReExpressRouteLink}};
  \code{\link{VisualizeRouteLink}};
  \code{\link{VisualizeSubsetStream}}
}
\concept{
dataMgmt
}
\keyword{manip}

