% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Rmse}
\alias{Rmse}
\title{Calculate root mean squared error.}
\usage{
Rmse(m, o, na.rm = TRUE)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The root mean squared error.
}
\description{
\code{Rmse} calculates the root mean squared error.
Calculate the root mean squared error for vectors
of modelled and observed values.
}
\keyword{internal}

