% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{RmseNorm}
\alias{RmseNorm}
\title{Calculate normalized root mean squared error.}
\usage{
RmseNorm(m, o)
}
\arguments{
\item{m}{The vector of modelled values.}

\item{o}{The vector of observed values.}
}
\value{
The nrmalized root mean squared error.
}
\description{
\code{RmseNorm} calculates the normalized root mean squared error.
Calculate the normalized root mean squared error for vectors
of modelled and observed values.
}
\keyword{internal}

