% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{RotateCw}
\alias{RotateCw}
\title{Rotate a matrix clock-wise.}
\usage{
RotateCw(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\description{
\code{RotateCw} Rotates a matrix clock-wise.
}
\examples{
x <- matrix(1:9, 3)
x
RotateCw(x)
RotateCw(RotateCw(x))
}
\keyword{internal}

