% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{subset.prettyUsgs}
\alias{subset.prettyUsgs}
\title{Subset prettyUsgs objects.}
\usage{
\method{subset}{prettyUsgs}(prettyUsgs, ...)
}
\arguments{
\item{prettyUsgs}{A dataframe of class \code{c("prettyUsgs", "data.frame")} returned from \code{PrettyUsgs}}

\item{...}{additional arguments to subset.data.frame}
}
\value{
A dataframe of class c("prettyUsgs", "data.frame")
# See vignette "Collect USGS stream observations and build a local database" for examples.
}
\description{
\code{subset.prettyUsgs} subsets prettyUsgs objects and retains their attributes.
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}};
  \code{\link{GetSiteHuc}}; \code{\link{GetUsgsHucData}};
  \code{\link{GetUsgsIvProduct}};
  \code{\link{ImproveHucMeta}}; \code{\link{LoadMetaDB}};
  \code{\link{PlotPrettyData}}; \code{\link{PrettyUsgs}};
  \code{\link{QueryHaveSite}}; \code{\link{QuerySiteData}};
  \code{\link{QuerySiteInfo}}; \code{\link{QuerySiteName}};
  \code{\link{QuerySiteProd}}; \code{\link{SaveHucData}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{manip}

