% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_ingest.R
\name{subset.prettyUsgs}
\alias{subset.prettyUsgs}
\title{Subset prettyUsgs objects.}
\usage{
\method{subset}{prettyUsgs}(prettyUsgs, ...)
}
\arguments{
\item{prettyUsgs}{A dataframe of class \code{c("prettyUsgs", "data.frame")}
returned from \code{PrettyUsgs}}

\item{...}{additional arguments to subset.data.frame}
}
\value{
A dataframe of class c("prettyUsgs", "data.frame") # See vignette
  "Collect USGS stream observations and build a local database" for examples.
}
\description{
\code{subset.prettyUsgs} subsets prettyUsgs objects and retains their
attributes.
}
\examples{
\dontrun{
dbPath <- '~/wrfHydroTestCases/usgsDb/'
threeSites <- QuerySiteInfo(path = dbPath, info = c("site_no"))[1:3,]$site_no
obs <- PrettyUsgs(QueryHaveSite(threeSites, path=dbPath, ret='00060', get=TRUE))
length(unique(obs$site_no))
df0 <- subset(obs,site_no==threeSites[1])
length(unique(df0$site_no))
ggplot2::ggplot(obs, ggplot2::aes(x=value,color=site_no)) + 
  ggplot2::geom_density() + ggplot2::scale_x_continuous(limits=c(0,7))
df1 <- subset(obs,value < 2)
length(unique(df1$site_no))
length(attr(df1,'siteInfo')$site_no)
plyr::ddply(obs, plyr::.(site_no), plyr::summarize, max=max(value))
df2 <- subset(obs,value > 9)
length(unique(df2$site_no))
length(attr(df2,'siteInfo')$site_no)
}
}
\seealso{
Other streamObs: \code{\link{FindUsgsStns}},
  \code{\link{GetSiteHuc}}, \code{\link{GetUsgsHucData}},
  \code{\link{GetUsgsIvProduct}},
  \code{\link{ImproveHucMeta}}, \code{\link{LoadMetaDB}},
  \code{\link{PlotPrettyUsgs}}, \code{\link{PrettyUsgs}},
  \code{\link{QueryHaveSite}}, \code{\link{QuerySiteData}},
  \code{\link{QuerySiteInfo}}, \code{\link{QuerySiteName}},
  \code{\link{QuerySiteProd}}, \code{\link{SaveHucData}},
  \code{\link{[.prettyUsgs}}
}
\concept{
dataMgmt usgsStreamObs
}
\keyword{internal}
\keyword{manip}

