% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_zone.R
\docType{data}
\name{timeZone}
\alias{timeZone}
\title{Time zone polygon}
\format{data.frame}
\usage{
timeZone
}
\description{
Polygon shapefile of world time zone classification.
}
\section{Citation}{

Online_Linkage: \link[http://efele.net/maps/tz/world/]{http://efele.net/maps/tz/world/} \cr
}
\concept{
data
}
\keyword{data}

