% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{TransTz}
\alias{TransTz}
\title{Translate (i.e. invert) timezones to the so calle Olson names used by POSIXct.}
\usage{
TransTz(tz)
}
\arguments{
\item{tz}{The timezone to be translated.}
}
\description{
Translate formatted timezones codes to the so-called "Olson names" used by POSIXct.
\code{TransTz} translates the formatted timezone codes (incl those from USGS) to Olson Names.
}
\examples{
as.POSIXct('2012-01-01')
as.POSIXct('2012-01-01', tz='US/Pacific')
format(as.POSIXct('2012-01-01', tz='US/Pacific'),'\%Z')
TransTz(format(as.POSIXct('2012-01-01', tz='US/Pacific'),'\%Z'))
lubridate::with_tz(as.POSIXct('2012-01-01'),TransTz(format(as.POSIXct('2012-01-01', tz='US/Pacific'),'\%Z')))
}
\keyword{internal}

