% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_expression.R
\name{VisualizeSubsetStream}
\alias{VisualizeSubsetStream}
\title{Visualize upstream or downstream links determined from GatherStreamInds}
\usage{
VisualizeSubsetStream(indDist, ncFile, comIds = TRUE, downstreamReExp = "",
  ...)
}
\arguments{
\item{indDist}{List containing indies and accumulated distance from start, obtained from GatherStreamInds}

\item{ncFile}{Route Link file read in/initially processed with VisualizeRouteLink()}

\item{comIds}{Logical, show the comIds or the link indices in the Route_Link.nc file.}

\item{downstreamReExp}{Character, eliminate searching by providing the re-expressed network file}

\item{...}{arguments to the function returned by VisualizeRouteLink.}
}
\value{
Map of Route Links with selected upstream/downstream links highlighted in red, starting location in black
}
\description{
Visualize upstream or downstream links determined from GatherStreamInds
}
\examples{
\dontrun{
 ## see example for GatherStream
 file <- '~/WRF_Hydro/DOMAIN_library/Boulder_Creek_100m_1km_2sqkm_full_2015_09_03/Route_Link.nc'
 VisualizeSubsetStream(upstreamInds, file)
 VisualizeSubsetStream(upstreamInds, file, com=FALSE, zoom=10, textColor='purple')
 VisualizeSubsetStream(downstreamInds, file, com=TRUE, zoom=10)
 VisualizeSubsetStream(downstreamInds, file, com=FALSE, zoom=10, linkColor='lightblue', maptype='satellite')
}
}
\seealso{
Other networkExpression nudging: \code{\link{GatherStreamInds}},
  \code{\link{NtwKReExToNcdf}}, \code{\link{ReExpNetwork}},
  \code{\link{ReExpressRouteLink}},
  \code{\link{ReIndexRouteLink}},
  \code{\link{VisualizeRouteLink}}
}
\concept{
nudging plot
}
\keyword{hplot}
