% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usgs_to_ncdf.R
\name{WriteNcTimeSlice}
\alias{WriteNcTimeSlice}
\title{Write a USGS discharge timeslice to a netcdf file.}
\usage{
WriteNcTimeSlice(dfByPosix, outPath)
}
\arguments{
\item{dfByPosix}{Dataframe, a data frame with the following columns:
\code{site_no}, \code{dateTime}, \code{code}, \code{queryTime}, \code{discharge.cms},
and \code{variance} where dateTime is the same for the entire dataframe.}

\item{outPath}{Character, the file name with path for the output netcdf file.}
}
\description{
Write a USGS discharge timeslice to a netcdf file.
}
\examples{
## See \\link{MkUsgsTimeSlice}.
}
\keyword{internal}

