% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_to_ncdf.R
\name{WriteNcTimeSlice}
\alias{WriteNcTimeSlice}
\title{Write a USGS discharge timeslice to a netcdf file.}
\usage{
WriteNcTimeSlice(dfByPosix, outPath, sliceResolution)
}
\arguments{
\item{dfByPosix}{Dataframe, a data frame with the following columns: 
\code{site_no}, \code{dateTime}, \code{dateTimeRound},
\code{queryTime}, \code{discharge.cms}, and \code{discharge.quality} where dateTimeRound
is the same for the entire dataframe.}

\item{outPath}{Character, the path for the output netcdf file.}

\item{sliceResolution}{The Temporal resolution.}
}
\description{
Write a USGS discharge timeslice to a netcdf file.
}
\examples{
## See \\link{MkUsgsTimeSlice}.
}

